% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{after_init}
\alias{after_init}
\title{Post Initialization}
\usage{
after_init(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
List consisting of:
 \itemize{
  \item \code{inp} Updated input data.
  \item \code{out} Updated output data.
  \item \code{method} Updated embedding method.
 }
}
\description{
Function called after input and output data have been initialized. Useful for
doing data-dependent initialization of stiffness parameters, e.g. based on
the parameterization of the input probabilities.
}

