% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{after_step}
\alias{after_step}
\title{Post Optimization Step}
\usage{
after_step(opt, inp, out, new_out, ok, iter)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data from the start of the iteration.}

\item{new_out}{Output data which will be the starting solution for the next
iteration of optimization. If the validation stage failed, then this may be
the same solution as \code{out}.}

\item{ok}{\code{TRUE} if the current iteration passed validation,
\code{FALSE} otherwise.}

\item{iter}{Current iteration number.}
}
\value{
A list containing
  \item{\code{opt}}{Updated optimizer.}
  \item{\code{inp}}{Input data.}
  \item{\code{out}}{Output data from the start of the iteration.}
  \item{\code{new_out}}{New output to be used in the next iteration.}
}
\description{
A function for the optimizer to run after updating the solution on each
iteration. As part of this process it will run any \code{after_step}
functions provided by the direction, step size and update methods of the
optimizer.
}

