% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolfe.R
\name{armijo_ok}
\alias{armijo_ok}
\title{Armijo Rule (or Sufficient Decrease Condition)}
\usage{
armijo_ok(phi_0, dphi_0, alpha, phi_a, c1)
}
\arguments{
\item{phi_0}{Value of phi at alpha = 0}

\item{dphi_0}{the directional derivative at alpha = 0}

\item{alpha}{the step length.}

\item{phi_a}{the value of phi at \code{alpha}.}

\item{c1}{the sufficient decrease constant.}
}
\value{
\code{TRUE} if the step \code{alpha} represents a sufficient decrease.
}
\description{
Armijo Rule (or Sufficient Decrease Condition)
}

