% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{asne_plugin}
\alias{asne_plugin}
\title{ASNE Method using Plugin Gradient}
\usage{
asne_plugin(beta = 1, eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{beta}{Bandwidth parameter of the exponential similarity kernel
function.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
An implementation of ASNE using the plugin gradient.
}
\seealso{
\code{asne} should give equivalent results, but is probably
a bit more efficient.

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{hsjse_plugin}}, \code{\link{hsjse}},
  \code{\link{hsnerv_plugin}}, \code{\link{hsnerv}},
  \code{\link{hssne_plugin}}, \code{\link{hssne}},
  \code{\link{jse_plugin}}, \code{\link{jse}},
  \code{\link{mmds}}, \code{\link{nerv_plugin}},
  \code{\link{nerv}}, \code{\link{plugin}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sammon_map}}, \code{\link{sjse_plugin}},
  \code{\link{sjse}}, \code{\link{smmds}},
  \code{\link{snerv_plugin}}, \code{\link{snerv}},
  \code{\link{ssne_plugin}}, \code{\link{ssne}},
  \code{\link{tasne_plugin}}, \code{\link{tasne}},
  \code{\link{tnerv_plugin}}, \code{\link{tnerv}},
  \code{\link{tsne_plugin}}, \code{\link{tsne}},
  \code{\link{uhsnerv_plugin}}, \code{\link{uhsnerv}},
  \code{\link{unerv_plugin}}, \code{\link{unerv}},
  \code{\link{usnerv_plugin}}, \code{\link{usnerv}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hsjse_plugin}}, \code{\link{hsjse}},
  \code{\link{hsnerv_plugin}}, \code{\link{hsnerv}},
  \code{\link{hssne_plugin}}, \code{\link{hssne}},
  \code{\link{jse_plugin}}, \code{\link{jse}},
  \code{\link{nerv_plugin}}, \code{\link{nerv}},
  \code{\link{plugin}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sjse_plugin}}, \code{\link{sjse}},
  \code{\link{snerv_plugin}}, \code{\link{snerv}},
  \code{\link{ssne_plugin}}, \code{\link{ssne}},
  \code{\link{tasne_plugin}}, \code{\link{tasne}},
  \code{\link{tnerv_plugin}}, \code{\link{tnerv}},
  \code{\link{tsne_plugin}}, \code{\link{tsne}},
  \code{\link{uhsnerv_plugin}}, \code{\link{uhsnerv}},
  \code{\link{unerv_plugin}}, \code{\link{unerv}},
  \code{\link{usnerv_plugin}}, \code{\link{usnerv}}
}

