% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{asymm_weights_to_pjoint}
\alias{asymm_weights_to_pjoint}
\title{Create Joint Probability Matrix from Asymmetric Weight Matrix}
\usage{
asymm_weights_to_pjoint(wm)
}
\arguments{
\item{wm}{Asymmetric matrix of weighted distances.}
}
\value{
Joint probability matrix.
}
\description{
A weight to probability function.
}
\details{
Creates a joint probability matrix: the grand sum of the elements are
one, and the matrix is symmetric. The weight matrix should be asymmetric.
A symmetric matrix is also a valid input, but it's more  efficient to use
\code{symm_weights_to_pjoint}.
}

