% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc_mat}
\alias{auc_mat}
\title{Average Area Under a Curve}
\usage{
auc_mat(dm, labels, auc_row_fn)
}
\arguments{
\item{dm}{Distance matrix.}

\item{labels}{Vector of labels, of the same size as the number of rows
(or columns) in the distance matrix.}

\item{auc_row_fn}{A function which can calculate the Area Under a Curve
for a particular quality measure. Should have the signature
\code{auc_row_fn(dm, labels, i)} where \code{i} is the ith row of the
distance matrix, and should return a scalar value giving the area under the
curve using the ith row of the distance matrix.}
}
\value{
A list containing:
\item{av_auc}{Area Under the curve, averaged over each observation.}
The list also contains the average AUC per class label, with each average
being named after the class label.
}
\description{
Embedding quality measure.
}
\details{
This function calculates a curve using the specified function, repeating the
procedure N times, where N is the number of the observations. Each time
a different row of the distance matrix is used. The label of the Nth
observation is set as the positive class and then the other observations are
ranked according to their distance from the Nth observation in the output
coordinates (lower distances being better). Observations with the same label
as the Nth observation count as positive observations. The final reported
result is the average over all observations.
}
\note{
Use of this function requires that the \code{PRROC} package be
installed.
}

