% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{back_nag}
\alias{back_nag}
\title{Nesterov Accelerated Gradient Optimizer with Backstepping Step Search}
\usage{
back_nag(min_step_size = sqrt(.Machine$double.eps), max_momentum = 1,
  normalize_direction = TRUE, linear_weight = TRUE)
}
\arguments{
\item{min_step_size}{Minimum step size allowed.}

\item{max_momentum}{Maximum value the momentum may take.}

\item{normalize_direction}{If \code{TRUE}, scale the length of the gradient to
one.}

\item{linear_weight}{If \code{TRUE}, then the contribution of the gradient
descent part of the update is scaled relative to the momentum part.}
}
\value{
Optimizer with NAG parameters and backstepping step size.
}
\description{
Optimizer factory function.
}
\details{
Wrapper around \code{make_opt} which mixes the NAG descent method and
momentum for non-strongly convex problems formulated by Sutskever et al.,
along with a backstepping method for step size.
}
\examples{
# Should be passed to the opt argument of an embedding function:
\dontrun{
 embed_prob(opt = back_nag(), ...)
}
}
\seealso{
\code{embed_prob} and \code{embed_dist} for how to use
 this function for configuring an embedding.

Other sneer optimization methods: \code{\link{back_nag_adapt}},
  \code{\link{bold_nag_adapt}}, \code{\link{bold_nag}},
  \code{\link{gradient_descent}}, \code{\link{make_opt}},
  \code{\link{nag}}, \code{\link{optimization_methods}},
  \code{\link{ropt}}, \code{\link{tsne_opt}}
}

