% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_backtrack.R
\name{backtracking_line_search}
\alias{backtracking_line_search}
\title{Backtracking Line Search}
\usage{
backtracking_line_search(opt, inp, out, method, iter, alpha = 1, c1 = 0.1,
  rho = 0.8, min_step_size = .Machine$double.eps)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}

\item{iter}{Iteration number.}

\item{alpha}{Initial step size.}

\item{c1}{Constant for the sufficient decrease condition.}

\item{rho}{Factor to decrease alpha by.}

\item{min_step_size}{Mininum step size.}
}
\value{
Step length satisfying the Wolfe conditions.
}
\description{
Finds a step length that satisifes the Wolfe conditions.
}
\details{
This function implements backtracking line search to find a step length
which fulfils both the curvature and sufficient decrease (aka Armijo)
conditions. It starts the line search at a step length given by the
\code{alpha} parameter, and if the sufficient decrease condition is not
met, decreases the step length by a factor of \code{rho} until it does.
}
\references{
Nocedal, J., & Wright, S. (2006).
\emph{Numerical optimization.}
Springer Science & Business Media.
}

