% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{bnx_crm}
\alias{bnx_crm}
\title{Intrusions and Extrusions for K-ary Neighborhoods (BNX)}
\usage{
bnx_crm(crm, k)
}
\arguments{
\item{crm}{Co-ranking matrix. Create from a pair of distance matrices with
\code{\link{coranking_matrix}}.}

\item{k}{Neighborhood size.}
}
\value{
BNX for \code{k}.
}
\description{
BNX measures the degree of intrusions versus extrusions that contributes
to the QNX measure of embedding error. If BNX > 0 this means that intrusions
dominate over extrusions: i.e. non-neighbors in the input space are neighbors
in the output space. BNX < 0 means that extrusions dominate over intrusions:
neighbors in the input space tend to be non-neighbors in the output space.
}

