% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_step.R
\name{bold_driver}
\alias{bold_driver}
\title{Bold driver step size.}
\usage{
bold_driver(inc_mult = 1.1, dec_mult = 0.5, inc_fn = partial(`*`,
  inc_mult), dec_fn = partial(`*`, dec_mult), init_step_size = 1,
  min_step_size = sqrt(.Machine$double.eps), max_step_size = NULL)
}
\arguments{
\item{inc_mult}{Multiplier of the current step size when the cost
decreases. Should be greater than one to increase the step size. This
parameter is ignored if \code{inc_fun} is supplied.}

\item{dec_mult}{Multiplier of the current step size when the cost
increases. Should be smaller than one to decrease the step size. This
parameter is ignored if \code{dec_fun} is supplied.}

\item{inc_fn}{Function to apply to the current step size when the cost
decreases. Should return a value greater than the current step size.}

\item{dec_fn}{Function to apply to the current step size when the cost
increases. Should return a value smaller than the current step size.}

\item{init_step_size}{Step size to attempt on the first step of
optimization.}

\item{min_step_size}{Minimum step size.}

\item{max_step_size}{Maximum step size.}
}
\value{
Step size method, to be used by the Optimizer. A list containing:
 \item{\code{inc_fn}}{Function to invoke to increase the step size.}
 \item{\code{dec_fn}}{Function to invoke to decrease the step size.}
 \item{\code{init_step_size}}{Initial step size.}
 \item{\code{min_step_size}}{Minimum step size.}
 \item{\code{max_step_size}}{Maximum step size.}
 \item{\code{init}}{Function to do any needed initialization.}
 \item{\code{get_step_size}}{Function to return the current step size.}
 \item{\code{validate}}{Function to validate whether the current step was
 successful or not.}
 \item{\code{after_step}}{Function to do any needed updating or internal state
 before the next optimization step.}
}
\description{
This function configures the 'Bold Driver' method for step size selection.
If the costdecreases after an optimization step occurs, then the step
size will be increased (normally by a conservative amount). If the cost
increases, then the step size is decreased (normally by a more drastic
amount).
}
\seealso{
Other sneer optimization step size methods: \code{\link{jacobs}},
  \code{\link{tsne_jacobs}}
}

