% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{bold_nag_opt}
\alias{bold_nag_opt}
\title{Nesterov Accelerated Gradient optimizer with bold driver}
\usage{
bold_nag_opt(min_step_size = sqrt(.Machine$double.eps), init_step_size = 1,
  max_momentum = 1)
}
\arguments{
\item{min_step_size}{Minimum step size allowed.}

\item{init_step_size}{Initial step size.}

\item{max_momentum}{Maximum value the momentum may take.}
}
\value{
Optimizer with NAG parameters and bold driver step size.
}
\description{
Convenience factory function which makes a very performant optimizer. Mixes
the NAG descent method and momentum for non-strongly convex problems
formulated by Sutkever et al., along with the bold driver method for step
size.
}
\examples{
# Should be passed to the opt argument of an embedding function:
\dontrun{
 embed_sim(opt = bold_nag_opt(), ...)
}
}
\seealso{
\code{\link{embed_sim}} for how to use this function for configuring
an embedding.

Other sneer optimization methods: \code{\link{make_opt}},
  \code{\link{tsne_opt}}
}

