% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{calculate_cost}
\alias{calculate_cost}
\title{Evaluate Cost Function}
\usage{
calculate_cost(method, inp, out)
}
\arguments{
\item{method}{Embedding method containing a cost function.}

\item{inp}{Input data.}

\item{out}{Output data.}
}
\value{
Scalar cost function value.
}
\description{
Evaluate Cost Function
}

