% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{calculate_gradient}
\alias{calculate_gradient}
\title{Gradient Calculation}
\usage{
calculate_gradient(opt, inp, out, method, iter)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data containing the desired position.}

\item{method}{Embedding method.}

\item{iter}{Iteration number.}
}
\value{
List containing:
\item{\code{km}}{Stiffness matrix.}
\item{\code{gm}}{Gradient matrix.}
}
\description{
Calculate the gradient of the cost function for a specified position, and
a given type of optimization.
}

