% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{calculate_importance}
\alias{calculate_importance}
\title{Degree Centrality}
\usage{
calculate_importance(inp, out, method)
}
\arguments{
\item{inp}{Input data}

\item{out}{Output data}

\item{method}{Embedding method}
}
\value{
list containing the updated method.
}
\description{
Calculates the degree centrality of a probability matrix.
}
\details{
The input probability matrix is weighted to create the symmetrized nearest
neighbor graph. The degree centrality of node i is then the sum of row
and column i, where the element \code{m[i, j]} of matrix \code{m} can be
interpreted as the value of the weighted edge from point i to j.
}
\references{
Yang, Z., Peltonen, J., & Kaski, S. (2014).
Optimization equivalence of divergences improves neighbor embedding.
In \emph{Proceedings of the 31st International Conference on Machine Learning (ICML-14)}
(pp. 460-468).
}

