% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{centrality}
\alias{centrality}
\title{Centrality Measure for Input Probability}
\usage{
centrality(inp, method)
}
\arguments{
\item{inp}{Input method}

\item{method}{Embedding method.}
}
\value{
Vector containing the centrality measure for each observation in the
data set.
}
\description{
Converts the input probability to a nearest neighbour adjacency matrix
and then calculates a centrality measure.
}

