% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{check_symmetry}
\alias{check_symmetry}
\title{Ensure the Kernel has the Correct Symmetry}
\usage{
check_symmetry(kernel)
}
\arguments{
\item{kernel}{A similarity kernel.}
}
\value{
Kernel with the type attribute of the function correctly set for
its parameters.
}
\description{
This function should be called when the parameters of a kernel (e.g.
the beta parameter of the exponential) are changed from a scalar to
a vector or (vice versa). Such a change may result in the symmetry of
the kernel changing from asymmetric to symmetric (and vice versa again).
}

