% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clamp}
\alias{clamp}
\title{Clamp Numerical Values}
\usage{
clamp(x, min_val = .Machine$double.eps, max_val = NULL)
}
\arguments{
\item{x}{Matrix.}

\item{min_val}{Minimum value allowed for any element in the matrix.}

\item{max_val}{Maximum value allowed for any element in the matrix.}
}
\value{
Matrix with the clamped values.
}
\description{
Values are truncated so that they lie within (\code{min_val, max_val}). In
embedding this is used to prevent individual probabilities values getting
too small and causing underflow or some other horrible explosion.
}

