% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{classical_grad_pos}
\alias{classical_grad_pos}
\title{Classical Gradient Calculation}
\usage{
classical_grad_pos(opt, inp, out, method)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
List containing:
 \item{\code{km}}{Stiffness matrix.}
 \item{\code{gm}}{Gradient matrix.}
}
\description{
If the solution is currently at \code{out$ym}, this function calculates the
gradient is calculated at this position. Contrast this with Nesterov
Accelerated Gradient Descent.
}

