% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{classical_gradient}
\alias{classical_gradient}
\title{Classical Gradient}
\usage{
classical_gradient()
}
\value{
Classical gradient calculation method.
}
\description{
Factory function for creating an optimizer gradient method.
}
\details{
Calculates the gradient at the current location of the solution.
Can also provide the position at which the gradient is calculated, useful
for some step length calculation methods.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# gradient method:
make_opt(gradient = classical_gradient())
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{\link{optimization_gradient_interface}}
for details on the functions and values defined for this method.

Other sneer optimization gradient methods: \code{\link{nesterov_gradient}},
  \code{\link{optimization_gradient}}
}

