% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{classical_position}
\alias{classical_position}
\title{Classical Gradient Position}
\usage{
classical_position(opt, inp, out, method, iter)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}

\item{iter}{Iteration number.}
}
\value{
List containing:
 \item{\code{km}}{Stiffness matrix.}
 \item{\code{gm}}{Gradient matrix.}
}
\description{
Function for calculating the position to evaluate the gradient at.
}
\details{
If the solution is currently at \code{out$ym}, this function returns that
position. In standard gradient descent optimization, this is the location
where the gradient would be calculated.
}
\seealso{
\code{\link{nesterov_position}} for an alternative location to
calculate the gradient at.
}

