% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{colorBrewerPalette}
\alias{colorBrewerPalette}
\title{Interpolated Color Brewer Palette}
\usage{
colorBrewerPalette(name, n)
}
\arguments{
\item{name}{Name of the palette.}

\item{n}{Number of colors desired.}
}
\value{
Vector of \code{n} colors from the palette.
}
\description{
Returns a vector of colors from the specified palette, interpolated if the
number of requested colors is larger than the number of colors in the
palette. Sequential and Diverging palettes are suitable for numerical scales.
The Qualitiative palettes are intended for categorical values.
}
\details{
Sequential palettes names:
 Blues BuGn BuPu GnBu Greens Greys Oranges OrRd PuBu PuBuGn PuRd Purples
 RdPu Reds YlGn YlGnBu YlOrBr YlOrRd
Diverging palette names:
 BrBG PiYG PRGn PuOr RdBu RdGy RdYlBu RdYlGn Spectral
Qualitative:
 Accent Dark2 Paired	Pastel1 Pastel2	Set1 Set2	Set3
}
\seealso{
More information on ColorBrewer is available at its website,
\url{http://www.colorbrewer2.org}.
}

