% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{constant_momentum}
\alias{constant_momentum}
\title{Constant Momentum}
\usage{
constant_momentum(momentum)
}
\arguments{
\item{momentum}{Momentum value to use.}
}
\value{
Constant momentum update method, to be used by the optimizer.
}
\description{
Factory function for creating an optimizer update method.
}
\details{
Create a callback for the optimizer to use with a constant momentum term.
This may be useful in the final stages of an optimization for fine tuning
a solution.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# update method:
make_opt(update = constant_momentum())
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{\link{optimization_update_interface}}
for details on the functions and values defined for this method.

Other sneer optimization update methods: \code{\link{linear_momentum}},
  \code{\link{nesterov_nsc_momentum}},
  \code{\link{no_momentum}},
  \code{\link{optimization_update}},
  \code{\link{step_momentum}}
}

