% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_step.R
\name{constant_step_size}
\alias{constant_step_size}
\title{Fixed Step Size}
\usage{
constant_step_size(step_size = 1)
}
\arguments{
\item{step_size}{Value of the step_size.}
}
\description{
Factory function for creating an optimizer step size method.
}
\details{
This function creates a step size method where the step size is always
a constant value.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# step size method:
make_opt(step_size = constant_step_size(0.1))
}
\seealso{
Other sneer optimization step size methods: \code{\link{bold_driver}},
  \code{\link{jacobs}},
  \code{\link{optimization_step_size}},
  \code{\link{tsne_jacobs}}
}

