% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{coords_to_dist2}
\alias{coords_to_dist2}
\title{Squared (Euclidean) Distance Matrix}
\usage{
coords_to_dist2(xm)
}
\arguments{
\item{xm}{a matrix of coordinates}
}
\value{
Squared distance matrix.
}
\description{
Creates a matrix of squared Euclidean distances from a coordinate matrix.
}
\details{
Probability-based embedding techniques use the squared Euclidean distance
as input to their weighting functions.
}

