% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{coords_to_weights}
\alias{coords_to_weights}
\title{Create Weight Matrix from Coordinates Matrix}
\usage{
coords_to_weights(ym, weight_fn)
}
\arguments{
\item{ym}{Matrix of coordinates.}

\item{weight_fn}{Function with signature \code{weight_fn(d2m)}
where \code{d2m} is a matrix of squared distances between the output
coordinates. It should return a weight matrix.}
}
\value{
Weight matrix.
}
\description{
Create Weight Matrix from Coordinates Matrix
}

