% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{coranking_matrix}
\alias{coranking_matrix}
\title{Co-ranking Matrix}
\usage{
coranking_matrix(din, dout)
}
\arguments{
\item{din}{Input distance matrix.}

\item{dout}{Output distance matrix.}
}
\value{
Co-ranking matrix.
}
\description{
Calculates the co-ranking matrix for an embedding.
}
\details{
The co-ranking matrix is the basic data structure used for calculating
various quality metrics, such as \code{\link{qnx_crm}},
\code{\link{rnx_crm}} and \code{\link{bnx_crm}}.

The co-ranking matrix is an N x N matrix where N is the number of
observations (so it's the same size as the distance matrices its
constructed from). The element (i, j) in the co-ranking matrix is the
number of times an ith-nearest neighbor of an observation in the input
distance matrix was the jth-nearest neighbor in the output space.

The lower diagonal represents "intrusions". This is when observations
have a larger rank in the input space than in the output space,
i.e. non-neighbors are falsely marked as neighbors in the output space.

The upper diagonal represents "extrusions". This occurs when observations
have a smaller rank in the input space than in the output space,
i.e. true neighbors are falsely marked as non-neighbors in the output space.
}
\references{
Lee, J. A., & Verleysen, M. (2009).
Quality assessment of dimensionality reduction: Rank-based criteria.
\emph{Neurocomputing}, \emph{72(7)}, 1431-1443.
}

