% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost_gradient_fd}
\alias{cost_gradient_fd}
\title{Finite Difference Gradient Calculation}
\usage{
cost_gradient_fd(inp, out, method, diff = 1e-04)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data containing the desired position.}

\item{method}{Embedding method.}

\item{diff}{Step size to take in finite difference calculation.}
}
\value{
Gradient matrix.
}
\description{
Calculate the gradient of the cost function for a specified position using
a finite difference.
}
\details{
Only intended for testing that analytical gradients have been calculated
correctly.
}

