% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_backtrack.R
\name{cost_step_length}
\alias{cost_step_length}
\title{Cost Evaluated at a Step Length Along a Search Direction}
\usage{
cost_step_length(opt, inp, method, y0, pm, alpha)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{method}{Embedding method.}

\item{y0}{Coordinate start point of the line search.}

\item{pm}{Direction vector to search along.}

\item{alpha}{Step length along the direction.}
}
\value{
Cost at the location given by \code{y0 + alpha * pm}.
}
\description{
This function calculates the cost function when the solution is updated
based on moving a specified step length along a search direction.
}

