% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{cost_validate}
\alias{cost_validate}
\title{Cost Validation}
\usage{
cost_validate(opt, inp, out, proposed_out, method, iter)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}

\item{iter}{Iteration number.}
}
\value{
List containing:
\item{ok}{\code{TRUE} if the cost did not increase with the proposed
 solution.}
\item{opt}{Optimizer with any updated data (e.g. cached costs).}
}
\description{
Validation function for assessing a proposed solution from the optimizer.
}
\details{
This function validates a solution by looking for a non-increasing cost
function between iterations. Used by adaptive schemes, e.g. the
\code{\link{bold_driver}} step size method and the
\code{\link{adaptive_restart}} momentum scheme.

Because multiple methods may use this function, some caching is carried out
to prevent wasteful recalculation of the cost function. This function will
set the following members on the \code{opt} object:
\describe{
 \item{\code{old_cost_dirty}}{The cost associated with the previous solution,
   \code{out}. If this value already exists, and \code{opt$old_cost_dirty} is
   \code{FALSE}, it is assumed that the old cost is up to date and it won't
   be recalculated.}
 \item{\code{cost}}{The cost associated with the proposed solution,
   \code{proposed_out}. If this value already exists, and
   \code{opt$cost_iter} exists and is equal to the current value of
   \code{iter}, it is assumed that the cost is up to data and it won't be
   recalculated.}
 \item{\code{old_cost_dirty}}{A flag indicating that cost data from the
   previous iteration has been invalidated. If the function recalculates the
   old cost, this flag will be reset to false.}
 \item{\code{cost_iter}}{The iteration at which the current cost was
   calculated. If the function recalculates the cost of the proposed
   solution, this value will be set to the current value of \code{iter}.}
 \item{\code{cost_ok}}{\code{TRUE} if the cost for the proposed solution has not
   increased that of the old solution.}
}
}
\keyword{internal}

