% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolfe.R
\name{curvature_ok}
\alias{curvature_ok}
\title{Curvature Condition}
\usage{
curvature_ok(d0, da, c2)
}
\arguments{
\item{d0}{Value of gradient of the line at the starting position of the line
search.}

\item{da}{Value of gradient of the line at a distance along the line.}

\item{c2}{Sufficient decrease constant.}
}
\value{
\code{TRUE} if the standard curvature condition is met.
}
\description{
Curvature Condition
}

