% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{d_to_p_perp_bisect}
\alias{d_to_p_perp_bisect}
\title{Find Row Probabilities by Perplexity Bisection Search}
\usage{
d_to_p_perp_bisect(dm, perplexity = 15, weight_fn, tol = 1e-05,
  max_iters = 50, verbose = TRUE)
}
\arguments{
\item{dm}{Distance matrix.}

\item{perplexity}{Target perplexity value.}

\item{weight_fn}{Function which maps squared distances to weights. Should
have the following signature: \code{function(d2m, beta)}}

\item{tol}{Convergence tolerance for perplexity.}

\item{max_iters}{Maximum number of iterations to carry out the search.}

\item{verbose}{If TRUE, logs information about the beta values.}
}
\value{
List with the following members:
 \item{\code{pm}}{Row probability matrix. Each row is a probability
 distribution with a perplexity within \code{tol} of \code{perplexity}.}
 \item{\code{beta}}{Matrix of beta parameters used with \code{weight_fn} that
 generated \code{pm}.}
}
\description{
For each row, finds the value of beta which generates the probability
with the desired perplexity.
}

