% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{dist2_to_weights}
\alias{dist2_to_weights}
\title{Create Weight Matrix from Squared Distances}
\usage{
dist2_to_weights(d2m, weight_fn)
}
\arguments{
\item{d2m}{Matrix of squared distances.}

\item{weight_fn}{Function with signature \code{weight_fn(d2m)} where
\code{d2m} is a matrix of squared distances.}
}
\value{
Weights matrix. The diagonal (i.e. self-weight) is enforced to be
zero.
}
\description{
Weights are subsequently normalized to probabilities in probability-based
embedding. This function guarantees that the self-weight is 0.
}

