% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{distance_embedding_methods}
\alias{distance_embedding_methods}
\title{Distance-Based Embeddings}
\description{
The available embedding methods which work by reproducing input distances.
}
\examples{
\dontrun{
embed_dist(method = mmds(), ...)
embed_dist(method = sammon_map(eps = 1e-4), ...)
}
}
\seealso{
Other sneer distance embedding methods: \code{\link{mmds}},
  \code{\link{sammon_map}}, \code{\link{smmds}}
}
\keyword{internal}

