% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Euclidean Distance Matrix}
\usage{
distance_matrix(xm, min_dist = .Machine$double.eps)
}
\arguments{
\item{xm}{Matrix of coordinates.}

\item{min_dist}{Truncate any inter-point distances in \code{xm} less than
this value.}
}
\value{
Distance matrix.
}
\description{
Creates an Euclidean distance matrix with the type "\code{matrix}", rather
than an  object of class "\code{dist}", which the \code{stats} function
\code{dist} produces.
}

