% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dot}
\alias{dot}
\title{Dot Product}
\usage{
dot(a, b)
}
\arguments{
\item{a}{A vector or matrix.}

\item{b}{Another vector or matrix of the same dimensions as \code{a}.}
}
\value{
Dot product of \code{a} and \code{b}.
}
\description{
Returns the dot product of two vectors. But will also treat two matrices
of equal dimensions as if they were two vectors. For the purposes of an
emedding, this allows the various matrices involved to keep their "natural"
shape, i.e. the gradient can be calculated as an N x N matrix without having
to convert it into a vector, as normally required in optimization methods.
}
\examples{
\dontrun{
 a <- c(1, 2, 3, 4)
 b <- c(5, 6, 7, 8)
 dot(a, b) # 70
 am <- matrix(a, nrow = 2) # turn them into matrices
 bm <- matrix(b, nrow = 2)
 dot(am, bm) # still 70
}
}

