% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricks.R
\name{early_exaggeration}
\alias{early_exaggeration}
\title{Early Exaggeration}
\usage{
early_exaggeration(exaggeration = 4, off_iter = 50, verbose = TRUE)
}
\arguments{
\item{exaggeration}{Size of the exaggeration factor: input probabilities will
be multiplied by this value.}

\item{off_iter}{Iteration step at which the input probabilities are returned
to their original values.}

\item{verbose}{If \code{TRUE} report a message when exaggeration is turned
off.}
}
\value{
Trick callback. Should be passed to \code{make_tricks} when
configuring an embedding.
}
\description{
A "trick" for improving the quality of the embedding.
}
\details{
This trick is for use with probability-based embedding. It scales up the
input probabilities for the first few iterations of the embedding to
encourage close distances to form between very similar points.
}
\examples{
\dontrun{
# exaggerate for the first 100 iterations
embed_prob(make_tricks(early_exaggeration(off_iter = 100)), ...)
}
}
\seealso{
Other sneer tricks: \code{\link{late_momentum}},
  \code{\link{tricks}}
}

