% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed_main}
\alias{embed_main}
\title{Generic Embedding}
\usage{
embed_main(xm, method, init_inp, init_out, opt, max_iter = 1000,
  tricks = NULL, reporter = NULL, preprocess = make_preprocess(),
  export = NULL, after_embed = NULL)
}
\arguments{
\item{xm}{A matrix or data frame to embed.}

\item{method}{Embedding method. Set by assigning the result value of one of
the configuration functions listed in
\code{\link{embedding_methods}}.}

\item{init_inp}{Input initializer. Set by assigning the result value of
calling one of the available \code{\link{input_initializers}}.}

\item{init_out}{Output initializer. Set by assigning the result value of
calling one of the available \code{\link{output_initializers}}.}

\item{opt}{Optimization method. Set by assigning the result of value of one
of the configuration functions listed in
\code{\link{optimization_methods}}.}

\item{max_iter}{Maximum number of optimization steps to take.}

\item{tricks}{Optional collection of heuristics. Set by assigning the result
value of \code{\link{make_tricks}} or a related wrapper. See
\code{\link{tricks}} for the available tricks.}

\item{reporter}{Reporter callback. Set by assigning the result value of
\code{\link{make_reporter}}.}

\item{preprocess}{Input data preprocess callback. Set by assigning the
result value of \code{\link{make_preprocess}}.}

\item{export}{Vector of names, specifying data in addition to of the output
data to export provide in the return value. Possible names are:
\describe{
  \item{"\code{inp}"}{The input data, based on the argument passed to
    \code{init_inp}. See the help text for the specific
    \code{\link{input_initializers}} for details.}
  \item{"\code{report}"}{The result of the last report. See the help
    text for \code{\link{make_reporter}} for details.}
  \item{"\code{opt}"}{The optimizer. See the help text for the optimizer
  passed to \code{opt} for details.}
}}

\item{after_embed}{Callback to run on input and output data before output
data is returned.}
}
\value{
The output data. A list containing:
 \itemize{
  \item \code{ym} Embedded coordinates.
  \item \code{cost} The cost value associated with \code{ym}.
  If the parameter \code{export} was used, additional elements will be
  present. See the help text for the \code{export} parameter for more
  details.
}
}
\description{
Carry out an embedding of a dataset using any embedding method. The most
generic embedding function, and conversely the one with the fewest helpful
defaults.
}
\seealso{
\itemize{
\item \code{\link{embedding_methods}} for configuring \code{method}
\item \code{\link{make_preprocess}} for configuring \code{preprocess}
\item \code{\link{input_initializers}} for configuring \code{init_inp}
\item \code{\link{output_initializers}} for configuring \code{init_out}
\item \code{\link{optimization_methods}} for configuring \code{opt}
\item \code{\link{make_tricks}} for configuring \code{tricks}
\item \code{\link{make_reporter}} for configuring \code{reporter}
}
}

