% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{equal_factors}
\alias{equal_factors}
\title{Split A Vector Into Equally Populated Factors}
\usage{
equal_factors(x, nfactors)
}
\arguments{
\item{x}{Numeric vector}

\item{nfactors}{Number of factors required.}
}
\value{
factor-encoded vector specifying the level for each item in the
vector.
}
\description{
Assigns each member of a vector to a factor, based on the quantiles of the
distribution, so that each factor is equal populated. Levels range from
one to \code{nfactors}.
}

