% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{exp_gr}
\alias{exp_gr}
\title{Exponential Gradient}
\usage{
exp_gr(d2m, beta = 1)
}
\arguments{
\item{d2m}{Matrix of squared distances.}

\item{beta}{exponential parameter.}
}
\value{
Matrix containing the gradient of (with respect to \code{d2m}).
}
\description{
Similarity Kernel Gradient.
}
\details{
Calculates the gradient of the exponential function with respect to d2m,
the matrix of squared distances.
}

