% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{exp_kernel}
\alias{exp_kernel}
\title{Exponential Kernel Factory Function}
\usage{
exp_kernel(beta = 1)
}
\arguments{
\item{beta}{exponential parameter.}
}
\value{
Exponential function and gradient.
}
\description{
Similarity Kernel factory function.
}
\details{
Creates a list implementing the exponential kernel function and gradient.
}
\seealso{
Other sneer similiarity kernels: \code{\link{heavy_tail_kernel}},
  \code{\link{tdist_kernel}}
}

