% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{exp_weight}
\alias{exp_weight}
\title{Exponential weight.}
\usage{
exp_weight(d2m, beta = 1)
}
\arguments{
\item{d2m}{Matrix of squared distances.}

\item{beta}{exponential parameter.}
}
\value{
Weight matrix.
}
\description{
Applies exponential weighting to the squared distances.
}
\details{
The weight matrix, \eqn{W} is generated from the squared distances matrix,
\eqn{D2} by:

\deqn{W = e^{-\beta D2}}{exp(-beta * D2)}
}
\seealso{
Other sneer weight functions: \code{\link{sqrt_exp_weight}},
  \code{\link{tdist_weight}}
}

