% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{find_beta}
\alias{find_beta}
\title{Find Beta Parameter for One Row of Matrix}
\usage{
find_beta(d2mi, i, perplexity, beta_init = 1, weight_fn, tol = 1e-05,
  max_iters = 50)
}
\arguments{
\item{d2mi}{Row of a squared distance matrix.}

\item{i}{Index of the row in the squared distance matrix.}

\item{perplexity}{Target perplexity value.}

\item{beta_init}{Initial guess for beta.}

\item{weight_fn}{Function which maps squared distances to weights. Should
have the following signature: \code{function(D2i, beta)}}

\item{tol}{Convergence tolerance for perplexity.}

\item{max_iters}{Maximum number of iterations to carry out the search.}
}
\value{
List with the following members:
 \item{\code{pm}}{Matrix with one row containing a probability distribution
 with a perplexity within \code{tol} of \code{perplexity}.}
 \item{\code{beta}}{Beta parameter used with \code{weight_fn} that generated
 \code{pm}.}
 \item{\code{perp}}{Final perplexity of the probability, differing from
 \code{perplexity} only if \code{max_iters} was exceeded.}
}
\description{
Find the value of beta that produces a specific perplexity when a row of
a distance matrix is converted to probabilities.
}

