% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{frey_faces}
\alias{frey_faces}
\title{Frey Faces dataset}
\format{A data frame with 1965 rows and 561 variables.}
\usage{
frey_faces()
}
\value{
the Frey Faces dataset as a dataframe.
}
\description{
Sneer benchmarking data.
}
\details{
Returns the Frey Faces dataset in a data frame reformatted to be suitable
for use with Sneer. This is a series of 1965 images (with dimension 20 x 28)
of Brendan Frey's face taken from sequential frames of a video.

The variables are as follows:
\itemize{
\item \code{px1}, \code{px2}, \code{px3} ... \code{px560} 8-bit grayscale
pixel values (0-255). The pixel index starts at the top right of the image
(\code{px1}) and are then stored row-wise.
\item \code{Label} An integer in the range (1-1965) indicating the frame.
Provides the same information as the row id, but included to be consistent
with other sneer datasets.
}
}
\note{
requires the \code{RnavGraphImageData} package.
\url{https://cran.r-project.org/web/packages/RnavGraphImageData/index.html}
to be installed and loaded.
}
\seealso{
Saul Roweis' dataset web page: \url{http://www.cs.nyu.edu/~roweis/data.html}.
Each row can be visualized as an image using \code{\link{show_frey_face}}.
}

