% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gauss}
\alias{gauss}
\title{d-Dimensional Gaussian Data Set}
\usage{
gauss(n = 1000, d = 3, nlabels = 5)
}
\arguments{
\item{n}{Number of points to create.}

\item{d}{Dimension of the Gaussian.}

\item{nlabels}{Number of labels.}
}
\value{
Data frame.
}
\description{
Simulation data.
}
\details{
Creates a series of points randomly sampled from an d-Dimensional Gaussian
with mean zero and variance 1.

Labels go from \code{0} to \code{nlabels - 1}. Points are classified into
labels based on their distance from the origin, where that distance is
divided into \code{nlabels} equal lengths. That results in unequal volumes
and hence unequal populations of the labels.
}

