% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{gradient}
\alias{gradient}
\title{Calculate the gradient of the cost function for the current configuration.}
\usage{
gradient(inp, out, method, mat_name = "ym")
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}

\item{mat_name}{Name of the matrix in the output data list that contains the
embedded coordinates.}
}
\value{
List containing:
\item{\code{km}}{Stiffness matrix.}
\item{\code{gm}}{Gradient matrix.}
}
\description{
Calculate the gradient of the cost function for the current configuration.
}

