% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Steepest Descent Optimizer with No Momentum}
\usage{
gradient_descent()
}
\value{
Pure gradient (steepest) descent optimizer.
}
\description{
Optimizer factory function.
}
\details{
Wrapper around \code{\link{make_opt}} that creates a simple (some might say
boring) optimizer that only does steepest descent, without any momentum
term. The gradient is normalized to length 1 and the bold driver step
size method is used to adaptively select the step size.
}
\examples{
# Should be passed to the opt argument of an embedding function:
\dontrun{
 embed_prob(opt = gradient_descent(), ...)
}
}
\seealso{
\code{\link{embed_prob}} and \code{\link{embed_dist}} for how to use
 this function for configuring an embedding.

Other sneer optimization methods: \code{\link{bold_nag}},
  \code{\link{make_opt}}, \code{\link{nag}},
  \code{\link{optimization_methods}},
  \code{\link{tsne_opt}}
}

