% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{gradient_fd}
\alias{gradient_fd}
\title{Finite Difference Gradient Calculation}
\usage{
gradient_fd(opt, inp, out, method, mat_name = "ym", diff = 1e-04)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data containing the desired position.}

\item{method}{Embedding method.}

\item{mat_name}{Name of the matrix in the output data list that contains the
embedded coordinates.}

\item{diff}{Step size to take in finite difference calculation.}
}
\value{
List containing:
\item{\code{gm}}{Gradient matrix.}
}
\description{
Calculate the gradient of the cost function for a specified position using
a finite difference.
}
\details{
Only intended for testing that analytical gradients have been calculated
correctly.
}

