% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{gradient_update_term}
\alias{gradient_update_term}
\title{Gradient term of an update}
\usage{
gradient_update_term(opt)
}
\arguments{
\item{opt}{Optimizer.}
}
\value{
Gradient update.
}
\description{
Calculates the gradient term of an update.
}

