% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{handle_prob}
\alias{handle_prob}
\title{Probability Type Conversion}
\usage{
handle_prob(pm, method)
}
\arguments{
\item{pm}{Probability Matrix. Must have a "type" attribute with one of the
following values:
\describe{
\item{"row"}{A probability where each row sums to one.}
\item{"cond"}{A conditional probablilty matrix where the grand sum of the
entire matrix is 1.}
\item{"joint"}{A joint probability matrix where the grand sum of the entire
matrix is 1, and \code{pm[i, j]} = \code{pm[j, i]} for all pairs i and j.}
}
See \code{\link{probability_matrices}} for more details.}

\item{method}{Embedding method, which must have a "type" attribute with one
of the values described for \code{pm}. This represents the type of
probability matrix that \code{pm} will be converted into.}
}
\value{
Converted probability matrix.
}
\description{
Given a probability matrix and an embedding method, this function applies
a conversion from the type of the probability matrix into that required
by the embedding method.
}

