% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{heavy_tail_gr}
\alias{heavy_tail_gr}
\title{Heavy Tail Kernel Gradient.}
\usage{
heavy_tail_gr(d2m, beta = 1, alpha = 1.5e-08)
}
\arguments{
\item{d2m}{Matrix of squared distances.}

\item{beta}{The precision of the function. Becomes equivalent to the
precision in the Gaussian distribution of distances as \code{alpha}
approaches zero.}

\item{alpha}{Tail heaviness. Must be greater than zero.}
}
\value{
Matrix containing the gradient (with respect to \code{d2m}).
}
\description{
Calculates the gradient of the Student-t distribution with one degree of
freedom with respect to d2m, the matrix of squared distances.
}

