% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{heavy_tail_kernel}
\alias{heavy_tail_kernel}
\title{Heavy Tailed Kernel Factory Function}
\usage{
heavy_tail_kernel(beta = 1, alpha = 0)
}
\arguments{
\item{beta}{The bandwidth of the function. Becomes equivalent to the
precision of the exponential distribution of squared distances as
\code{alpha} approaches zero.}

\item{alpha}{Tail heaviness. Must be greater than zero.}
}
\value{
Heavy tailed function and gradient.
}
\description{
Similarity Kernel factory function.
}
\details{
Creates a list implementing a heavy tailed (compared to an exponential)
function and gradient.
}
\seealso{
Other sneer similiarity kernels: \code{\link{exp_kernel}},
  \code{\link{tdist_kernel}}
}

