% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{heavy_tail_weight}
\alias{heavy_tail_weight}
\title{Heavy-Tailed Similarity}
\usage{
heavy_tail_weight(d2m, beta = 1, alpha = 1.5e-08)
}
\arguments{
\item{d2m}{Matrix of squared distances.}

\item{beta}{The precision of the function. Becomes equivalent to the
precision in the Gaussian distribution of distances as \code{alpha}
approaches zero.}

\item{alpha}{Tail heaviness. Must be greater than zero.}
}
\value{
Weight matrix.
}
\description{
A similarity function for probability-based embedding.
}
\details{
Applies a "heavy-tailed" similarity function that represents
a generalization of the similarity functions used in SNE and t-SNE. The
heavy-tailedness is with respect to that of the exponential functon.
The weight matrix, \eqn{W} is generated from the squared distances matrix,
\eqn{D^2}{D2} by:
\deqn{W  = [(\alpha \beta D^2) + 1]^{-\frac{1}{\alpha}}}{W = ((alpha * beta * D2) + 1) ^ (-1 / alpha)}

\eqn{\alpha \to 0}{alpha approaches 0}, the weighting function becomes
exponential, and behaves like \code{\link{exp_weight}}. At
\eqn{\alpha = 1}{alpha = 1}, the weighting function is the t-distribution
with one degree of freedom, \code{\link{tdist_weight}}. Intermediate
values provide an intermediate degree of tail heaviness.
The \eqn{\beta}{beta} parameter is equivalent to that in the
\code{\link{exp_weight}} function. This is set to one in methods suchs as
SSNE and t-SNE.
}
\examples{
# make a matrix of squared distances
d2m <- dist(matrix(rnorm(12), nrow = 3)) ^ 2

# exponential weighting
wm_exp <-   heavy_tail_weight(d2m, alpha = 1.5e-8)

# t-distributed weighting
wm_tdist <- heavy_tail_weight(d2m, alpha = 0.0)

# exponential weighting with a non-standard beta value
wm_expb2 <- heavy_tail_weight(d2m, alpha = 1.5e-8, beta = 2.0)
}
\references{
Yang, Z., King, I., Xu, Z., & Oja, E. (2009).
Heavy-tailed symmetric stochastic neighbor embedding.
In \emph{Advances in neural information processing systems} (pp. 2169-2177).
}
\seealso{
Other sneer weight functions: \code{\link{exp_weight}},
  \code{\link{sqrt_exp_weight}}, \code{\link{tdist_weight}}
}

