% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{helix}
\alias{helix}
\title{Toroidal Helix Data Set}
\usage{
helix(n = 1000, rmajor = 2, rminor = 1, nwinds = 8, nlabels = 5)
}
\arguments{
\item{n}{Number of points to create.}

\item{rmajor}{Major radius.}

\item{rminor}{Minor radius.}

\item{nwinds}{Number of winds the helix makes.}

\item{nlabels}{Number of labels.}
}
\value{
Data frame.
}
\description{
Simulation data.
}
\details{
Creates a series of points sampled from a 3D helix with the ends joined
to each other.

Unlike \code{\link{ball}} and \code{\link{sphere}}, this data set is not
randomly sampled.

Labels go from \code{0} to \code{nlabels - 1}. Points are classified into
labels based on their z-coordinate, where that distance is divided into
\code{nlabels} equal lengths. That results in unequal volumes and hence
unequal populations of the labels.
}
\references{
A dataset like this was used in:
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}

