% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jse.R
\name{hsjse_stiffness}
\alias{hsjse_stiffness}
\title{HSJSE Stiffness Function}
\usage{
hsjse_stiffness(qm, zm, wm, kl_qz, kappa = 0.5, alpha = 1.5e-08, beta = 1,
  eps = .Machine$double.eps)
}
\arguments{
\item{qm}{Output probabilty matrix.}

\item{zm}{Mixture matrix, weighted combination of input probability and
output probability matrix \code{qm}.}

\item{wm}{Output weight probability matrix.}

\item{kl_qz}{KL divergence between \code{qm} and \code{zm}. \code{qm} is the
reference probability.}

\item{kappa}{Mixture parameter. Should be a value between 0 and 1 and be the
same value used to produce the mixture matrix \code{zm}.}

\item{alpha}{Tail heaviness of the weighting function.}

\item{beta}{The precision of the weighting function.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix.
}
\description{
HSJSE Stiffness Function
}

