% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{hsnerv}
\alias{hsnerv}
\title{Heavy-tailed Symmetric Neighbor Retrieval Visualizer (HSNeRV)}
\usage{
hsnerv(lambda = 0.5, alpha = 0, beta = 1, eps = .Machine$double.eps,
  verbose = TRUE)
}
\arguments{
\item{lambda}{Weighting factor controlling the emphasis placed on precision
(set \code{lambda} to 0), versus recall (set \code{lambda} to 1). If set to
1, then the method is equivalent to t-SNE. Must be a value between 0 and 1.}

\item{alpha}{Tail heaviness. Must be greater than zero. When set to a small
value this method is equivalent to SSNE or SNeRV (depending on the value
of \code{lambda}. When set to one to one, this method behaves like
t-SNE/t-NeRV.}

\item{beta}{The precision of the function. Becomes equivalent to the
precision in the Gaussian distribution of distances as \code{alpha}
approaches zero.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
HSNeRV is a hybrid of \code{\link{snerv}} and \code{\link{hssne}}. It has
the \code{lambda} parameter of SNeRV, allowing for the control of precision
versus recall, and the \code{alpha} and \code{beta} parameters of HSSNE which
give the behavior of SSNE/SNeRV when \code{alpha} is close to zero,
and of t-SNE/t-NeRV when \code{alpha = 1}.

The probability matrix used in HSNeRV:

\itemize{
 \item{represents one probability distribution, i.e. the grand sum of the
 matrix is one.}
 \item{is symmetric, i.e. \code{P[i, j] == P[j, i]} and therefore the
 probabilities are joint probabilities.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix based on the weight matrix
 \code{wm}.}
}
}
\examples{
\dontrun{
# equivalent to SSNE
embed_prob(method = hsnerv(lambda = 1, alpha = 1.5e-8), ...)

# equivalent to t-SNE
embed_prob(method = hsnerv(lambda = 1, alpha = 1), ...)

# equivalent to default SNeRV
embed_prob(method = hsnerv(lambda = 0.5, alpha = 1.5e-8), ...)

# equivalent to default t-NeRV
embed_prob(method = hsnerv(lambda = 0.5, alpha = 1), ...)

}
}
\references{
Venna, J., Peltonen, J., Nybo, K., Aidos, H., & Kaski, S. (2010).
Information retrieval perspective to nonlinear dimensionality reduction for
data visualization.
\emph{Journal of Machine Learning Research}, \emph{11}, 451-490.
}
\seealso{
HSNeRV uses the \code{\link{nerv_cost}} cost function and the
  \code{\link{heavy_tail_weight}} similarity function for converting
  distances to probabilities.

The return value of this function should be used with the
  \code{\link{embed_prob}} embedding function.

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{hssne}},
  \code{\link{jse}}, \code{\link{mmds}},
  \code{\link{nerv}}, \code{\link{rasne}},
  \code{\link{rssne}}, \code{\link{rtsne}},
  \code{\link{sammon_map}}, \code{\link{smmds}},
  \code{\link{snerv}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hssne}}, \code{\link{jse}},
  \code{\link{nerv}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne}}, \code{\link{rssne}},
  \code{\link{rtsne}}, \code{\link{snerv}},
  \code{\link{ssne}}, \code{\link{tasne}},
  \code{\link{tnerv}}, \code{\link{tsne}}
}

