% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{hssne_stiffness}
\alias{hssne_stiffness}
\title{HSSNE Stiffness Function}
\usage{
hssne_stiffness(pm, qm, wm, alpha = 1.5e-08, beta = 1)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}

\item{wm}{Output weight probability matrix.}

\item{alpha}{Tail heaviness of the weighting function.}

\item{beta}{The precision of the weighting function.}
}
\value{
Stiffness matrix.
}
\description{
HSSNE Stiffness Function
}

