% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{iembed}
\alias{iembed}
\title{Initialize Embedding}
\usage{
iembed(method, xm = iris[1:50, 1:4],
  preprocess = make_preprocess(range_scale_matrix = TRUE, verbose = FALSE),
  init_inp = inp_from_perp(perplexity = 20, verbose = FALSE),
  init_out = out_from_PCA(verbose = FALSE), opt = gradient_descent())
}
\arguments{
\item{method}{Embedding method.}

\item{xm}{Input data matrix.}

\item{preprocess}{Preprocessing method.}

\item{init_inp}{Input initializer.}

\item{init_out}{Output initializer.}

\item{opt}{Optimizer.}
}
\value{
A list containing:
  \item{\code{inp}}{Initialized input}
  \item{\code{out}}{Initialized output}
  \item{\code{method}}{Initialized embedding method}
  \item{\code{opt}}{Initialized optimizer}
  \item{\code{report}}{Initialized report}
}
\description{
A convenience function which intializes an embedder with a small amount
of data. Useful for interactive exploration of e.g. gradients.
}

