% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{imp_kernel}
\alias{imp_kernel}
\title{Convert an existing similarity kernel to an importance weighted version.}
\usage{
imp_kernel(kernel)
}
\arguments{
\item{kernel}{Similarity kernel to convert.}
}
\value{
Kernel with importance weighting.
}
\description{
The original kernel is used to produce the weight matrix (or the derivative)
as usual, then multipled by the importance matrix. This should be stored
as \code{kernel$im}.
}

