% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{indegree_centrality}
\alias{indegree_centrality}
\title{Indegree Centrality}
\usage{
indegree_centrality(m)
}
\arguments{
\item{m}{Nearest neighbor graph matrix.}
}
\value{
Indegree centrality of \code{m}.
}
\description{
A node importance measure for directed graphs.
}
\details{
Sums the (possibly weighted) edges that are directed towards each node. For
probability matrices, column i is interpreted as containing the weighted
edges that are directed to node i.
}

