% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{init_embed}
\alias{init_embed}
\title{Embedding Initialziation}
\usage{
init_embed(xm, method, preprocess, init_inp, init_out, opt)
}
\arguments{
\item{xm}{A matrix or data frame to embed.}

\item{method}{Embedding method. Set by assigning the result value of one of
the configuration functions listed in
\code{\link{embedding_methods}}.}

\item{preprocess}{Input data preprocess callback. Does any necessary
filtering and scaling of the input data.}

\item{init_inp}{Input initializer. Creates the input data required by
embedding algorithm.}

\item{init_out}{Output initializer. Creates the initial set of output
coordinates.}

\item{opt}{Optimization method.}
}
\value{
A list containing:
  \item{\code{inp}}{Initialized input}
  \item{\code{out}}{Initialized output}
  \item{\code{method}}{Initialized embedding method}
  \item{\code{opt}}{Initialized optimizer}
  \item{\code{report}}{Initialized report}
}
\description{
Initializes all data, ready to be optimized.
}

