% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{init_out}
\alias{init_out}
\title{Output Initializer Wrapper}
\usage{
init_out(initializer)
}
\arguments{
\item{initializer}{Initializer function.}
}
\value{
Input data.
}
\description{
Wrapper function to creates the input data list and runs the specific
initializer function provided as a parameter.
}

