% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{initialize_optimizer}
\alias{initialize_optimizer}
\title{Optimizer initialization.}
\usage{
initialize_optimizer(opt, inp, out, method)
}
\arguments{
\item{opt}{Optimizer to initialize.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
a list containing:
\item{\code{opt}}{Optimizer with its components initialized.}
}
\description{
The direction, step size and update components of the optimizer may all
provide initialization methods to set their internal state before the first
optimization iteration This function will run those initialization functions
as part of the overall optimizer intialization step.
}

