% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{inp_from_dint_max}
\alias{inp_from_dint_max}
\title{Initialize With Perplexity that Maximizes Intrinsic Dimensionality}
\usage{
inp_from_dint_max(perplexities = NULL, input_weight_fn = exp_weight,
  modify_kernel_fn = NULL, verbose = TRUE)
}
\arguments{
\item{perplexities}{List of perplexities to use. If not provided, then
a series of perplexities in decreasing powers of two are used, starting
with the power of two closest to the number of observations in the dataset
divided by four.}

\item{input_weight_fn}{Weighting function for distances. It should have the
signature \code{input_weight_fn(d2m, beta)}, where \code{d2m} is a matrix
of squared distances and \code{beta} is a real-valued scalar parameter
which will be varied as part of the search to produce the desired
perplexity. The function should return a matrix of weights
corresponding to the transformed squared distances passed in as arguments.}

\item{modify_kernel_fn}{Function to create a new similarity kernel based
on the new perplexity. Will be called every time a new input probability
matrix is generated.}

\item{verbose}{If \code{TRUE} print message about initialization during the
embedding.}
}
\value{
Input initializer for use by an embedding function.
}
\description{
An initialization method for creating input probabilities.
}
\details{
This technique uses a lot of the same ideas from multiscale embedding,
except rather than use probabilities from multiple perplexities directly
in the input and output spaces, it selects for the input probability matrix
that which corresponds to the perplexity that gives the maximum intrinsic
dimensionality.

A list of perplexities may be provided to this function. Otherwise, the
perplexities used are decreasing powers of 2, e.g. 16, 8, 4, 2
with the maximum perplexity given by the formula:

\deqn{\lfloor{\log_{2}(N/4)}\rceil}{round(log2(N / 4)}

The perplexities are used in the order they are provided in the
\code{perplexities} vector. As soon as the intrinsic dimensionality begins
to decrease, the procedure is aborted and the probability matrix from the
previously calculated perplexity is used.

The parameter \code{modify_kernel_fn} can be used to modify the output kernel
based on the results of the perplexity calculation. If provided, then the
signature of \code{modify_kernel_fn} must be:

\code{modify_kernel_fn(inp, out, method)}

where \code{inp} is the input data, \code{out} is the current output data,
\code{method} is the embedding method.

This function will be called once for each perplexity, and an updated
kernel should be returned.
}
\seealso{
\code{inp_from_perps_multi} for caveats on using a
 non-default version of \code{input_weight_fn}.

Other sneer input initializers: \code{\link{inp_from_perps_multi}},
  \code{\link{inp_from_step_perp}}
}

