% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{inp_updated}
\alias{inp_updated}
\title{Update Embedding Internals When Input Data Changes}
\usage{
inp_updated(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
a list containing:
\item{inp}{Updated input data.}
\item{out}{Updated output data.}
\item{method}{Updated embedding method.}
}
\description{
Called when the input data changes, normally when the input probability
matrix is calculated. Some embedding method's output has an explicit
dependency on such data: for example, the JSE (\code{\link{jse}}) cost
function uses a mixture matrix of the input and output probability matrices.
}
\details{
Normally this will only be called once when the input data is first
initialized, but is using a technique where the input probabilities change,
such as multiscaling, then this should be called every time such a change
occurs.
}

