% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{input_initializers}
\alias{input_initializers}
\title{Input Initializers}
\description{
These methods deal with converting the input coordinates or distances
into the structures required by different embedding methods. For instance,
probability-based embedding methods require the construction of a probability
matrix from the input distances.
}
\examples{
# initializer that uses bisection search to create input probability
# distribution with a perplexity of 50 - pass to make_init_inp:
make_init_inp(prob_perp_bisect(perplexity = 50))

\dontrun{
# in turn, pass the result of make_init_inp to an embedding routine
embed_prob(method = tsne(),
           init_inp = make_init_inp(prob_perp_bisect(perplexity = 50)))
}
}
\seealso{
Input initializers should be passed to the
\code{\link{make_init_inp}} function.

Other sneer input initializers: \code{\link{prob_perp_bisect}}
}
\keyword{internal}

