% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_step.R
\name{jacobs_step_size}
\alias{jacobs_step_size}
\title{Calculate a new step size matrix based on the sign of the gradient versus
that of the previous step.}
\usage{
jacobs_step_size(gm, step_size, update, inc_fn, dec_fn)
}
\arguments{
\item{gm}{Gradient matrix.}

\item{step_size}{Step size for the previous iteration.}

\item{update}{update matrix for the previous iteration.}

\item{inc_fn}{Function to apply to \code{step_size} to increase its elements.}

\item{dec_fn}{Function to apply to \code{step_size} to decrease its elements.}
}
\value{
the new step size.
}
\description{
Calculate a new step size matrix based on the sign of the gradient versus
that of the previous step.
}
\details{
For each direction, the sign of the gradient is compared with that
of the update in the previous time step. If it's the same sign, increase the
step size; if the sign has changed, then it's assumed that the minimum was
missed and the current location brackets the minimum. In this case, the step
size is decreased.
}

