% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jse.R
\name{js_mixture}
\alias{js_mixture}
\title{Jensen-Shannon Mixture Matrix}
\usage{
js_mixture(pm, qm, kappa = 0.5)
}
\arguments{
\item{pm}{Probability matrix.}

\item{qm}{Probability matrix.}

\item{kappa}{Mixture parameter.}
}
\value{
Mixture matrix.
}
\description{
Creates a mixture matrix, \eqn{Z}, comprised of a linear weighted mixture of
\eqn{P} and \eqn{Q}:
}
\details{
\deqn{Z = \kappa P + (1 - \kappa)Q}{Z = kappa * P + (1 - kappa) * Q}
}
\references{
Lee, J. A., Renard, E., Bernard, G., Dupont, P., & Verleysen, M. (2013).
Type 1 and 2 mixtures of Kullback-Leibler divergences as cost functions in
dimensionality reduction based on similarity preservation.
\emph{Neurocomputing}, \emph{112}, 92-108.
}

