% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jse.R
\name{jse_cost_gr}
\alias{jse_cost_gr}
\title{JSE Cost Gradient}
\usage{
jse_cost_gr(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Gradient of the JSE cost.
}
\description{
Calculates the gradient of the JSE cost of an embedding with respect to the
output probabilities.
}

