% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jse.R
\name{jse_cost_norm}
\alias{jse_cost_norm}
\title{Normalized JSE Cost Function}
\usage{
jse_cost_norm(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
JSE divergence between \code{inp$pm} and \code{out$qm}.
}
\description{
A measure of embedding quality between input and output data.
}
\details{
Normalizes the JSE cost using the cost when the output probability matrix
\code{out$qm} is uniform. Also recalculates the mixture matrix \code{out$zm}
too. Intended to be used in the reporter function of sneer as a custom
normalized cost function, not as a main objective function.
}

