% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jse.R
\name{jse_fg}
\alias{jse_fg}
\title{JSE Cost}
\usage{
jse_fg(kappa = 0.5)
}
\arguments{
\item{kappa}{Mixture parameter. If set to 0, then the cost behaves like the
Kullback Leibler divergence. If set to 1, then the cost behaves like the
reverse KL divergence.}
}
\value{
JSE function and gradient.
}
\description{
Cost wrapper factory function.
}
\details{
Creates the a list containing the required functions for using the JSE cost
in an embedding.

Provides the cost function and its gradient (with respect to Q).
}
\seealso{
Other sneer cost wrappers: \code{\link{nerv_fg}},
  \code{\link{reverse_kl_fg}}
}

