% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jse.R
\name{jse_inp_update}
\alias{jse_inp_update}
\title{Update JSE Output Data When Input Data Changes}
\usage{
jse_inp_update(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
a list containing:
\item{inp}{Updated input data.}
\item{out}{Updated output data.}
\item{method}{Updated embedding method.}
}
\description{
Because JSE explicitly couples the input data to the output data via the
JS mixture matrix (\code{js_mixture}), whenever the input data
changes, this function should be called.
}

