% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{k_largest_ind}
\alias{k_largest_ind}
\title{Indexes of the k-largest numbers.}
\usage{
k_largest_ind(x, k)
}
\arguments{
\item{x}{Vector of numbers.}

\item{k}{Top k results to return}
}
\value{
Vector of the indexes of the \code{k} largest values in \code{x}.
}
\description{
Given a vector of numbers, return the indexes of the k-largest
values.
}

