% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{k_shared_nbrs_ind}
\alias{k_shared_nbrs_ind}
\title{Indexes of the shared neighbors between two distance vectors}
\usage{
k_shared_nbrs_ind(di, dj, k)
}
\arguments{
\item{di}{list of distances}

\item{dj}{list of distances}

\item{k}{The size of the shared neighborhood}
}
\value{
Vector of the indexes of the elements which are among both the
\code{k}-smallest values of \code{di} and the \code{k}-smallest
values of \code{dj}.
}
\description{
Return the indexes of shared k-closest neighbors in two lists of distances.
}

