% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{k_smallest_ind}
\alias{k_smallest_ind}
\title{Indexes of the k-smallest numbers in a vector.}
\usage{
k_smallest_ind(x, k)
}
\arguments{
\item{x}{Vector of numbers.}

\item{k}{Top k results to return}
}
\value{
Vector of the indexes of the \code{k} smallest values in \code{x}.
}
\description{
Given a vector of numbers, return the indexes of the k-smallest
values.
}

