% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{kernel_gr_fd}
\alias{kernel_gr_fd}
\title{Finite Difference Gradient of Kernel}
\usage{
kernel_gr_fd(kernel, d2m, diff = 1e-04)
}
\arguments{
\item{kernel}{A similarity kernel.}

\item{d2m}{Matrix of squared distances.}

\item{diff}{Step size to take in finite difference calculation.}
}
\value{
Gradient matrix.
}
\description{
Calculates the gradient of a similarity kernel by finite difference.
Only intended for testing purposes.
}

