% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{kl_cost_gr}
\alias{kl_cost_gr}
\title{Kullback Leibler Cost Gradients}
\usage{
kl_cost_gr(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Gradient of the KL divergence from \code{inp$pm} to \code{out$qm}.
}
\description{
Measures the gradient of the KL divergence of an embedding, with respect
to the probabilities of the ouput probabilities.
}

