% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{kl_divergence}
\alias{kl_divergence}
\title{Calculate the Kullback Leibler divergence between two matrices of
probabilities.}
\usage{
kl_divergence(pm, qm, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Probability Matrix. First probability in the divergence.}

\item{qm}{Probability Matrix. Second probability in the divergence.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
KL divergence between \code{pm} and \code{qm}.
}
\description{
If the matrices represent row probabilities, then the returned divergence
is the average over the divergence for each row.
}

