% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{kl_divergence_gr}
\alias{kl_divergence_gr}
\title{Kullback Leibler Gradient}
\usage{
kl_divergence_gr(pm, qm, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Probability Matrix. First probability in the divergence.}

\item{qm}{Probability Matrix. Second probability in the divergence.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Gradient of the KL divergence from \code{pm} to \code{qm}.
}
\description{
Calculates the gradient of the KL divergence with respect to the
probability Q in KL(P||Q).
}

