% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{kl_fg}
\alias{kl_fg}
\title{Kullback-Leibler Divergence Cost Wrapper Factory Function}
\usage{
kl_fg()
}
\value{
KL divergence function and gradient.
}
\description{
Cost wrapper factory function.
}
\details{
Creates the a list containing the required functions for using the Kullback
Leibler divergence, KL(P||Q), in an embedding.

Provides the cost function and its gradient (with respect to Q).
}
\seealso{
Other sneer cost wrappers: \code{\link{jse_fg}},
  \code{\link{nerv_fg}}, \code{\link{reverse_kl_fg}}
}

