% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{klqp_update_pjoint}
\alias{klqp_update_pjoint}
\title{Updates the Kullback Leibler Divergence for Joint Probabilities.}
\usage{
klqp_update_pjoint(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
\code{out} updated with the KL divergence from {\code{inp$pm}} to
\code{out$qm}.
}
\description{
Calculates and stores the KL divergence from P (input probabilities) to Q
(output probabilities) on the output data. Used by those embedding methods
where the KL divergence is used to calculate the stiffness matrix in a
gradient calculation (e.g. \code{snerv}).
}
\details{
Only appropriate for embedding methods that use joint probabilities.
}

