% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jse.R
\name{klqz_update_pjoint}
\alias{klqz_update_pjoint}
\title{Updates the Kullback Leibler Divergence Q||Z for Joint Probabilities.}
\usage{
klqz_update_pjoint(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
\code{out} updated with the KL divergence from {\code{out$qm}} to
\code{inp$zm}.
}
\description{
Calculates and stores the mixture probability Z and calculates the KL
divergence from Q (output probabilities) to Z on the output data. Used by
those embedding methods where this KL divergence is used to calculate the
stiffness matrix in a gradient calculation (e.g. \code{sjse}).
}
\details{
Only appropriate for embedding methods that use joint probabilities.
}

