% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{kruskal_stress}
\alias{kruskal_stress}
\title{Kruskal Stress Type-1}
\usage{
kruskal_stress(dxm, dym, eps = .Machine$double.eps)
}
\arguments{
\item{dxm}{Distance matrix.}

\item{dym}{Distance matrix, must be of the same dimensions as \code{dxm}.}

\item{eps}{Small floating point value to avoid numerical problems.}
}
\value{
Kruskal stress.
}
\description{
A measure of embedding quality between input and output distance matrices.
}
\details{
The Kruskal stress is normally used in non-metric MDS applications, but can
be defined for metric embeddings:

\deqn{K = \sqrt{\frac{STRESS}{\sum_{i<j} dym_{ij}^2}}}{K = sqrt(STRESS/sum(dym^2))}

where \eqn{dym} is the input distance matrix and \eqn{STRESS} is the
\code{\link{metric_stress_cost}} between the input and output distances. It
is dimensionless.
}

