% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{kruskal_stress_cost}
\alias{kruskal_stress_cost}
\title{Kruskal Stress Type-1 Cost Function}
\usage{
kruskal_stress_cost(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Kruskal stress.
}
\description{
A measure of embedding quality between input and output data.
}
\details{
The Kruskal stress is normally used in non-metric MDS applications, but can
be defined for metric embeddings:

\deqn{K = \sqrt{\frac{STRESS}{\sum_{i<j} d_{ij}^2}}}{K = sqrt(STRESS/sum(dij^2))}

where \eqn{d_{ij}}{dij} is the output distance between points \eqn{i} and
\eqn{j}, and \eqn{STRESS} is the \code{metric_stress} between the
input and output distance matrices. Unlike the raw STRESS, it is
dimensionless.

This cost function requires the following matrices to be defined:
\describe{
 \item{\code{inp$dm}}{Input distances.}
 \item{\code{out$dm}}{Output distances.}
}
}
\seealso{
Other sneer cost functions: \code{\link{jse_cost}},
  \code{\link{mean_relative_error_cost}},
  \code{\link{metric_sstress_cost}},
  \code{\link{metric_stress_cost}},
  \code{\link{nerv_cost}},
  \code{\link{normalized_stress_cost}},
  \code{\link{reverse_kl_cost}},
  \code{\link{rms_metric_stress_cost}},
  \code{\link{sammon_stress_cost}},
  \code{\link{sammon_stress_unnorm_cost}}
}

