% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{length_vec}
\alias{length_vec}
\title{Length of a vector (or matrix)}
\usage{
length_vec(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
Length (2-norm) of the matrix.
}
\description{
Length of a vector (or matrix)
}

