% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{linear_momentum}
\alias{linear_momentum}
\title{Linear momentum.}
\usage{
linear_momentum(max_iter, init_momentum = 0, final_momentum = 0.9)
}
\arguments{
\item{max_iter}{Number of iterations to scale the momentum over from
\code{init_momentum} to \code{final_momentum}.}

\item{init_momentum}{Momentum value for the first \code{switch_iter}
iterations.}

\item{final_momentum}{Momentum value after \code{switch_iter} iterations.}
}
\value{
Linear momentum update method, to be used by the optimizer.
}
\description{
Factory function for creating an optimizer update method.
}
\details{
Create a callback for the optimizer to use to update the embedding solution.
Update is in the form of a linear momentum function.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# update method:
make_opt(update = linear_momentum())
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{\link{optimizer_update}} for details
on the functions and values defined for this method.

Other sneer optimization update methods: \code{\link{nesterov_nsc_momentum}},
  \code{\link{no_momentum}}, \code{\link{step_momentum}}
}

