% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lreplace}
\alias{lreplace}
\title{Replace Members of a List}
\usage{
lreplace(l, ...)
}
\arguments{
\item{l}{A list.}

\item{...}{Named arguments to add to the list.}
}
\value{
The list \code{l} with the contents of the rest of the function
arguments added to it.
}
\description{
Updates a list with a set of named arguments, adding to, or replacing the
current contents.
}
\details{
Given a list and some named arguments, the contents of the list will
augmented by the named arguments. If the list already contains a value with
a name in the named arguments, it will be replaced.
}
\examples{
\dontrun{
 nato <- list(b = "bravo", m = "moto", x = "xray")
 lapd <- lreplace(nato,  b = "boy", m = "mary", s = "sam")
 # lapd is list(b = "boy", m = "mary", x = "xray", s = "sam")
}
}

