% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{make_after_step}
\alias{make_after_step}
\title{Create callback to be invoked after the solution is updated.}
\usage{
make_after_step(recenter = TRUE, direction_after_step_fn = NULL,
  step_size_after_step_fn = NULL, update_after_step_fn = NULL)
}
\arguments{
\item{recenter}{If \code{TRUE}, translate the embedded coordinates so they
are centered around the origin.}

\item{direction_after_step_fn}{Function provided by the direction method to
be invoked after the solution has been updated.}

\item{step_size_after_step_fn}{Function provided by the step size method to
be invoked after the solution has been updated.}

\item{update_after_step_fn}{Function provided by the update method to be
invoked after the solution has been updated.}
}
\value{
Callback to be invoked by the optimizer after the solution has been
updated.
}
\description{
The direction, step size and update methods of the optimizer may all have
to update their internal state after the solution has been updated (e.g.
updating step sizes or momentum terms according to a schedule). This function
accumulates all the provided function into a single callback that the
optimizer will invoke whenever it updates the solution.
}

