% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_embedding_plot}
\alias{make_embedding_plot}
\title{Embedding Plot Using \code{graphics} Library}
\usage{
make_embedding_plot(x, attr_name, label_fn, cex = 1)
}
\arguments{
\item{x}{Data frame containing label information for the embedded data.}

\item{attr_name}{Name of the label column in \code{x}.}

\item{label_fn}{Function with the signature \code{label_fn(labels)} where
\code{labels} is a vector of labels for each point in the data set. The
function should return a vector of labels suitable for displaying in the
plot.}

\item{cex}{Numeric \strong{c}haracter \strong{ex}pansion factor;
multiplied by \code{\link[graphics]{par}("cex")} yields the final
character size of the labels.}
}
\value{
Function which will take an output list, and produce a 2D plot of
the embedding.
}
\description{
Create a function which when invoked on a 2D matrix, plots the embedding
with color-coded labels.
}
\examples{
\dontrun{
# Create a plot function for the Iris dataset
iris_plot <- make_embedding_plot(iris, "Species")

# PCA on iris
pca_iris <- prcomp(iris[, 1:4], center = TRUE, retx = TRUE)
# view the first two scores:
iris_plot(pca_iris$x[, 1:2])

# TSNE on iris
tsne_iris <- embed_prob(iris[, 1:4])
# view the TSNE embedding
iris_plot(tsne_iris$ym)

# Same plot, with smaller labels
make_embedding_plot(iris, "Species", cex = 0.5)(tsne_iris$ym)
}
}
\seealso{
\code{\link{make_label}} for an example of a suitable
argument to \code{label_fn}.
}

