% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_embedding_qplot}
\alias{make_embedding_qplot}
\title{Embedded Coordinates Plot Using \code{ggplot2} Library}
\usage{
make_embedding_qplot(df, attr_name = "Label", size = 1, palette = "Set1",
  legend = TRUE, legend_rows = NULL)
}
\arguments{
\item{df}{Data frame containing label information for the embedded data.}

\item{attr_name}{Name of the label column in \code{df}.}

\item{size}{Size of the points.}

\item{palette}{String giving the name of a ColorBrewer Palette. To see the
available palettes run the function
\code{RColorBrewer::display.brewer.all()}. Although subject to change, the
following palettes from the "qualitative" subset are suggested:
\itemize{
  \item \code{Set1}
  \item \code{Set2}
  \item \code{Set3}
  \item \code{Pastel1}
  \item \code{Pastel2}
  \item \code{Dark2}
}}

\item{legend}{If \code{TRUE}, then the legend will be displayed. Set to
\code{FALSE} if there are a lot of separate categories that would appear
in the legend, which can result in the legend taking up more space than
the actual plot.}

\item{legend_rows}{If non-null, then sets the number of rows to display
the legend items in. If the legend is taking up too much space, you may
want to experiment with setting the number of rows manually, rather than
just setting the \code{legend} parameter to \code{FALSE}.}
}
\value{
Function with signature \code{plot_fn(ym)} where \code{ym} is a
2D matrix of embedded coordinates of data set \code{x}. On invocation, the
data will be plotted.
}
\description{
Creates a ggplot2 function for embeddings of a given dataset with coloring
from RColorBrewer palettes.
}
\note{
Use of this function requires that the \code{ggplot2} and
\code{RColorBrewer} packages be installed.
}
\examples{
\dontrun{
# make two different embeddings of the iris dataset
prcomp_iris <- prcomp(iris[, 1:4], scale. = TRUE, retx = TRUE)
mds_iris <- embed_dist(iris[, 1:4], method = mmds(eps = 1e-4),
                       opt = bold_nag(),
                       init_out = out_from_matrix(prcomp_iris$x[, 1:2]),
                       max_iter = 40)
iris_view <- make_embedding_qplot(iris, "Species")
iris_view(prcomp_iris$x)
iris_view(mds_iris$ym)
}
}
\seealso{
More information on ColorBrewer is available at its website,
\url{http://www.colorbrewer2.org}.
}

