% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make_final_reporter}
\alias{make_final_reporter}
\title{Report Embedding Quality After Embedding Terminates}
\usage{
make_final_reporter(labels = NULL)
}
\arguments{
\item{labels}{Vector of labels associated with the dataset, one for each
observation, in the same order that the data is presented to the embedding
algorithm.}
}
\value{
A function suitable for calling after embedding, which will in turn
return the output data, with the quality metrics appended in a list called
\code{quality}.
}
\description{
Calculates the AUC of the log RNX metric, the scaled neighbor retrieval
calculated for multiple neighbor values, with a logarithmic weighting to
favour smaller neighbor values.
}
\details{
Additionally, if a set of labels are provided for the dataset, the average
ROC AUC and PR AUC will be calculated for each observation in the dataset,
using its label as the positive class. If there are more than two labels
in the dataset, then the other classes are all treated as negative.
}
\seealso{
\code{\link{pr_auc}}, \code{\link{roc_auc}} and
\code{\link{rnx_auc}} for definitions of these metrics.
}

