% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{make_init_inp_dist}
\alias{make_init_inp_dist}
\title{Input Initialization for Distance Embedding}
\usage{
make_init_inp_dist()
}
\value{
A callback function with signature \code{fn(xm)} where \code{xm}
is the input coordinates or a distance matrix. This function will return
a list containing:
 \item{\code{xm}}{Input coordinates if these were provided.}
 \item{\code{dm}}{Input distances.}
}
\description{
Factory function for input initialization callbacks. During input
initialization, the high dimensional distances or coordinates will be used
to generate the a distance matrix.
}
\examples{
# Should be passed to the init_inp argument of an embedding function:
\dontrun{
 embed_dist(init_inp = make_init_inp_dist(), ...)
}
}
\seealso{
\code{\link{embed_dist}} for how to use this function for
configuring an embedding.
}

