% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{make_init_out}
\alias{make_init_out}
\title{Output Initialization}
\usage{
make_init_out(k = 2, init_config = NULL, stdev = 1e-04,
  from_PCA = FALSE, from_input_name = NULL, mat_name = "ym",
  verbose = TRUE)
}
\arguments{
\item{k}{Number of output dimensions. For 2D visualization this is always 2.}

\item{init_config}{Input data to initialize the coordinates from. Must
be a matrix with the same dimensions as the desired output coordinates.}

\item{stdev}{The standard deviation of the gaussian used to generate the
output coordinates. Used only if \code{from_PCA} is \code{FALSE},
\code{init_config} is \code{NULL} and \code{from_input_name} is
\code{NULL}.}

\item{from_PCA}{If \code{TRUE}, then the first \code{k} scores of the PCA
of the input coordinates are used to initialize the embedded coordinates.
Input coordinates are centered but not scaled before the PCA is carried out.
This option can only be used if the input data is in the form of coordinates,
and not a distance matrix.}

\item{from_input_name}{Name of a matrix in the input data list \code{inp}
which contains suitable initialization data. Must be a matrix with the same
dimensions as the desired output coordinates.}

\item{mat_name}{Name of the matrix on the output list to contain the
initialized coordinates. Ensure that if you change this from the default that
other callbacks that need this information (e.g. optimization routines,
plotting function in reporter callbacks) are also passed the same value.}

\item{verbose}{If \code{TRUE} information about the initialization will be
logged to screen.}
}
\value{
Callback to be used by the embedding routine to initialize
the output data. The callback has the signature \code{init_out(inp)} where
\code{inp} is the initialized input data list. The return value of the
callback is the initialized output data, a list containing:
\item{\code{ym}}{A matrix containing the initialized output coordinates. The
 name of the matrix can be changed by setting the \code{mat_name} parameter.}
Depending on the type of embedding carried out, the \code{out} list will
accumulate other auxiliary data associated with the embedded coordinates,
e.g. distances, probabilities, divergences or other data required for
stiffness calculations.
}
\description{
Factory function that creates a callback that used by the embedding routine.
When invoked, it will initialize the output data.
}
\examples{
# Initialize from PCA scores matrix (normally a decent starting point and
# reproducible)
make_init_out(from_PCA = TRUE)

# Initialize from random small distances (used in t-SNE paper)
make_init_out(stdev = 1e-4)

# Should be passed to the init_out argument of an embedding function:
\dontrun{
 embed_prob(init_out = make_init_out(from_PCA = TRUE), ...)
}
}
\seealso{
\code{\link{embed_prob}} for how to use this function for configuring
an embedding.
}

