% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_iris_plot}
\alias{make_iris_plot}
\title{Iris dataset plot factory function.}
\usage{
make_iris_plot(num_label_chars = 1)
}
\arguments{
\item{num_label_chars}{The number of characters to plot from the label
for each data point.}
}
\value{
Function for plotting the embedded iris data set.
}
\description{
Creates a function which can be used to visualize embeddings of the iris
dataset.
}
\details{
Wrapper function for visualizing the iris data set. If embedding the iris
data set, the result of invoking this can be passed to the \code{plot_fn}
parameter of the \code{make_reporter} function.
}
\seealso{
Other sneer plot functions: \code{\link{make_plot}}
}

