% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_label}
\alias{make_label}
\title{Create a label function for 2D embedding plot.}
\usage{
make_label(num_label_chars = 1)
}
\arguments{
\item{num_label_chars}{The number of characters to plot from the label
for each data point.}
}
\value{
a function which can be passed to the \code{label_fn} parameter
of the \code{make_plot} function.
}
\description{
Even quite short labels can create a crowded looking embedding plot.
Use this function to only print the first few characters of each label.
}

