% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{make_opt_init}
\alias{make_opt_init}
\title{Create optimizer initialization.}
\usage{
make_opt_init(direction_fn = NULL, step_size_fn = NULL, update_fn = NULL)
}
\arguments{
\item{direction_fn}{Function provided by the direction method to be invoked
during initialization.}

\item{step_size_fn}{Function provided by the step size method to be invoked
during initialization.}

\item{update_fn}{Function provided by the update method to be invoked during
initialization.}
}
\value{
Callback to be invoked by the optimizer during initialization.
}
\description{
Create callback to be invoked when the optimizer is initialized.
}
\details{
The direction, step size and update methods of the optimizer may all provide
initialization methods to set their internal state before the first
optimization step. This function accumulates all the provided function into
a single callback that the optimizer will invoke whenever it updates the
solution.
}

