% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_external.R
\name{make_optim_f}
\alias{make_optim_f}
\title{Convert Cost Function}
\usage{
make_optim_f(opt, inp, method, iter)
}
\arguments{
\item{opt}{Optimizer}

\item{inp}{Input data.}

\item{method}{Embedding method.}

\item{iter}{Iteration number.}
}
\value{
Cost function taking a vector of parameters and returning the scalar
 cost.
}
\description{
This function takes the current state of the embedding and returns a
cost function which can be used by \code{\link[stats]{optim}} as the
\code{fn} parameter.
}

