% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_external.R
\name{make_optim_g}
\alias{make_optim_g}
\title{Convert Gradient Function}
\usage{
make_optim_g(opt, inp, method, iter)
}
\arguments{
\item{opt}{Optimizer}

\item{inp}{Input data.}

\item{method}{Embedding method.}

\item{iter}{Iteration number.}
}
\value{
Gradient function taking a vector of parameters and returning
 a vector of gradient values.
}
\description{
This function takes the current state of the embedding and returns a
gradient function which can be used by \code{\link[stats]{optim}} as the
\code{gr} parameter.
}

