% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_plot}
\alias{make_plot}
\title{Create a plotting function.}
\usage{
make_plot(x, attr_name, label_fn = function(labels) {     labels },
  mat_name = "ym")
}
\arguments{
\item{x}{Data frame containing label information for the embedded data.}

\item{attr_name}{Name of the label column in \code{x}.}

\item{label_fn}{Function with the signature \code{label_fn(labels)} where
\code{labels} is a vector of labels for each point in the data set, and
the function returns a vector of labels suitable for displaying in the plot
(e.g. with each label truncated so that only the first three characters)
will be used.}

\item{mat_name}{The name of the matrix containing the embedded data in the
output list \code{out} which will be passed to the plot function.}
}
\value{
Function which will take an output list, and produce a 2D plot of
the embedding.
}
\description{
Factory function for a plotting callback which can be used by the reporter
function of an embedding to plot the current (two-dimensional embedding).
}
\examples{
# For s1k dataset, plot 2D embedding with "Label" factor to identify each
# point on the plot
make_plot(s1k, "Label")

# For iris dataset, plot 2D embedding with first two characters of the
# "Species" factor to identify each point on the plot
make_plot(iris, "Species", make_label(2))

# Should be passed to the plot_fn argument of the reporter factory function:
\dontrun{
 embed_sim(reporter = make_reporter(report_every = 100, calc_stress = TRUE,
                                    plot_fn = make_plot(iris, "Species")),
                                     ...)
}
}
\seealso{
\code{\link{make_reporter}} for how to use this function for
configuring visualization of the progress of an embedding, and
\code{\link{make_iris_plot}} for a helper function when using the iris
dataset that involves less typing.
}

