% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{make_preprocess}
\alias{make_preprocess}
\title{Preprocessing}
\usage{
make_preprocess(range_scale_matrix = FALSE, range_scale = FALSE, rmin = 0,
  rmax = 1, auto_scale = FALSE, whiten = FALSE, zwhiten = FALSE,
  whiten_dims = 30, scale_distances = FALSE, verbose = TRUE)
}
\arguments{
\item{range_scale_matrix}{If \code{TRUE}, input coordinates will be scaled
so that elements are between (\code{rmin}, \code{rmax}).}

\item{range_scale}{If \code{TRUE}, input coordinates will be scaled}

\item{rmin}{Minimum value if using \code{range_scale_matrix} and
\code{range_scale}.}

\item{rmax}{Maximum value if using \code{range_scale_matrix} and
\code{range_scale}.}

\item{auto_scale}{If \code{TRUE}, input coordinates will be centered and
scaled so that each column has a mean of 0 and a standard deviation of 1.}

\item{whiten}{If \code{TRUE}, input coordinates will be whitened.}

\item{zwhiten}{If \code{TRUE}, input coordinates will be whitened using the
ZCA transform.}

\item{whiten_dims}{Number of components to use in the whitening preprocess.
Only use if \code{whiten} or \code{zwhiten} is \code{TRUE}.}

\item{scale_distances}{If \code{TRUE}, the distance matrix will be scaled
such that the mean distance is 1.}

\item{verbose}{If \code{TRUE}, log information about preprocessing.}
}
\value{
A preprocessor for use by the embedding routine.
}
\description{
Creates a callback for use in the embedding routine. The resulting
preprocessor will apply the specified transformations to the input data,
and, if necessary, convert the input to a distance matrix.
}
\details{
Pass the result of this factory function to the preprocess argument of an
embedding function, e.g. \code{embed_prob}. Note that most of the options
are applicable only if the input data is in the form of coordinates. In that
case, the distance matrix will also be generated as part of the
preprocessing after any coordinate-specific processing has occurred.
Then, any distance matrix-specific preprocessing will be applied.

A variety of common preprocessing options are available. The range scaling
and auto scaling options are mutually exclusive, but can be combined with
whitening if required.
}
\examples{
# Scale the input data so the smallest element is 0, and the largest is 1.
make_preprocess(range_scale_matrix = TRUE)

# Scale the input data so the smallest element in each column is -1, and
# the largest is 1.
make_preprocess(range_scale = TRUE, rmin = -1)

# Autoscale each column in the input data, to mean 0 and the sd 1.
make_preprocess(auto_scale = TRUE)

# Whiten the data after range scaling the matrix.
make_preprocess(range_scale_matrix = TRUE, whiten = TRUE)

# Whiten the data using 10 components.
make_preprocess(range_scale_matrix = TRUE, whiten = TRUE, whiten_dims = 10)

# Whiten the data with the ZCA technique, using 10 components.
make_preprocess(range_scale_matrix = TRUE, zwhiten = TRUE, whiten_dims = 10)

# Should be passed to the preprocess argument of an embedding function:
\dontrun{
 embed_prob(preprocess = make_preprocess(range_scale = TRUE, rmin = -1), ...)
}
}
\seealso{
\code{\link{embed_prob}} for how to use this function for
  configuring an embedding.
}

