% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_qplot}
\alias{make_qplot}
\title{Embedding Plot Using \code{ggplot2} Library}
\usage{
make_qplot(x, attr_name, mat_name = "ym")
}
\arguments{
\item{x}{Data frame containing label information for the embedded data.}

\item{attr_name}{Name of the label column in \code{x}.}

\item{mat_name}{The name of the matrix containing the embedded data in the
output list \code{out} which will be passed to the plot function.}
}
\value{
Function with signature \code{plot_fn(out)} where \code{out} is
a return value from a sneer embedding function. On invocation, the
data will be plotted.
}
\description{
Creates a function which can be used to visualize embeddings from sneer
output results for a particular dataset using ggplot2.
}
\note{
Use of this function requires that the \code{ggplot2} package be
installed.
}
\examples{
\dontrun{
mds_iris <- embed_dist(iris[, 1:4])
iris_view <- make_qplot(iris, "Species")
iris_view(mds_iris)
}
}

