% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make_quality_reporter}
\alias{make_quality_reporter}
\title{Create Embedding Quality Reporter}
\usage{
make_quality_reporter(fs, labels)
}
\arguments{
\item{fs}{Vector of quality functions.}

\item{labels}{Vector of labels to apply to the data.}
}
\value{
A function suitable for calling after embedding, which will in turn
return the output data, with the quality metrics appended in a list called
\code{quality}.
}
\description{
Quality functions should have the signature \code{f(inp, out)} where
\code{inp} is the input data and \code{out} is the output data, and return
a list consisting of: \code{name}, the name of the quality measure; and
\code{value}, the value of the quality measure.
}

