% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricks.R
\name{make_tricks}
\alias{make_tricks}
\title{Create tricks callback.}
\usage{
make_tricks(early_exaggeration = TRUE, P_exaggeration = 4,
  exaggeration_off_iter = 50, verbose = TRUE)
}
\arguments{
\item{early_exaggeration}{If \code{TRUE}, then apply early exaggeration.}

\item{P_exaggeration}{The size of the exaggeration.}

\item{exaggeration_off_iter}{Iteration step to turn exaggeration off.}

\item{verbose}{If \code{TRUE} log messages about trick application.}
}
\value{
Tricks callback.
}
\description{
Creates a callback which the embedding routine will call before each
optimization step. The callback has the signature
\code{inp, out, method, opt, iter} where:
\itemize{
 \item \code{inp} Input data.
 \item \code{out} Output data.
 \item \code{method} Embedding method.
 \item \code{opt} Optimizer.
 \item \code{iter} Iteration number.
}
and returns a list containing:
\itemize{
 \item \code{inp} Updated input data.
 \item \code{out} Updated output data.
 \item \code{method} Updated embedded method.
 \item \code{opt} Updated optimizer.
}
}
\examples{
# Use early exaggeration as described in the t-SNE paper
make_tricks(early_exaggeration = TRUE, P_exaggeration = 4,
            exaggeration_off_iter = 50)

# Should be passed to the tricks argument of an embedding function:
\dontrun{
 embed_sim(tricks = make_tricks(early_exaggeration = TRUE,
                                P_exaggeration = 4,
                                exaggeration_off_iter = 50), ...)
}
}
\seealso{
\code{\link{embed_sim}} for how to use this function for configuring
an embedding, and \code{\link{tsne_tricks}} for a convenience function that
supplies the parameters used in the t-SNE paper.
}

