% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{make_update_out_ms}
\alias{make_update_out_ms}
\title{Output Update Factory Function for Multiscale Probability}
\usage{
make_update_out_ms()
}
\value{
The output update function which will be invoked as part of the
embedding.
}
\description{
In a multiscale embedding, the distance matrix is calculated once
when the coordinates change, and then multiple similarity kernels
are used to produce multiple weights matrices. These in turn
create multiple probability matrices, which are averaged to create
the final probability matrix.
}
\details{
The weights and probabilities associated with the lth kernel is stored on
the \code{wms[[l]]} and \code{qms[[l]]} list on \code{out}. The averaged
probability matrix is stored as \code{out$qm} as usual.
}

