% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{map2color}
\alias{map2color}
\title{Map Numbers to Colors}
\usage{
map2color(x, name = "Blues", n, limits = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{name}{Name of the ColorBrewer palette.}

\item{n}{Number of unique colors to map values in \code{x} to.}

\item{limits}{The range that the colors should map over. If not specified,
then the range of \code{x}. This is useful if there is some external
absolute scale that should be used.}
}
\description{
Maps a numeric vector to an equivalent set of colors based on the specified
ColorBrewer palette. Use the diverging or sequential.
}
\details{
Sequential palettes names:
 Blues BuGn BuPu GnBu Greens Greys Oranges OrRd PuBu PuBuGn PuRd Purples
 RdPu Reds YlGn YlGnBu YlOrBr YlOrRd
Diverging palette names:
 BrBG PiYG PRGn PuOr RdBu RdGy RdYlBu RdYlGn Spectral
}
\note{
Use of this function requires that the \code{RColorBrewer} packages be
 installed.

This function is based off a Stack Overflow answer by user "Dave X":
 \url{http://stackoverflow.com/a/18749392}
}
\examples{
\dontrun{
# Plot Iris dataset sepal width vs length, colored by petal length, using
# 20 colors ranging from Purple to Green (PRGn):
plot(iris[, c("Sepal.Length", "Sepal.Width")],
 col = map2color(iris$Petal.Length, "PRGn", 20), pch = 20, cex = 1.5)
}
}
\seealso{
More information on ColorBrewer is available at its website,
\url{http://www.colorbrewer2.org}.
}

