% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_external.R
\name{mat_to_par}
\alias{mat_to_par}
\title{Convert Matrix to 1D Parameter Vector}
\usage{
mat_to_par(mat)
}
\arguments{
\item{mat}{Matrix to convert.}
}
\value{
Matrix in vector form.
}
\description{
This function takes a matrix used by sneer internally (e.g. output
coordinates or gradient matrix) and converts into a 1D vector, as used
by \code{\link[stats]{optim}}. The matrix is converted columnwise.
}

