% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{mean_relative_error}
\alias{mean_relative_error}
\title{Mean Relative Error of Embedding}
\usage{
mean_relative_error(dxm, dym)
}
\arguments{
\item{dxm}{Distance matrix.}

\item{dym}{Distance matrix, must be of the same dimensions as \code{dxm}.}
}
\value{
Mean relative error.
}
\description{
A measure of embedding quality between input and output distance matrices.
}
\details{
The mean relative error of an embedding is defined as:

\deqn{MRE=\frac{2}{n(n-1)}\sum_{i<j}\left|\frac{dxm_{ij} - dym_{ij}}{r_{ij}}\right|}{MRE = (2/(n*(n-1)))*sum(abs((dxm - dym)/dxm))}

where \eqn{dxm} is the input distance matrix, \eqn{dym} is the corresponding
output distance and \eqn{n} is the number of points in the distance matrix.
The constant pre-sum factor is a normalization to account for the number of
unique distances (ignoring self-distances) in the matrix. You can treat the
MRE as a percentage by multiplying by 100.
}

