% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{mean_relative_error_cost}
\alias{mean_relative_error_cost}
\title{Mean Relative Error of Embedding Cost Function}
\usage{
mean_relative_error_cost(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Mean Relative Error.
}
\description{
A measure of embedding quality between input and output data.
}
\details{
The mean relative error of an embedding is defined as:

\deqn{MRE=\frac{2}{n(n-1)}\sum_{i<j}\left|\frac{r_{ij} - d_{ij}}{r_{ij}}\right|}{MRE = (2/(n*(n-1)))*sum(abs((rij- dij)/rij))}

where \eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j}, \eqn{d_{ij}}{dij} is the corresponding output distance and \eqn{n}
is the number of points in the distance matrix. The constant pre-sum factor
is a normalization to account for the number of unique distances (ignoring
self-distances) in the matrix. You can treat the MRE as a percentage
by multiplying by 100.

This cost function requires the following matrices to be defined:
\describe{
 \item{\code{inp$dm}}{Input distances.}
 \item{\code{out$dm}}{Output distances.}
}
}
\seealso{
Other sneer cost functions: \code{\link{kl_cost}},
  \code{\link{kruskal_stress_cost}},
  \code{\link{metric_sstress_cost}},
  \code{\link{metric_stress_cost}},
  \code{\link{nerv_cost}},
  \code{\link{normalized_stress_cost}},
  \code{\link{reverse_kl_cost}},
  \code{\link{rms_metric_stress_cost}},
  \code{\link{sammon_stress_cost}}
}

