% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{metric_sstress}
\alias{metric_sstress}
\title{Squared Distance Metric Stress Function (SSTRESS)}
\usage{
metric_sstress(dxm, dym)
}
\arguments{
\item{dxm}{Input distance matrix.}

\item{dym}{Embedded distance matrix.}
}
\value{
Metric stress.
}
\description{
A measure of embedding quality between distance matrices.
}
\details{
SSTRESS is a cost function used in metric MDS, related to the
\code{\link{metric_stress}}. It is defined as:

\deqn{SSTRESS = \sum_{i<j} ((dx_{ij}^2 - dy_{ij})^2)^2}{SSTRESS = sum(dxij^2-dyij^2)^2}

\eqn{dx_{ij}}{dxij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{dy_{ij}}{dij} is the corresponding output distance.
SSTRESS differs from STRESS in the distances being squared in the loss
function.
}

