% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{metric_sstress_cost}
\alias{metric_sstress_cost}
\title{Squared Distance STRESS Cost Function (SSTRESS)}
\usage{
metric_sstress_cost(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Metric stress.
}
\description{
A measure of embedding quality between input and output data.
}
\details{
SSTRESS is a cost function used in metric MDS, related to the
\code{\link{metric_stress}}. It is defined as:

\deqn{SSTRESS = \sum_{i<j} ((r_{ij}^2 - d_{ij})^2)^2}{SSTRESS = sum(rij^2-dij^2)^2}

\eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{d_{ij}}{dij} is the corresponding output distance.
SSTRESS differs from STRESS in the distances being squared in the loss
function.

This cost function requires the following matrices to be defined:
\describe{
 \item{\code{inp$dm}}{Input distances.}
 \item{\code{out$dm}}{Output distances.}
}
}
\seealso{
Other sneer cost functions: \code{\link{kl_cost}},
  \code{\link{kruskal_stress_cost}},
  \code{\link{mean_relative_error_cost}},
  \code{\link{metric_stress_cost}},
  \code{\link{normalized_stress_cost}},
  \code{\link{rms_metric_stress_cost}},
  \code{\link{sammon_stress_cost}}
}

