% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{metric_stress}
\alias{metric_stress}
\title{Metric Stress (STRESS)}
\usage{
metric_stress(dxm, dym)
}
\arguments{
\item{dxm}{Input distance matrix.}

\item{dym}{Embedded distance matrix.}
}
\value{
Metric stress.
}
\description{
A measure of embedding quality between distance matrices.
}
\details{
The metric stress, known in the MDS literature as STRESS, is the residual
sum of squares between two distance matrices.

\deqn{STRESS = \sum_{i<j} (dx_{ij} - dy_{ij})^2}{STRESS = sum(dxij-dyij)^2}

\eqn{dx_{ij}}{dxij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{dy_{ij}}{dij} is the corresponding output distance.
}

