% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{metric_stress_cost}
\alias{metric_stress_cost}
\title{Metric stress cost function, STRESS.}
\usage{
metric_stress_cost(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Metric stress.
}
\description{
A measure of embedding quality between input and output data.
}
\details{
The metric stress, known in the MDS literature as STRESS, is the residual
sum of squares between the input and output distances:

\deqn{STRESS = \sum_{i<j} (r_{ij} - d_{ij})^2}{STRESS = sum(rij-dij)^2}

\eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{d_{ij}}{dij} is the corresponding output distance.

This cost function requires the following matrices to be defined:
\describe{
 \item{\code{inp$dm}}{Input distances.}
 \item{\code{out$dm}}{Output distances.}
}
}
\seealso{
Other sneer cost functions: \code{\link{kl_cost}},
  \code{\link{kruskal_stress_cost}},
  \code{\link{mean_relative_error_cost}},
  \code{\link{metric_sstress_cost}},
  \code{\link{normalized_stress_cost}},
  \code{\link{rms_metric_stress_cost}},
  \code{\link{sammon_stress_cost}}
}

