% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{momentum_scheme}
\alias{momentum_scheme}
\title{Momentum Update Method}
\usage{
momentum_scheme(mu_fn, msg_fn = NULL, verbose = TRUE)
}
\arguments{
\item{mu_fn}{Momentum function with signature \code{mu_fn(iter)} where
\code{iter} is the iteration number and the function returns the momentum
value for that iteration.}

\item{msg_fn}{Function with signature \code{msg_fn(iter)} where
\code{iter} is the iteration number. Called on each iteration if
\code{verbose} is \code{TRUE}. Can be used to provide information.}

\item{verbose}{If \code{TRUE}, then \code{msg_fn} will be invoked on each
iteration.}
}
\value{
Momentum scheme update method.
}
\description{
Create an update method using a momentum scheme
}

