% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{momentum_update}
\alias{momentum_update}
\title{Solution update with momentum term.}
\usage{
momentum_update(opt, inp, out, method)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Update matrix, consisting of gradient update and momentum term.
}
\description{
Carries out a solution update using a momentum term in addition to the
gradient update.
}

