% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{momentum_update_term}
\alias{momentum_update_term}
\title{Momentum term of an update}
\usage{
momentum_update_term(opt, inp, out, method, iter)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}

\item{iter}{Iteration number.}
}
\value{
Momentum update.
}
\description{
Calculates the momentum term of an update.
}

