% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{ms_perps}
\alias{ms_perps}
\title{Perplexity Values for Multiscale Embedding}
\usage{
ms_perps(x, num_perps = max(round(log2(nrow(x)/4)), 1))
}
\arguments{
\item{x}{Data frame or matrix.}

\item{num_perps}{Number of perplexities.}
}
\value{
Vector containing \code{num_perps} perplexities.
}
\description{
Returns a vector of perplexity values scaled according to the size of the
data frame or matrix suitable for use in a step wise scaling of perplexity.
}
\details{
The perplexities will be returned in descending order, starting at the power
of two closest to N/4, where N is the number of observations in the data set,
and ending at 2, with the perplexities in taking up decreasing powers of 2.
}

