% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{nbr_pres}
\alias{nbr_pres}
\title{Neighborhood Preservation Between Distance Matrices}
\usage{
nbr_pres(din, dout, k)
}
\arguments{
\item{din}{Distance matrix. The "ground truth" or reference distances.}

\item{dout}{Distance matrix. A set of distances to compare to the reference
distances.}

\item{k}{The size of the neighborhood, where k is the number of neighbors to
include in the neighborhood.}
}
\value{
Vector of preservation values, one for each row of the distance
 matrix.
}
\description{
Calculates the neighborhood preservation for each observation in a dataset,
represented by two distance matrices. The first matrix is the "ground truth",
the second being the estimation or approximation.
The neighborhood preservation is calculated for each row where each element
d[i, j] is taken to be the distance between observation i and j.
}
\details{
The neighborhood preservation can vary between 0 (no neighbors in common)
and 1 (perfect preservation). However, random performance gives an
approximate value of k / (k - 1), where k is the size of the neighborhood.
}
\note{
This is not a very efficient way to calculate the preservation if you
 want to calculate the value for multiple values of \code{k}. For more
 global measures of preservation, see \code{\link{rnx_auc_embed}}.
}
