% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{nbr_pres_i}
\alias{nbr_pres_i}
\title{Neighborhood Preservation}
\usage{
nbr_pres_i(di, dj, k)
}
\arguments{
\item{di}{Vector of distances.}

\item{dj}{Vector of distances.}

\item{k}{Size of the neighborhood to consider.}
}
\value{
The number of shared neighbors in the equivalent neighbor lists of
\code{di} and \code{dj}.
}
\description{
For the K nearest neighbors in one set of distances, returns the number of
those neighbors which are also K nearest neighbors in another list,
normalized with respect to K.
}
\details{
The neighborhood preservation can vary between 0 (no neighbors in common)
and 1 (perfect preservation). However, random performance gives an
approximate value of K / K - 1.
}

