% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{nerv_cost_gr}
\alias{nerv_cost_gr}
\title{NeRV Cost Gradient}
\usage{
nerv_cost_gr(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Gradient of the NeRV cost.
}
\description{
Calculates the gradient of the NeRV cost of an embedding with respect to the
output probabilities.
}

