% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{nerv_fg}
\alias{nerv_fg}
\title{NeRV Cost}
\usage{
nerv_fg(lambda = 0.5)
}
\arguments{
\item{lambda}{Weighting factor controlling the emphasis placed on precision
(set \code{lambda} to 0), versus recall (set \code{lambda} to 1). If set to
1, then the cost behaves like the Kullback Leibler divergence. If set to 0,
the cost behaves like the "reverse" KL divergence.}
}
\value{
NeRV cost function and gradient.
}
\description{
Cost wrapper factory function.
}
\details{
Creates the a list containing the required functions for using the NeRV cost
in an embedding.

Provides the cost function and its gradient (with respect to Q).
}
\seealso{
Other sneer cost wrappers: \code{\link{jse_fg}},
  \code{\link{kl_fg}}, \code{\link{reverse_kl_fg}}
}

