% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{nerv_inp_update}
\alias{nerv_inp_update}
\title{NeRV input update function}
\usage{
nerv_inp_update(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
List containing the update embedding method, where the kernel beta
parameters have been updated to be the same as that for the \code{input}
data.
}
\description{
Update function to run when the input data has changed. Unlike most other
probability-based embedding methods, NeRV sets the output kernels to have
the same parameters as the equivalent input kernels (in the case of NeRV, the
exponential parameter).
}

