% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{nerv_stiffness}
\alias{nerv_stiffness}
\title{NeRV Stiffness Function}
\usage{
nerv_stiffness(pm, qm, rev_kl, lambda = 0.5, beta = 1,
  eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Input probability matrix.}

\item{qm}{Output probabilty matrix.}

\item{rev_kl}{"Reverse" KL divergence between \code{pm} and \code{qm}.}

\item{lambda}{NeRV weighting factor controlling the emphasis placed on
precision versus recall.}

\item{beta}{Precision of the weighting function.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix.
}
\description{
NeRV Stiffness Function
}

