% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{nesterov_gradient}
\alias{nesterov_gradient}
\title{Nesterov Accelerated Gradient.}
\usage{
nesterov_gradient()
}
\value{
NAG method for gradient calculation.
}
\description{
Factory function for creating an optimizer gradient method.
}
\details{
Calculates the gradient according to the Nesterov Accelerated Gradient
method.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# gradient method:
make_opt(gradient = nesterov_gradient())
}
\references{
Sutskever, I., Martens, J., Dahl, G. and Hinton, G. E.
On the importance of momentum and initialization in deep learning.
30th International Conference on Machine Learning, Atlanta, USA, 2013.
JMLR: W&CP volume 28.
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{\link{optimizer_gradient}} for details
on the functions and values defined for this method.

Other sneer optimization gradient methods: \code{\link{classical_gradient}}
}

