% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{nn_graph}
\alias{nn_graph}
\title{Convert a Probability Matrix to a Nearest Neighbor Adjacency Graph.}
\usage{
nn_graph(m, k)
}
\arguments{
\item{m}{Probability matrix.}

\item{k}{The number of neighbors.}
}
\value{
Nearest neighbor graph.
}
\description{
Weights a probability matrix so that it approximates a (possibly symmetrized)
nearest neighbor graph. In the context of probability-based embedding, the
nearest neighbor adjancency graph is the same as the weight matrix that would
result if a step function was used, rather than the usual exponential
weighting. Given a row probability matrix we can then map back to the nearest
neighbor graph by multiplying each probability by the number of neighbours
required, which is equivalent to the perplexity.
}
\details{
If the matrix is already normalized and symmetrized then the resulting
neighbor graph is also symmetrized.
}

