% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{no_momentum}
\alias{no_momentum}
\title{An update schedule with no momentum.}
\usage{
no_momentum()
}
\value{
A solution update method for use by the optimizer. A list consisting
of:
 \item{\code{init}}{Function to do any needed initialization.}
 \item{\code{get_update}}{Function to return the current update, which will
 be added to current solution matrix.}
}
\description{
Create an callback for the optimizer to use that uses strict gradient
descent with no momentum term.
}
\seealso{
Other sneer optimization update methods: \code{\link{linear_momentum}},
  \code{\link{nesterov_nsc_momentum}},
  \code{\link{step_momentum}}
}

