% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{normalized_stress}
\alias{normalized_stress}
\title{Normalized stress.}
\usage{
normalized_stress(dxm, dym, eps = .Machine$double.eps)
}
\arguments{
\item{dxm}{Distance matrix.}

\item{dym}{Distance matrix, must be of the same dimensions as \code{dxm}.}

\item{eps}{Small floating point value to avoid numerical problems.}
}
\value{
Normalized stress.
}
\description{
A measure of embedding quality between input and output distance matrices.
}
\details{
A normalized version of the \code{\link{metric_stress}}:

\deqn{STRESS_{RN} = \frac{STRESS}{\sum_{i<j} dxm_{ij}^2}}{STRESS/sum(dxm^2)}

where \eqn{dxm} is the input distance matrix and \eqn{STRESS} is the
\code{\link{metric_stress}} between the input and output distances.
This gives a dimensionless value similar to the \code{\link{kruskal_stress}},
except the normalization uses the input distances, not the output distances,
much like the \code{\link{sammon_stress}}. It can be interpreted as the
proportion of the sum of squares of the input distances unaccounted for by
the output distances. Borg and Groenen defined this in their book on MDS.
}
\references{
Ingwer Borg, Patrick J. F. Groenen
Modern Multidimensional Scaling: Theory and Applications
Springer Science & Business Media, Aug 4, 2005
}

