% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{null_model_dist}
\alias{null_model_dist}
\title{Null Model for Distance Matrices}
\usage{
null_model_dist(dm)
}
\arguments{
\item{dm}{Distance matrix, the dimensions of which will be used to create
the null matrix.}
}
\value{
Matrix of zeros of the same dimensions as \code{dm}.
}
\description{
Calculates a distance matrix that represents a "null" model, i.e. one
where no information from the input distances has been used. In this case,
the null model would mean all distances equal, which can only be achieved
if there is more than one distance by setting all distances to zero.
}

