% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{null_model_prob}
\alias{null_model_prob}
\title{Null Model for Probability Matrices}
\usage{
null_model_prob(pm)
}
\arguments{
\item{pm}{Probability matrix. Can be row, joint or conditional.}
}
\value{
Probability matrix where all elements are equal.
}
\description{
For a given probability matrix, return the equivalent "null" model, i.e. one
where all probabilities are equal.
}

