% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{olivetti_faces}
\alias{olivetti_faces}
\title{Olivetti Faces dataset}
\format{A data frame with 400 rows and 4097 variables.}
\usage{
olivetti_faces()
}
\value{
the Olivetti Faces dataset as a dataframe.
}
\description{
Sneer benchmarking data.
}
\details{
Returns the Olivetti Faces dataset in a data frame reformatted to be suitable
for use with Sneer. This is a series of 400 images (with dimension 64 x 64)
of 40 individual's faces, with ten different poses per person.

The variables are as follows:
\itemize{
\item \code{px1}, \code{px2}, \code{px3} ... \code{px560} 8-bit grayscale
pixel values (0-255). The pixel index starts at the top right of the image
(\code{px1}) and are then stored column-wise.
\item \code{Label} An integer in the range (1-40) indicating the person.
}

Each row has a name with the format "<face>_<pose>", where \code{<face>} is
the index of the face, and \code{<pose>} is the index of the pose, e.g.
the row with name \code{20_10} is the tenth pose of the twentieth face.
}
\note{
requires the \code{RnavGraphImageData} package.
\url{https://cran.r-project.org/web/packages/RnavGraphImageData/index.html}
to be installed and loaded.
}
\seealso{
Saul Roweis' dataset web page: \url{http://www.cs.nyu.edu/~roweis/data.html}.
Each row can be visualized as an image using
\code{\link{show_olivetti_face}}.
}

