% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{opt_report}
\alias{opt_report}
\title{Optimizer Diagnostics}
\usage{
opt_report(opt)
}
\arguments{
\item{opt}{Optimizer to report on.}
}
\value{
a list containing (very) brief numeric summary of parts of the
optimizer:
\item{\code{grad_length}}{Length of the gradient.}
\item{\code{step_size}}{Value of the step size (or length of the step size
 vector).}
\item{\code{momentum}}{Momentum (if any).}
}
\description{
Simple diagnostics of the state of the optimization.
}

