% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_direction.R
\name{optimization_direction}
\alias{optimization_direction}
\title{Optimization Direction Methods}
\description{
The available direction methods that can be used by the optimization routines
in sneer.
}
\examples{
make_opt(direction = steepest_descent())

}
\seealso{
Other sneer optimization direction methods: \code{\link{steepest_descent}}
}
\keyword{internal}

