% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{optimization_gradient}
\alias{optimization_gradient}
\title{Optimization Gradient Methods}
\description{
The available gradient methods that can be used by the optimization routines
in sneer.
}
\examples{
make_opt(gradient = classical_gradient())
make_opt(gradient = nesterov_gradient())
}
\seealso{
Other sneer optimization gradient methods: \code{\link{classical_gradient}},
  \code{\link{nesterov_gradient}}
}
\keyword{internal}

