% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optimization_methods}
\alias{optimization_methods}
\title{Optimization Methods}
\description{
The available methods and wrappers that can be used to create optimization
routines in sneer.
}
\examples{
make_opt(step_size = bold_driver())
make_opt(step_size = jacobs())
}
\seealso{
Other sneer optimization methods: \code{\link{bold_nagger}},
  \code{\link{gradient_descent}}, \code{\link{make_opt}},
  \code{\link{tsne_opt}}
}
\keyword{internal}

