% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{optimization_update}
\alias{optimization_update}
\title{Optimization Update Methods}
\description{
The available update methods that can be used by the optimization routines
in sneer.
}
\examples{
make_opt(update = step_momentum(init_momentum = 0.4, final_momentum = 0.8))
make_opt(update = no_momentum())

}
\seealso{
Other sneer optimization update methods: \code{\link{constant_momentum}},
  \code{\link{linear_momentum}},
  \code{\link{nesterov_nsc_momentum}},
  \code{\link{no_momentum}}, \code{\link{step_momentum}}
}
\keyword{internal}

