% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optimize_step}
\alias{optimize_step}
\title{One Step of Optimization}
\usage{
optimize_step(opt, method, inp, out, iter)
}
\arguments{
\item{opt}{Optimizer}

\item{method}{Embedding method.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{iter}{Iteration number.}
}
\value{
List consisting of:
  \item{\code{opt}}{Updated optimizer.}
  \item{\code{inp}}{Updated input.}
  \item{\code{out}}{Updated output.}
}
\description{
One Step of Optimization
}

