% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_direction.R
\name{optimizer_direction}
\alias{optimizer_direction}
\title{Optimizer direction methods.}
\description{
Part of the optimizer that finds the direction of descent.
}
\section{Interface}{

A direction method is a list containing:
\describe{
\item{\code{init(opt, inp, out, method)}}{Optional initialization function
with the following arguments:
 \describe{
   \item{\code{opt}}{Optimizer.}
   \item{\code{inp}}{Input data.}
   \item{\code{out}}{Output data.}
   \item{\code{method}}{Embedding method.}
 }
 The function should set any needed state on \code{opt$direction} and return
 a list containing:
 \describe{
   \item{\code{opt}}{Optimizer containing initialized \code{direction}.}
 }
}
\item{\code{calculate(opt, inp, out, method, iter)}}{Calculation function
with the following arguments:
 \describe{
   \item{\code{opt}}{Optimizer.}
   \item{\code{inp}}{Input data.}
   \item{\code{out}}{Output data.}
   \item{\code{method}}{Embedding method.}
   \item{\code{iter}}{Iteration number.}
 }
 The function should \code{opt$direction$value} with the current direction
 of descent and return a list containing:
 \describe{
   \item{\code{opt}}{Optimizer containing updated \code{direction$value}.}
 }
}
\item{\code{value}}{The current direction. It should be a matrix with
the same dimensions as the gradient.}
}
}
\keyword{internal}

