% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_step.R
\name{optimizer_step_size}
\alias{optimizer_step_size}
\title{Optimizer step size methods.}
\description{
Part of the optimizer that finds the step size of the descent.
}
\section{Interface}{

A step size method is a list containing:
\describe{
 \item{\code{value}}{The current step size. It should either be a scalar or
 a matrix with the same dimensions as the gradient.}
 \item{\code{calculate(opt, inp, out, method)}}{Calculation function
 with the following arguments:
   \describe{
     \item{\code{opt}}{Optimizer.}
     \item{\code{inp}}{Input data.}
     \item{\code{out}}{Output data.}
     \item{\code{method}}{Embedding method.}
   }
   The function should set \code{opt$step_size$value} with the current
   step size and return a list containing:
   \describe{
     \item{\code{opt}}{Optimizer containing updated \code{step_size$value}.}
   }
 }
 \item{\code{init(opt, inp, out, method)}}{Optional initialization function
 with the following arguments:
   \describe{
     \item{\code{opt}}{Optimizer.}
     \item{\code{inp}}{Input data.}
     \item{\code{out}}{Output data.}
     \item{\code{method}}{Embedding method.}
   }
   The function should set any needed state on \code{opt$step_size} and
   return a list containing:
   \describe{
     \item{\code{opt}}{Optimizer containing initialized \code{step_size}
     method.}
   }
 }
 \item{\code{validate(opt, inp, out, proposed_out, method)}}{Optional
 validation function with the following arguments:
   \describe{
     \item{\code{opt}}{Optimizer.}
     \item{\code{inp}}{Input data.}
     \item{\code{out}}{Output data from the start of the iteration.}
     \item{\code{proposed_out}}{Proposed updated output for this iteration.}
     \item{\code{method}}{Embedding method.}
   }
   The function should do any validation required by this method on the state
   of \code{proposed_out}, e.g. check that the proposed solution reduces the
   cost function. In addition it should update the state of any of the other
   arguments passed to the validation function on the basis of the pass or
   failure of the validation.
   The return value of the function should be a list containing:
   \describe{
     \item{\code{opt}}{Optimizer.}
     \item{\code{inp}}{Input data.}
     \item{\code{out}}{Output data from the start of the iteration.}
     \item{\code{proposed_out}}{Proposed updated output for this iteration.}
     \item{\code{method}}{Embedding method.}
     \item{\code{ok}}{Logical value, \code{TRUE} if \code{proposed_out}
     passed validation, \code{FALSE} otherwise}
   }
   Note that if any validation functions fail the proposed solution by
   setting \code{ok} to \code{FALSE} in their return value, the optimizer
   will reject \code{proposed_out} and use \code{out} as the starting point
   for the next iteration of the optimization process.
 }
 \item{\code{after_step(opt, inp, out, new_out, ok, iter)}}{Optional function
 to invoke after the solution has been updated with the following arguments:
   \describe{
     \item{\code{opt}}{Optimizer.}
     \item{\code{inp}}{Input data.}
     \item{\code{out}}{Output data from the start of the iteration.}
     \item{\code{new_out}}{Output data which will be the starting solution
     for the next iteration of optimization. If the validation stage failed,
     then this may be the same solution as \code{out}.}
     \item{\code{ok}}{\code{TRUE} if the current iteration passed validation,
     \code{FALSE} otherwise.}
     \item{\code{iter}}{Current iteration number.}
   }
   The function should do any processing of this method's internal state to
   prepare for the next iteration and call to \code{calculate}. The
   return value of the function should be a list containing:
   \describe{
     \item{\code{opt}}{Updated optimizer.}
     \item{\code{inp}}{Input data.}
     \item{\code{out}}{Output data from the start of the iteration.}
     \item{\code{new_out}}{New output to be used in the next iteration.}
   }
 }
}
}
\keyword{internal}

