% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{out_from_matrix}
\alias{out_from_matrix}
\title{Initialize Output Coordinates from Matrix}
\usage{
out_from_matrix(init_config, k = 2, verbose = TRUE)
}
\arguments{
\item{init_config}{Configuration to initialize the coordinates from. Must
be a matrix with the same dimensions as the desired output coordinates.}

\item{k}{Number of output dimensions. For 2D visualization this is always 2.}

\item{verbose}{If \code{TRUE}, log information about the initialization.}
}
\value{
Output initializer.
}
\description{
Output initialization function.
}
\details{
Creates output data and initialize coordinates from the specified matrix.
}
\examples{
\dontrun{
# create a scores matrix using R PCA
pca_scores <- prcomp(iris[, 1:4], center = TRUE, retx = TRUE)$x[, 1:2]

# Should be passed to the init_out argument of an embedding function:
embed_dist(init_out = out_from_matrix(pca_scores), ...)
}
}
\seealso{
\code{\link{embed_dist}} and \code{\link{embed_prob}}
  for how to use this function to configure an embedding.

Other sneer output intializers: \code{\link{out_from_PCA}},
  \code{\link{out_from_rnorm}},
  \code{\link{out_from_runif}}
}

