% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{out_from_PCA}
\alias{out_from_PCA}
\title{Initialize Output Coordinates from PCA}
\usage{
out_from_PCA(k = 2, verbose = TRUE)
}
\arguments{
\item{k}{Number of output dimensions. For 2D visualization this is always 2.}

\item{verbose}{If \code{TRUE}, log information about the initialization.}
}
\value{
Output initializer.
}
\description{
Output initialization function.
}
\details{
The first \code{k} scores of the PCA of the input coordinates are used to
initialize the \code{k} dimensions of the embedded coordinates. Input
coordinates are centered but not scaled before the PCA is carried out.
This initialization can only be used if the input  data is in the form of
coordinates, not a distance matrix.
}
\examples{
\dontrun{
# Should be passed to the init_out argument of an embedding function:
embed_dist(init_out = out_from_PCA(), ...)
}
}
\seealso{
\code{embed_dist} and \code{embed_prob}
  for how to use this function to configure an embedding.

Other sneer output intializers: \code{\link{out_from_matrix}},
  \code{\link{out_from_rnorm}},
  \code{\link{out_from_runif}}
}

