% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{out_from_rnorm}
\alias{out_from_rnorm}
\title{Initialize Output Coordinates from Normal Distribution}
\usage{
out_from_rnorm(k = 2, sd = 1e-04, verbose = TRUE)
}
\arguments{
\item{k}{Number of output dimensions. For 2D visualization this is always 2.}

\item{sd}{The standard deviation of the distribution.}

\item{verbose}{If \code{TRUE}, log information about the initialization.}
}
\value{
Output initializer.
}
\description{
Output initialization function.
}
\details{
Creates output data and initializes embedding coordinates from a normal
distribution centered at zero.
}
\examples{
\dontrun{
# Should be passed to the init_out argument of an embedding function:
 embed_dist(init_out = out_from_rnorm(sd = 1e-4), ...)
}
}
\seealso{
\code{\link{embed_dist}} and \code{\link{embed_prob}}
  for how to use this function to configure an embedding.

Other sneer output intializers: \code{\link{out_from_PCA}},
  \code{\link{out_from_matrix}},
  \code{\link{out_from_runif}}
}

