% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{out_from_runif}
\alias{out_from_runif}
\title{Initialize Output Coordinates from Uniform Distribution}
\usage{
out_from_runif(k = 2, min = 0, max = 1, verbose = TRUE)
}
\arguments{
\item{k}{Number of output dimensions. For 2D visualization this is always 2.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}

\item{verbose}{If \code{TRUE}, log information about the initialization.}
}
\value{
Output initializer.
}
\description{
Output initialization function.
}
\details{
Creates output data and initializes embedding coordinates from a random
uniform distribution. This is the initialization method suggested for use
by the authors of the \code{nerv} method.
}
\examples{
\dontrun{
# Should be passed to the init_out argument of an embedding function:
 embed_dist(init_out = out_from_runif(min = -10, max = 10), ...)
}
}
\references{
Venna, J., Peltonen, J., Nybo, K., Aidos, H., & Kaski, S. (2010).
Information retrieval perspective to nonlinear dimensionality reduction for
data visualization.
\emph{Journal of Machine Learning Research}, \emph{11}, 451-490.
}
\seealso{
\code{embed_dist} and \code{embed_prob}
  for how to use this function to configure an embedding.

Other sneer output intializers: \code{\link{out_from_PCA}},
  \code{\link{out_from_matrix}},
  \code{\link{out_from_rnorm}}
}

