% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{outdegree_centrality}
\alias{outdegree_centrality}
\title{Outdegree Centrality}
\usage{
outdegree_centrality(m)
}
\arguments{
\item{m}{Nearest neighbor graph matrix.}
}
\value{
Outdegree centrality of \code{m}.
}
\description{
A node centrality measure for graphs.
}
\details{
Sums the (possibly weighted) edges that are directed from each node. For
probability matrices, row i is interpreted as containing the weighted
edges that are directed from node i.
}

