% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{output_initializers}
\alias{output_initializers}
\title{Output Initializers}
\description{
These methods deal with initializing the output coordinates of an embedding.
They have names that begin with 'out_from' to indicate that they are only
used to initialize the output data, not the input data.
}
\examples{

\dontrun{
# pass to the init_out parameter of an embedding function

# initialize from PCA
embed_dist(init_out = out_from_PCA(), ...)

# initialize from small random differences
embed_prob(init_out = out_from_rnorm(sd = 1e-4), ...)
}
}
\seealso{
The return value of the initializers should be assigned to the
\code{init_out} parameter of an embedding function such as
\code{\link{embed_dist}} and \code{\link{embed_prob}}.
}
\keyword{internal}

