% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_external.R
\name{par_to_out}
\alias{par_to_out}
\title{Convert 1D Parameter to Sneer Output}
\usage{
par_to_out(par, opt, inp, method, nrow)
}
\arguments{
\item{par}{Vector of embedded coordinates.}

\item{opt}{Optimizer}

\item{inp}{Input data.}

\item{method}{Embedding method.}

\item{nrow}{Number of rows in the sneer output matrix.}
}
\value{
Output data with coordinates converted from \code{par}.
}
\description{
This function converts the 1D parameter format expected by
\code{\link[stats]{optim}} into the matrix format used by sneer.
}

