% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{partial}
\alias{partial}
\title{Partially apply a function.}
\usage{
partial(f, ...)
}
\arguments{
\item{f}{Function to partially apply.}

\item{...}{params of \code{f} to apply.}
}
\value{
Partially applied version of \code{f}.
}
\description{
Partially apply a function.
}

