% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{perplexity_rows}
\alias{perplexity_rows}
\title{Perplexity for Row Probability Matrices}
\usage{
perplexity_rows(pm, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Row probability matrix.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Vector of perplexities, one per row of the matrix.
}
\description{
Calculates the perplexity (antilogarithm of the Shannon Entropy) per row of
a row probability matrix. Each row of the matrix should consist of
probabilities summing to 1.
}

