% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugin_stiffness}
\alias{plugin_stiffness}
\title{Plugin Stiffness}
\usage{
plugin_stiffness(method, inp, out)
}
\arguments{
\item{method}{Embedding method.}

\item{inp}{Input data.}

\item{out}{Output data.}
}
\value{
Stiffness matrix.
}
\description{
Calculates the Stiffness matrix of an embedding method using the plugin
gradient formulation.
}

