% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{plugin_stiffness_ms}
\alias{plugin_stiffness_ms}
\title{Multiscale Plugin Stiffness}
\usage{
plugin_stiffness_ms(method, inp, out)
}
\arguments{
\item{method}{Embedding method.}

\item{inp}{Input data.}

\item{out}{Output data.}
}
\value{
Stiffness matrix.
}
\description{
Calculates the stiffness matrix of an embedding method using the multiscale
plugin gradient formulation.
}

