% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{plugin_stiffness_ms_row}
\alias{plugin_stiffness_ms_row}
\title{Multiscale Plugin Stiffness for Row Probability}
\usage{
plugin_stiffness_ms_row(method, inp, out)
}
\arguments{
\item{method}{Embedding method.}

\item{inp}{Input data.}

\item{out}{Output data.}
}
\value{
Stiffness matrix.
}
\description{
Calculates the multiscale stiffness matrix for row probability based
embedding methods.
}

