% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugin_stiffness_row}
\alias{plugin_stiffness_row}
\title{Plugin Stiffness for Row Probabilities}
\usage{
plugin_stiffness_row(method, inp, out)
}
\arguments{
\item{method}{Embedding method.}

\item{inp}{Input data.}

\item{out}{Output data.}
}
\value{
Stiffness matrix.
}
\description{
Calculates the stiffness matrix for row probability based embedding methods.
}

