% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugin_stiffness_symm_joint}
\alias{plugin_stiffness_symm_joint}
\title{Plugin Stiffness for Joint Probabilities}
\usage{
plugin_stiffness_symm_joint(method, inp, out)
}
\arguments{
\item{method}{Embedding method.}

\item{inp}{Input data.}

\item{out}{Output data.}
}
\value{
Stiffness matrix.
}
\description{
Calculates the stiffness matrix for joint probability based embedding
methods.
}

