% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{prec_to_bandwidth}
\alias{prec_to_bandwidth}
\title{Convert Precisions to Sigma (Bandwidths)}
\usage{
prec_to_bandwidth(prec)
}
\arguments{
\item{prec}{Precision (beta).}
}
\value{
Bandwidth (sigma).
}
\description{
Gaussian-type functions have a parameter associated with them
that is either thought of as a "precision", i.e. how "tight" the distribution
is, or its counterpart, the bandwidth, which measure how spread-out the
distribution is. For gaussian functions, a direct comparison can be made
with the standard deviation of the function. Exact definitions of the
parameters differ in different discussions. In sneer, the gaussian function
is defined as:
}
\details{
\deqn{e(-\beta D^2)}{exp(-beta * D^2)}

where \eqn{\beta}{beta} is the precision, or alternatively:

\deqn{e^{-\frac{D^2}{2\sigma^2}}}{exp[-(D ^ 2)/(2 * sigma ^ 2)]}

so that \eqn{\sigma = \frac{1}{\sqrt{2\beta}}}{sigma = 1/sqrt(2*beta)}

This function performs the conversion of the parameter between precision and
bandwidth.
}

