% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{prow_to_pcond}
\alias{prow_to_pcond}
\title{Conditional Probability Matrix from Row Probability Matrix}
\usage{
prow_to_pcond(prow)
}
\arguments{
\item{prow}{Row probability matrix.}
}
\value{
Conditional probability matrix.
}
\description{
Given a row probability matrix (elements of each row are non-negative and
sum to one), this function scales each element by the sum of the matrix so
that the elements of the entire matrix sum to one.
}

