% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{qnx_crm}
\alias{qnx_crm}
\title{Average Normalized Agreement Between K-ary Neighborhoods (QNX)}
\usage{
qnx_crm(crm, k)
}
\arguments{
\item{crm}{Co-ranking matrix. Create from a pair of distance matrices with
\code{coranking_matrix}.}

\item{k}{Neighborhood size.}
}
\value{
QNX for \code{k}.
}
\description{
QNX measures the degree to which an embedding preserves the local
neighborhood around each observation. For a value of K, the K closest
neighbors of each observation are retrieved in the input and output space.
For each observation, the number of shared neighbors can vary between 0
and K. QNX is simply the average value of the number of shared neighbors,
normalized by K, so that if the neighborhoods are perfectly preserved, QNX
is 1, and if there is no neighborhood preservation, QNX is 0.
}
\details{
For a random embedding, the expected value of QNX is approximately
K / (N - 1) where N is the number of observations. Using RNX
(\code{rnx_crm}) removes this dependency on K and the number of
observations.
}
\references{
Lee, J. A., & Verleysen, M. (2009).
Quality assessment of dimensionality reduction: Rank-based criteria.
\emph{Neurocomputing}, \emph{72(7)}, 1431-1443.
}

