% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{random_matrix_norm}
\alias{random_matrix_norm}
\title{Random Matrix (Normal Distribution)}
\usage{
random_matrix_norm(nrow, ncol = 2, sd = 1e-04)
}
\arguments{
\item{nrow}{Number of rows of the matrix.}

\item{ncol}{Number of columns of the matrix.}

\item{sd}{Standard deviation of the distribution.}
}
\value{
Random matrix with \code{nrow} rows and \code{ncol} columns.
}
\description{
Creates a matrix of normally distributed data with zero mean.
}
\examples{
\dontrun{
# matrix with 5 rows, 3 columns and standard deviation of 0.1
xm <- random_matrix_norm(5, 3, 0.1)
# matrix with 100 rows
xm <- random_matrix_norm(100)
}
}

