% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{random_matrix_unif}
\alias{random_matrix_unif}
\title{Random Matrix (Uniform Distribution)}
\usage{
random_matrix_unif(nrow, ncol = 2, min = 0, max = 1)
}
\arguments{
\item{nrow}{Number of rows of the matrix.}

\item{ncol}{Number of columns of the matrix.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}
}
\value{
Random matrix with \code{nrow} rows and \code{ncol} columns.
}
\description{
Creates a matrix of uniformly distributed data.
}
\examples{
\dontrun{
# matrix with 5 rows, 3 columns distribution between -10 and 10
xm <- random_matrix_unif(nrow = 5, ncol = 3, min = -10, max = 10)
# 2D matrix with 100 rows and default range of 0 to 1
xm <- random_matrix_unif(100)
}
}

