% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{range_scale}
\alias{range_scale}
\title{Range scales columns of numeric matrix.}
\usage{
range_scale(xm, rmin = 0, rmax = 1)
}
\arguments{
\item{xm}{Matrix to range scale.}

\item{rmin}{Minimum value per column in the scaled matrix.}

\item{rmax}{Maximum value per column in the scaled matrix.}
}
\value{
Range scaled matrix.
}
\description{
Values are scaled so that for each column, values are within
(\code{min}, \code{max}).
}

