% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{range_scale_matrix}
\alias{range_scale_matrix}
\title{Range scales numeric matrix.}
\usage{
range_scale_matrix(xm, rmin = 0, rmax = 1)
}
\arguments{
\item{xm}{Matrix to range scale.}

\item{rmin}{Minimum value in the scaled matrix.}

\item{rmax}{Maximum value in the scaled matrix.}
}
\value{
Range scaled data.
}
\description{
Elements are scaled so that they are within (\code{rmin}, \code{rmax}).
}

