% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{rasne_plugin}
\alias{rasne_plugin}
\title{RASNE Method using Plugin Gradient}
\usage{
rasne_plugin(beta = 1, eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{beta}{Bandwidth of the exponential similarity kernel
function.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
An implementation of RASNE using the plugin gradient.
}
\seealso{
\code{rasne} should give equivalent results, but is probably
a bit more efficient.

Other sneer embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsnerv_plugin}}, \code{\link{hssne_plugin}},
  \code{\link{hssne}}, \code{\link{jse_plugin}},
  \code{\link{nerv_plugin}}, \code{\link{plugin}},
  \code{\link{rasne}}, \code{\link{rssne_plugin}},
  \code{\link{rssne}}, \code{\link{rtsne_plugin}},
  \code{\link{rtsne}}, \code{\link{sjse_plugin}},
  \code{\link{snerv_plugin}}, \code{\link{ssne_plugin}},
  \code{\link{ssne}}, \code{\link{tasne_plugin}},
  \code{\link{tasne}}, \code{\link{tnerv_plugin}},
  \code{\link{tsne_plugin}}, \code{\link{tsne}},
  \code{\link{uhsnerv_plugin}}, \code{\link{unerv_plugin}},
  \code{\link{usnerv_plugin}}

Other sneer probability embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsnerv_plugin}}, \code{\link{hssne_plugin}},
  \code{\link{hssne}}, \code{\link{jse_plugin}},
  \code{\link{nerv_plugin}}, \code{\link{plugin}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne}}, \code{\link{rssne_plugin}},
  \code{\link{rssne}}, \code{\link{rtsne_plugin}},
  \code{\link{rtsne}}, \code{\link{sjse_plugin}},
  \code{\link{snerv_plugin}}, \code{\link{ssne_plugin}},
  \code{\link{ssne}}, \code{\link{tasne_plugin}},
  \code{\link{tasne}}, \code{\link{tnerv_plugin}},
  \code{\link{tsne_plugin}}, \code{\link{tsne}},
  \code{\link{uhsnerv_plugin}}, \code{\link{unerv_plugin}},
  \code{\link{usnerv_plugin}}
}

