% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{reltol}
\alias{reltol}
\title{Relative tolerance.}
\usage{
reltol(x, y)
}
\arguments{
\item{x}{real value.}

\item{y}{real value.}
}
\value{
the relative tolerance between the two values.
}
\description{
Relative tolerance.
}

