% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_nulls}
\alias{remove_nulls}
\title{Remove NULL Members from a List}
\usage{
remove_nulls(l)
}
\arguments{
\item{l}{List.}
}
\value{
List with NULL members removed.
}
\description{
Remove NULL Members from a List
}
\examples{
\dontrun{
mylist <- list(foo = "bar", cleesh = NULL, baz = "qux", nitfol = NULL)
names(remove_nulls(mylist)) == c("foo", "baz")
}
}

