% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{rev_asne_stiffness}
\alias{rev_asne_stiffness}
\title{"Reverse" ASNE Stiffness Function}
\usage{
rev_asne_stiffness(pm, qm, rev_kl, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Input probability matrix.}

\item{qm}{Output probabilty matrix.}

\item{rev_kl}{"Reverse" KL divergence between \code{pm} and \code{qm}.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix.
}
\description{
Uses the exponential weighting function for similarities, but the
"reverse" Kullback Leibler divergence as the cost function.
}

