% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{reverse_asne_stiffness}
\alias{reverse_asne_stiffness}
\title{"Reverse" ASNE Stiffness Function}
\usage{
reverse_asne_stiffness(pm, qm, rev_kl, beta = 1, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Input probability matrix.}

\item{qm}{Output probabilty matrix.}

\item{rev_kl}{"Reverse" KL divergence between \code{pm} and \code{qm}.}

\item{beta}{The precision of the weighting function.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix.
}
\description{
Uses the exponential weighting function for similarities, but the
"reverse" Kullback-Leibler divergence as the cost function.
}
\details{
The precision parameter \code{beta} is normally a scalar, but it can also
work with a vector, as long as the length of the vector is equal to the
number of rows  in the probability matrices. The \code{nerv} method
makes use of this property.
}

