% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{reverse_hssne_stiffness}
\alias{reverse_hssne_stiffness}
\title{"Reverse" HSSNE Stiffness Function}
\usage{
reverse_hssne_stiffness(pm, qm, wm, rev_kl, alpha = 1.5e-08, beta = 1,
  eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}

\item{wm}{Output weight probability matrix.}

\item{rev_kl}{"Reverse" KL divergence between \code{pm} and \code{qm}.}

\item{alpha}{Tail heaviness of the weighting function.}

\item{beta}{The precision of the weighting function.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix.
}
\description{
The precision parameter \code{beta} is normally left at its default value of
1. Note that unlike the \code{reverse_asne_stiffness} function, a
vector of precisions can not be used as input to \code{beta}: an incorrect
gradient will result.
}

