% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{reverse_kl_cost_gr}
\alias{reverse_kl_cost_gr}
\title{Reverse Kullback Leibler Divergence Cost Gradient}
\usage{
reverse_kl_cost_gr(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Gradient of the reverse KL divergence cost of the embedding with
respect to the output probabilities.
}
\description{
Calculates the gradient of the "reverse" KL divergence cost function of
an embedding with respect to the output probabilities.
}

