% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{reverse_kl_divergence_gr}
\alias{reverse_kl_divergence_gr}
\title{Reverse Kullback Leibler Gradient}
\usage{
reverse_kl_divergence_gr(pm, qm, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Probability Matrix. First probability in the divergence.}

\item{qm}{Probability Matrix. Second probability in the divergence.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Gradient of the KL divergence from \code{qm} to \code{pm}.
}
\description{
Calculates the gradient of the KL divergence with respect to the
probability Q in KL(Q||P).
}

