% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{reverse_kl_fg}
\alias{reverse_kl_fg}
\title{Reverse Kullback Leibler Divergence Cost}
\usage{
reverse_kl_fg()
}
\value{
KL divergence function and gradient.
}
\description{
Cost wrapper factory function.
}
\details{
Creates the a list containing the required functions for using "reverse"
Kullback Leibler Divergence, i.e. KL(Q||P), in an embedding.

Provides the cost function and its gradient (with respect to Q).
}
\seealso{
Other sneer cost wrappers: \code{\link{nerv_fg}}
}

