% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{reverse_ssne_stiffness}
\alias{reverse_ssne_stiffness}
\title{"Reverse" SSNE Stiffness Function}
\usage{
reverse_ssne_stiffness(pm, qm, rev_kl, beta = 1, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}

\item{rev_kl}{"Reverse" KL divergence between \code{pm} and \code{qm}.}

\item{beta}{The precision of the weighting function.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix.
}
\description{
Uses the exponential weighting function for similarities, but the
"reverse" Kullback-Leibler divergence as the cost function.
}
\details{
The precision parameter \code{beta} is normally left at its default value of
1. Note that unlike the \code{\link{reverse_asne_stiffness}} function, a
vector of precisions can not be used as input to \code{beta}: an incorrect
gradient will result.
}

