% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{rms_metric_stress}
\alias{rms_metric_stress}
\title{Root mean square deviation of the metric stress.}
\usage{
rms_metric_stress(dxm, dym)
}
\arguments{
\item{dxm}{Distance matrix.}

\item{dym}{Distance matrix, must be of the same dimensions as \code{dxm}.}
}
\value{
RMS deviation of the input and output distances.
}
\description{
A measure of embedding quality between distance matrices.
}
\details{
The RMS stress is the square root of the normalized
\code{\link{metric_stress}}:

\deqn{STRESS_{RMS} = \sqrt{\frac{2\times STRESS}{n(n-1)}}}{stress_rms = sqrt[(2*STRESS)/(n*(n-1))]}

The normalization divides the metric stress by
\eqn{0.5\times n(n-1)}{0.5*n*(n-1)}, where \eqn{n} is the number of points
in the distance matrix and hence the normalization factor accounts for the
number of unique distances (ignoring self distances) in the matrix. The
square root puts the units of the stress back into that of the original
distances.
}

