% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{rnx_auc}
\alias{rnx_auc}
\title{Area Under the RNX Curve}
\usage{
rnx_auc(inp, out)
}
\arguments{
\item{inp}{Input data. The input distance matrix will be calculated if it's
not present.}

\item{out}{Output data. The output distance matrix will be calculated if
it's not present.}
}
\value{
Area under the RNX curve.
}
\description{
The RNX curve is formed by calculating the \code{rnx_crm} metric for
different sizes of neighborhood. Each value of RNX is scaled according to
the natural log of the neighborhood size, to give a higher weight to smaller
neighborhoods. An AUC of 1 indicates perfect neighborhood preservation, an
AUC of 0 is due to random results.
}
\note{
Calculating the RNX curve requires calculating a co-ranking matrix,
 which is not a very efficient operation, because it requires iterating
 over every element of the distance matrix, the size of which scales with
 the square of the number of observations. Be sure you really want to
 calculate this for datasets with more than approximately 1,000 observations.
}
\references{
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}

