% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{rnx_auc_embed}
\alias{rnx_auc_embed}
\title{Area Under the RNX Curve}
\usage{
rnx_auc_embed(din, dout)
}
\arguments{
\item{din}{Input distance matrix.}

\item{dout}{Output distance matrix.}
}
\value{
Area under the RNX curve.
}
\description{
The RNX curve is formed by calculating the \code{rnx_crm} metric for
different sizes of neighborhood. Each value of RNX is scaled according to
the natural log of the neighborhood size, to give a higher weight to smaller
neighborhoods. An AUC of 1 indicates perfect neighborhood preservation, an
AUC of 0 is due to random results.
}
\references{
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}

