% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{rnx_crm}
\alias{rnx_crm}
\title{Rescaled Agreement Between K-ary Neighborhoods (RNX)}
\usage{
rnx_crm(crm, k)
}
\arguments{
\item{crm}{Co-ranking matrix. Create from a pair of distance matrices with
\code{coranking_matrix}.}

\item{k}{Neighborhood size.}
}
\value{
RNX for \code{k}.
}
\description{
RNX is a scaled version of QNX which measures the agreement between two
embeddings in terms of the shared number of k-nearest neighbors for each
observation. RNX gives a value of 1 if the neighbors are all preserved
perfectly and a value of 0 for a random embedding.
}
\references{
Lee, J. A., Renard, E., Bernard, G., Dupont, P., & Verleysen, M. (2013).
Type 1 and 2 mixtures of Kullback-Leibler divergences as cost functions in
dimensionality reduction based on similarity preservation.
\emph{Neurocomputing}, \emph{112}, 92-108.
}

