% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{roc_auc}
\alias{roc_auc}
\title{Average Area Under the ROC Curve}
\usage{
roc_auc(inp, out)
}
\arguments{
\item{inp}{Input data. This should be storing the class labels as
\code{inp$labels}, as vector with the labels ordered in the same way as
the observations in the distance matrices.}

\item{out}{Output data. If the output distance matrix is not stored as
\code{out$dm}, it will be calculated.}
}
\value{
Area Under the ROC curve, averaged over each observation.
}
\description{
Embedding quality measure.
}
\details{
The ROC curve plots the true positive rate vs false positive rate.
This function calculates the curve N times, where N is the number of the
observations. The label of the Nth observation is set as the positive class
and then the other observations are ranked according to their distance from
the Nth observation in the output coordinates (lower distances being better).
Observations with the same label as the Nth observation count as positive
observations. The final reported result is the average over all observations.

Perfect retrieval results in an AUC of 1. For random retrieval gives a value
of 0.5.
}
\note{
Use of this function requires that the \code{PRROC} package be
installed.
}

