% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{roc_auc_embed}
\alias{roc_auc_embed}
\title{Average Area Under the ROC Curve}
\usage{
roc_auc_embed(dm, labels)
}
\arguments{
\item{dm}{Distance matrix of an embedding.}

\item{labels}{Vector of labels for each observation in the dataset in the
same order as the observations in the distance matrix.}
}
\value{
Area Under the ROC curve, averaged over each observation.
}
\description{
Embedding quality measure.
}
\details{
The ROC curve plots the true positive rate vs false positive rate.
This function calculates the curve N times, where N is the number of the
observations. The label of the Nth observation is set as the positive class
and then the other observations are ranked according to their distance from
the Nth observation in the output coordinates (lower distances being better).
Observations with the same label as the Nth observation count as positive
observations. The final reported result is the average over all observations.

Perfect retrieval results in an AUC of 1. For random retrieval gives a value
of 0.5.
}
\note{
Use of this function requires that the \code{PRROC} package be
installed.
}
