% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{roc_auc_row}
\alias{roc_auc_row}
\title{Area Under the ROC Curve of an Observation}
\usage{
roc_auc_row(dm, labels, i)
}
\arguments{
\item{dm}{Distance matrix.}

\item{labels}{Vector of labels, of the same size as the number of rows
(or columns) in the distance matrix.}

\item{i}{The row of the distance matrix to use in the ROC calculation.}
}
\value{
Area Under the curve.
}
\description{
Embedding quality measure.
}
\details{
The ROC curve plots the true positive rate vs false positive rate.
This function calculates the curve with the label of the specified
observation set as the positive class. The other observations are then
ranked according to their distance from the ith observation
(lower distances being better). Observations with the same label as the
specified observation count as the positive observations.

Perfect retrieval results in an AUC of 1. For random retrieval gives a value
of 0.5.
}
\note{
Use of this function requires that the \code{PRROC} package be
installed.
}

