% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{root_bisect}
\alias{root_bisect}
\title{Find root of function by bisection.}
\usage{
root_bisect(fn, tol = 1e-05, max_iters = 50, x_lower, x_upper,
  x_init = (x_lower + x_upper)/2)
}
\arguments{
\item{fn}{Function to optimize. Should take one scalar numeric value and
return a list containing at least one element called \code{y}, which should
contain the scalar numeric value. The list may contain other elements, so
that other useful values can be returned from the search routine.}

\item{tol}{Tolerance. If \eqn{abs(y_target - y) < tol} then the search will
terminate.}

\item{max_iters}{Maximum number of iterations to search for.}

\item{x_lower}{Lower bracket of x.}

\item{x_upper}{Upper bracket of x.}

\item{x_init}{Initial value of x.}
}
\value{
a list containing:
 \item{\code{x}}{Optimized value of x.}
 \item{\code{y}}{Value of y at convergence or \code{max_iters}}
 \item{\code{iter}}{Number of iterations at which convergence was reached.}
 \item{\code{best}}{Return value of calling \code{fn} with the optimized
 value of \code{x}.}
}
\description{
the root.
}

