% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_external.R
\name{ropt_step}
\alias{ropt_step}
\title{One Round of Optimization}
\usage{
ropt_step(opt, method, inp, out, iter)
}
\arguments{
\item{opt}{Optimizer}

\item{method}{Embedding method.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{iter}{Iteration number.}
}
\value{
List consisting of:
  \item{\code{opt}}{Updated optimizer.}
  \item{\code{inp}}{Updated input.}
  \item{\code{out}}{Updated output.}
}
\description{
One Round of Optimization
}

