% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{rtsne}
\alias{rtsne}
\title{"Reverse" t-Distributed Stochastic Neighbor Embedding (RTSNE).}
\usage{
rtsne(eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
Like \code{\link{tsne}}, but with the cost function using the "reverse" form
of the Kullback-Leibler divergence, i.e. KL(Q||P).

The probability matrix in RTSNE:

\itemize{
 \item{represents one probability distribution, i.e. the grand sum of the
 matrix is one.}
 \item{is symmetric, i.e. \code{P[i, j] == P[j, i]} and therefore the
 probabilities are joint probabilities.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix based on the weight matrix
 \code{wm}.}
 \item{\code{wm}}{Weight matrix generated from the distances between points
 in \code{ym}.}
}
}
\examples{
\dontrun{
embed_prob(method = tsne(), ...)
}
}
\seealso{
RTSNE uses the \code{\link{reverse_kl_cost}} cost function and the
  \code{\link{tdist_weight}} similarity function. The return value of this
  function should be used with the \code{\link{embed_prob}} embedding
  function.

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{hsjse}},
  \code{\link{hsnerv}}, \code{\link{hssne}},
  \code{\link{jse}}, \code{\link{mmds}},
  \code{\link{nerv}}, \code{\link{rasne}},
  \code{\link{rssne}}, \code{\link{sammon_map}},
  \code{\link{sjse}}, \code{\link{smmds}},
  \code{\link{snerv}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hsjse}}, \code{\link{hsnerv}},
  \code{\link{hssne}}, \code{\link{jse}},
  \code{\link{nerv}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne}}, \code{\link{rssne}},
  \code{\link{sjse}}, \code{\link{snerv}},
  \code{\link{ssne}}, \code{\link{tasne}},
  \code{\link{tnerv}}, \code{\link{tsne}}
}

