% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{sammon_map}
\alias{sammon_map}
\title{Sammon Mapping}
\usage{
sammon_map(eps = .Machine$double.eps)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
a list containing:
 \item{\code{cost_fn}}{Cost function for the embedding: Sammon's stress.}
 \item{\code{stiffness_fn}}{Stiffness function for \code{cost_fn}. For
 Sammon's stress, we ignore the constant denominator.}
 \item{\code{update_out_fn}}{Function to calculate and store any needed
 data after a coordinate update.}
 \item{\code{eps}}{Small floating point value used to prevent numerical
 problems, e.g. in gradients and cost functions.}
}
\description{
A distance-based embedding method.
}
\details{
This function minimizes an embedding using a quadratic loss function on
the difference between the input distances and the output distances:

\deqn{S = \frac{\sum_{i<j}\frac{(r_{ij} - d_{ij})^2}{r_{ij}}}
{\sum_{i<j} r_{ij}}}{S = sum(((rij-dij)/rij)^2)/sum(rij)}

\eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{d_{ij}}{dij} is the corresponding output distance.

This puts a greater emphasis on short distances over long distances, compared
to MDS. Note that the denominator is a constant, and is just there to
normalize the stress. It's not used for calculating the stiffness matrix
or the gradient, but is included in the calculation of the cost function.
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{dm}}{Distance matrix generated from \code{ym}.}
}
}
\seealso{
\code{\link[MASS]{sammon}}, which also carries out Sammon mapping.
Results should be comparable with those of a sneer embedding. The
\code{stress} value is equivalent to the \code{\link{sammon_stress_cost}}
function in sneer.

Other sneer distance embedding methods: \code{\link{distance_embedding_methods}},
  \code{\link{mmds}}, \code{\link{smmds}}

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{hsnerv}},
  \code{\link{hssne}}, \code{\link{mmds}},
  \code{\link{nerv}}, \code{\link{smmds}},
  \code{\link{snerv}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}
}

