% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{sammon_stress}
\alias{sammon_stress}
\title{Sammon Stress}
\usage{
sammon_stress(dxm, dym, eps = .Machine$double.eps)
}
\arguments{
\item{dxm}{Distance matrix.}

\item{dym}{Distance matrix, must be of the same dimensions as \code{dxm}.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
the Sammon stress between the input and output distances.
}
\description{
A measure of embedding quality between input and output distance data.
}
\details{
The Sammon stress has a similar form to a normalized STRESS cost
function used in metric MDS:

\deqn{S = \frac{\sum_{i<j}\frac{(r_{ij} - d_{ij})^2}{r_{ij}}}
{\sum_{i<j} r_{ij}}}{S = sum(((rij-dij)/rij)^2)/sum(rij)}

where \eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{d_{ij}}{dij} is the corresponding output distance.

Like the Kruskal stress, the Sammon stress is dimensionless. The main
difference is that the individual stresses are weighted by the reciprocal of
the input distance. Compared to MDS, this places a greater weight on
reproducing short distances over long distances.
}

