% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{sammon_stress_cost}
\alias{sammon_stress_cost}
\title{Sammon Stress Cost Function}
\usage{
sammon_stress_cost(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Sammon Stress of the input and output distances.
}
\description{
A measure of embedding quality between input and output distance data.
}
\details{
The Sammon stress has a similar form to a normalized STRESS cost
function used in metric MDS:

\deqn{S = \frac{\sum_{i<j}\frac{(r_{ij} - d_{ij})^2}{r_{ij}}}
{\sum_{i<j} r_{ij}}}{S = sum(((rij-dij)/rij)^2)/sum(rij)}

where \eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{d_{ij}}{dij} is the corresponding output distance.

Like the Kruskal stress, the Sammon stress is dimensionless. The main
difference is that the individual stresses are weighted by the reciprocal of
the input distance. Compared to MDS, this places a greater weight on
reproducing short distances over long distances.

This cost function requires the following matrices to be defined:
\describe{
 \item{\code{inp$dm}}{Input distances.}
 \item{\code{out$dm}}{Output distances.}
}
}
\seealso{
Other sneer cost functions: \code{\link{jse_cost}},
  \code{\link{kruskal_stress_cost}},
  \code{\link{mean_relative_error_cost}},
  \code{\link{metric_sstress_cost}},
  \code{\link{metric_stress_cost}},
  \code{\link{nerv_cost}},
  \code{\link{normalized_stress_cost}},
  \code{\link{reverse_kl_cost}},
  \code{\link{rms_metric_stress_cost}},
  \code{\link{sammon_stress_unnorm_cost}}
}

