% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{sammon_stress_unnorm}
\alias{sammon_stress_unnorm}
\title{Unnormalized Sammon Stress}
\usage{
sammon_stress_unnorm(dxm, dym, eps = .Machine$double.eps)
}
\arguments{
\item{dxm}{Distance matrix.}

\item{dym}{Distance matrix, must be of the same dimensions as \code{dxm}.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
the unnormalized Sammon stress between the input and output
 distances.
}
\description{
A measure of embedding quality between input and output distance data.
}
\details{
The Sammon stress is defined as:

\deqn{S = \frac{\sum_{i<j}\frac{(r_{ij} - d_{ij})^2}{r_{ij}}}
{\sum_{i<j} r_{ij}}}{S = sum(((rij-dij)/rij)^2)/sum(rij)}

where \eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{d_{ij}}{dij} is the corresponding output distance.

Note that the denonimator of the stress only contains input distances, and
so is constant with respect to an embedding of a dataset. This version of
the function dispenses with that part of the calculation:

\deqn{S_{unnorm} = \sum_{i<j}\frac{(r_{ij} - d_{ij})^2}{r_{ij}}}
{S = sum(((rij-dij)/rij)^2)}

This is marginally faster, and is consistent with the analytical gradient
calculation in \code{\link{sammon_map}}.
}

