% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cost.R
\name{sammon_stress_unnorm_cost}
\alias{sammon_stress_unnorm_cost}
\title{Unnormalized Sammon Stress}
\usage{
sammon_stress_unnorm_cost(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Sammon Stress of the input and output distances.
}
\description{
A measure of embedding quality between input and output distance data.
}
\details{
The Sammon stress is defined as:

\deqn{S = \frac{\sum_{i<j}\frac{(r_{ij} - d_{ij})^2}{r_{ij}}}
{\sum_{i<j} r_{ij}}}{S = sum(((rij-dij)/rij)^2)/sum(rij)}

where \eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{d_{ij}}{dij} is the corresponding output distance.

Note that the denonimator of the stress only contains input distances, and
so is constant with respect to an embedding of a dataset. This version of
the function dispenses with that part of the calculation:

\deqn{S_{unnorm} = \sum_{i<j}\frac{(r_{ij} - d_{ij})^2}{r_{ij}}}
{S = sum(((rij-dij)/rij)^2)}

This is marginally faster, and is consistent with the analytical gradient
calculation in \code{\link{sammon_map}}.
}
\seealso{
Other sneer cost functions: \code{\link{jse_cost}},
  \code{\link{kl_cost}}, \code{\link{kruskal_stress_cost}},
  \code{\link{mean_relative_error_cost}},
  \code{\link{metric_sstress_cost}},
  \code{\link{metric_stress_cost}},
  \code{\link{nerv_cost}},
  \code{\link{normalized_stress_cost}},
  \code{\link{reverse_kl_cost}},
  \code{\link{rms_metric_stress_cost}},
  \code{\link{sammon_stress_cost}}
}

