% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scale_distances}
\alias{scale_distances}
\title{Input Distance Scaling}
\usage{
scale_distances(dm, verbose = TRUE)
}
\arguments{
\item{dm}{Distance matrix}

\item{verbose}{If \code{TRUE}, information about the scaled distances will be
logged.}
}
\value{
Scaled distance matrix.
}
\description{
Preprocess function for distance matrix.
}
\details{
Scales the input distances so that the mean distance is 1. This preprocessing
step is recommended as part of the \code{\link{nerv}} embedding method.
}
\references{
Venna, J., Peltonen, J., Nybo, K., Aidos, H., & Kaski, S. (2010).
Information retrieval perspective to nonlinear dimensionality reduction for
data visualization.
}

