% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{scale_input_distances}
\alias{scale_input_distances}
\title{Input Distance Scaling}
\usage{
scale_input_distances(verbose = TRUE)
}
\arguments{
\item{verbose}{If \code{TRUE}, information about the scaled distances will be
logged.}
}
\value{
Input initialization method for use by \code{\link{make_init_inp}}.
}
\description{
An initialization method for creating input probabilities.
}
\details{
This initialization method scales the input distances so that the average
distance is 1. Therefore, if applying other input initializers that use
the distances (e.g. to calculate probabilities), this initializer should
appear before them in the \code{\link{make_init_inp}} parameter list.
}
\examples{
# Should be passed to the init_inp argument of an embedding function:
init_inp = make_init_inp(scale_input_distances())

# Should appear before other use of the input distances
init_inp = make_init_inp(scale_input_distances(),
                         prob_perp_bisect(perplexity = 30,
                                         input_weight_fn = exp_weight))
}
\references{
Venna, J., Peltonen, J., Nybo, K., Aidos, H., & Kaski, S. (2010).
Information retrieval perspective to nonlinear dimensionality reduction for
data visualization.
}
\seealso{
\code{\link{make_init_inp}} for how to use this function as part
of an embedding.

Other sneer input initializers: \code{\link{input_initializers}},
  \code{\link{prob_perp_bisect}}
}

