% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{scale_prec_to_perp}
\alias{scale_prec_to_perp}
\title{Scale Output Kernel Precision Based on Intrinsic Dimensionality}
\usage{
scale_prec_to_perp(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\description{
Scale Output Kernel Precision Based on Intrinsic Dimensionality
}
\seealso{
Other sneer kernel modifiers: \code{\link{transfer_kernel_bandwidths}}
}

