% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scatterqplot}
\alias{scatterqplot}
\title{Scatterplot Using ggplot2 and ColorBrewer Palettes}
\usage{
scatterqplot(df, x, y, label_name = "Label", size = 1, palette = "Set1",
  x_label = "x", y_label = "y", legend = TRUE, legend_rows = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{x}{Vector of x values to plot for the x-coordinate.}

\item{y}{Vector of y values to plot for the y-coordinate.}

\item{label_name}{name of the column that contains a value to be used
to color each point.}

\item{size}{Size of the points.}

\item{palette}{String giving the name of a ColorBrewer Palette. To see the
available palettes run the function
\code{RColorBrewer::display.brewer.all()}. Although subject to change, the
following palettes from the "qualitative" subset are suggested:
\itemize{
  \item \code{Set1}
  \item \code{Set2}
  \item \code{Set3}
  \item \code{Pastel1}
  \item \code{Pastel2}
  \item \code{Dark2}
}}

\item{x_label}{Label for the x-axis.}

\item{y_label}{Label for the y-axis.}

\item{legend}{If \code{TRUE}, then the legend will be displayed. Set to
\code{FALSE} if there are a lot of separate categories that would appear
in the legend, which can result in the legend taking up more space than
the actual plot.}

\item{legend_rows}{If non-null, then sets the number of rows to display
the legend items in. If the legend is taking up too much space, you may
want to experiment with setting the number of rows manually, rather than
just setting the \code{legend} parameter to \code{FALSE}.}
}
\description{
Scatterplot using ggplot2 with coloring via categorical data and a palette
from RColorBrewer.
}
\note{
Use of this function requires that the \code{ggplot2} and
\code{RColorBrewer} packages be installed.
}
\seealso{
More information on ColorBrewer is available at its website,
\url{http://www.colorbrewer2.org}.
}

