% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sclamp}
\alias{sclamp}
\title{Clamp Scalar Numerical Value}
\usage{
sclamp(x, min, max)
}
\arguments{
\item{x}{Value.}

\item{min}{If \code{x} is smaller than this value, it will be truncated to
this value.}

\item{max}{If \code{x} is larger than this value, it will be truncated to
this value.}
}
\value{
Clamped value.
}
\description{
Value is truncated so that it lies within (\code{min, max}).
}

