% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_out.R
\name{scores_matrix}
\alias{scores_matrix}
\title{PCA Scores Matrix}
\usage{
scores_matrix(xm, ncol = min(nrow(xm), base::ncol(xm)), verbose = TRUE)
}
\arguments{
\item{xm}{Matrix to carry out PCA on. No scaling is carried out.}

\item{ncol}{The number of score columns to include in the output matrix.
Cannot be larger than the smaller of the number of rows or columns. Columns
are included in order of decreasing eigenvalue.}

\item{verbose}{If true, then information about variance explained by the
chosen number of columns will be logged.}
}
\value{
A column matrix of the PCA scores. The number of rows of the matrix
is the same as that of \code{xm}.
}
\description{
PCA Scores Matrix
}
\examples{
\dontrun{
# first two components of PCA
scores <- scores_matrix(iris[, 1:4], ncol = 2)
# all scores
scores <- scores_matrix(iris[, 1:4])
}
}

