% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scores_plot}
\alias{scores_plot}
\title{PCA Scores Plot}
\usage{
scores_plot(df, pca_indexes, label_name = "Label", center = TRUE,
  scale = FALSE, cex = 0.5)
}
\arguments{
\item{df}{Data frame.}

\item{pca_indexes}{Numeric vector containing the column indexes in the
data frame which should be used for the PCA calculation.}

\item{label_name}{name of the column that contains a value to be used
to color each point.}

\item{center}{If \code{TRUE}, mean-center the columns.}

\item{scale}{If \code{TRUE}, scale the columns to have unit variance.}

\item{cex}{Numeric \strong{c}haracter \strong{ex}pansion factor;
multiplied by \code{\link[graphics]{par}("cex")} yields the final
character size of the labels.}
}
\description{
Carries out PCA on a data frame (using the provided indexes) and then
plots the first two scores, coloring the points with a user-specified
label.
}
\examples{
\dontrun{
 # PCA on the scaled iris dataset, use the "Species" column to display
 scores_plot(iris, 1:4, "Species", scale = TRUE)

 # PCA on the s1k dataset, use default "Label" column and change label size
 # to 1.
 scores_plot(s1k, 1:9, cex = 1)
}
}

