% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scores_qplot}
\alias{scores_qplot}
\title{PCA Scores Plot Using ggplot2 and ColorBrewer Palettes}
\usage{
scores_qplot(df, pca_indexes = NULL, label_name = NULL, center = TRUE,
  scale = FALSE, size = 1, palette = "Set1", legend = TRUE,
  legend_rows = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{pca_indexes}{Numeric vector containing the column indexes in the
data frame which should be used for the PCA calculation. Default is
\code{NULL}, in which case all numeric columns will be used.}

\item{label_name}{name of the column that contains a value to be used
to color each point. Default is \code{NULL}, in which case the first
encountered factor column will be used.}

\item{center}{If \code{TRUE}, mean-center the columns.}

\item{scale}{If \code{TRUE}, scale the columns to have unit variance.}

\item{size}{Size of the points.}

\item{palette}{String giving the name of a ColorBrewer Palette. To see the
available palettes run the function
\code{RColorBrewer::display.brewer.all()}. Although subject to change, the
following palettes from the "qualitative" subset are suggested:
\itemize{
  \item \code{Set1}
  \item \code{Set2}
  \item \code{Set3}
  \item \code{Pastel1}
  \item \code{Pastel2}
  \item \code{Dark2}
}}

\item{legend}{If \code{TRUE}, then the legend will be displayed. Set to
\code{FALSE} if there are a lot of separate categories that would appear
in the legend, which can result in the legend taking up more space than
the actual plot.}

\item{legend_rows}{If non-null, then sets the number of rows to display
the legend items in. If the legend is taking up too much space, you may
want to experiment with setting the number of rows manually, rather than
just setting the \code{legend} parameter to \code{FALSE}.}
}
\description{
Carries out PCA on a dataset and then displays the first two scores in a
scatterplot using ggplot2 and a color palette from RColorBrewer.
}
\note{
Use of this function requires that the \code{ggplot2} and
\code{RColorBrewer} packages be installed.
}
\examples{
 \dontrun{
 # PCA on the scaled iris dataset, use the "Species" column to display
 scores_qplot(iris, 1:4, "Species", scale = TRUE)

 # PCA on s1k dataset, use all numeric indices and first factor (defaults)
 # for coloring with the "Set3" palette, set point size to 2,
 # and display three rows in the legend
 scores_qplot(s1k, size = 2, palette = "Set3", legend_rows = 3)
}
}
\seealso{
More information on ColorBrewer is available at its website,
\url{http://www.colorbrewer2.org}.
}

