% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{set_solution}
\alias{set_solution}
\title{Set Output Data Coordinates}
\usage{
set_solution(opt, inp, coords, method)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{coords}{Matrix of coordinates.}

\item{method}{Embedding method.}
}
\value{
Updated \code{out}.
}
\description{
This function sets the embedded coordinates in the output data, as well as
recalculating any auxiliary output data that is dependent on the coordinates
(e.g. distances and probabilities)
}

