% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{shannon_entropy_row}
\alias{shannon_entropy_row}
\title{Shannon Entropy per Row of a Matrix}
\usage{
shannon_entropy_row(pm, base = 2)
}
\arguments{
\item{pm}{Matrix of probabilities.}

\item{base}{Base of the logarithm to use in the entropy calculation.}
}
\value{
Vector of Shannon entropies, one per row of the matrix.
}
\description{
Each row of the matrix should consist of probabilities summing to 1.
}

