% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{should_update}
\alias{should_update}
\title{Check if Output Data Needs Updating}
\usage{
should_update(out)
}
\arguments{
\item{out}{Output data.}
}
\value{
\code{TRUE} if the output data matrices need updating.
}
\description{
When the coordinates of an embedding are updated via
\code{\link{set_solution}}, the other matrices and values that depend on
those coordinates should be updated automatically. If a change has occurred
to another part of the embedding data (most likely the input probabilities,
or the output kernel parameters), then you may need to call this function to
see if the output data needs updating before calculating the cost or
gradient.
}

