% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{show_frey_face}
\alias{show_frey_face}
\title{Visualize Frey face.}
\usage{
show_frey_face(df, n, col = gray(1/12:1))
}
\arguments{
\item{df}{Data frame containing the Frey faces.}

\item{n}{Frame index of the image to display.}

\item{col}{List of colors to use in the display.}
}
\description{
Display an image from the Frey faces dataset.
}

