% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{show_olivetti_face}
\alias{show_olivetti_face}
\title{Visualize Olivetti face.}
\usage{
show_olivetti_face(df, face, pose, col = gray(1/12:1))
}
\arguments{
\item{df}{Data frame containing the Olivetti faces.}

\item{face}{Face index of the image to display. Must be an integer between
1 and 400.}

\item{pose}{Pose index of the image to display. Must be an integer between
1 and 10.}

\item{col}{List of colors to use in the display.}
}
\description{
Display an image from the Olivetti faces dataset.
}

