% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{sjse_plugin}
\alias{sjse_plugin}
\title{An implementation of SJSE using the plugin gradient.}
\usage{
sjse_plugin(kappa = 0.5, beta = 1, eps = .Machine$double.eps)
}
\arguments{
\item{kappa}{Mixture parameter. If set to 0, then JSE behaves like SSNE. If
set to 1, then JSE behaves like RSSNE.}

\item{beta}{Precision parameter of the exponential similarity kernel
function.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
An implementation of SJSE using the plugin gradient.
}
\seealso{
\code{\link{sjse}} should give equivalent results, but is probably
a bit more efficient.

Other sneer embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{embedding_methods}},
  \code{\link{hsjse_plugin}}, \code{\link{hsjse}},
  \code{\link{hsnerv_plugin}}, \code{\link{hsnerv}},
  \code{\link{hssne_plugin}}, \code{\link{hssne}},
  \code{\link{jse_plugin}}, \code{\link{jse}},
  \code{\link{mmds}}, \code{\link{nerv_plugin}},
  \code{\link{nerv}}, \code{\link{plugin}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sammon_map}}, \code{\link{sjse}},
  \code{\link{smmds}}, \code{\link{snerv_plugin}},
  \code{\link{snerv}}, \code{\link{ssne_plugin}},
  \code{\link{ssne}}, \code{\link{tasne_plugin}},
  \code{\link{tasne}}, \code{\link{tnerv_plugin}},
  \code{\link{tnerv}}, \code{\link{tsne_plugin}},
  \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsjse}}, \code{\link{hsnerv_plugin}},
  \code{\link{hsnerv}}, \code{\link{hssne_plugin}},
  \code{\link{hssne}}, \code{\link{jse_plugin}},
  \code{\link{jse}}, \code{\link{nerv_plugin}},
  \code{\link{nerv}}, \code{\link{plugin}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sjse}}, \code{\link{snerv_plugin}},
  \code{\link{snerv}}, \code{\link{ssne_plugin}},
  \code{\link{ssne}}, \code{\link{tasne_plugin}},
  \code{\link{tasne}}, \code{\link{tnerv_plugin}},
  \code{\link{tnerv}}, \code{\link{tsne_plugin}},
  \code{\link{tsne}}
}

