% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jse.R
\name{sjse_stiffness}
\alias{sjse_stiffness}
\title{Symmetric JSE Stiffness Function}
\usage{
sjse_stiffness(qm, zm, kl_qz, kappa = 0.5, eps = .Machine$double.eps)
}
\arguments{
\item{qm}{Output probabilty matrix.}

\item{zm}{Mixture matrix, weighted combination of input probability and
output probability matrix \code{qm}.}

\item{kl_qz}{KL divergence between \code{qm} and \code{zm}. \code{qm} is the
reference probability.}

\item{kappa}{Mixture parameter. Should be a value between 0 and 1 and be the
same value used to produce the mixture matrix \code{zm}.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix.
}
\description{
Symmetric JSE Stiffness Function
}

