% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sphere}
\alias{sphere}
\title{Sphere data set.}
\usage{
sphere(n = 1000, nlabels = 5)
}
\arguments{
\item{n}{The number of points to create.}

\item{nlabels}{Number of labels.}
}
\value{
Data frame.
}
\description{
Simulation data.
}
\details{
Creates a series of points sampled from a 3D spherical surface.

Labels go from \code{0} to \code{nlabels - 1}. Points are classified into
labels based on the angle theta in their spherical coordinate presentation.
You can think of it as dividing the sphere surface like the segments of an
orange.
}
\references{
A dataset like this was used in:
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}

