% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{sqrt_exp_weight}
\alias{sqrt_exp_weight}
\title{Exponential weighting of the distances.}
\usage{
sqrt_exp_weight(d2m, beta = 1)
}
\arguments{
\item{d2m}{Matrix of squared distances.}

\item{beta}{Exponential parameter.}
}
\value{
Weight matrix.
}
\description{
Applies exponential weighting to the distances, rather than the square of
the distances. Included so results can be compared with other implementations
of TSNE, although you'd normally use the squared distances.
}
\details{
\eqn{W = e^{-\beta D}}
}

