% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_direction.R
\name{steepest_descent}
\alias{steepest_descent}
\title{Steepest Descent direction.}
\usage{
steepest_descent()
}
\value{
Steepest Descent direction method. A list consisting of:
\itemize{
 \item \code{get_direction} Function invoked by the optimizer to find the direction
 to move in the gradient descent part of the solution update.
}
}
\description{
Creates a gradient descent direction method for use by the optimizer.
Uses the direction of steepest descent.
}

