% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolfe.R
\name{step_armijo_ok}
\alias{step_armijo_ok}
\title{Does Step Fulfil the Armijo Rule}
\usage{
step_armijo_ok(step0, step, c1)
}
\arguments{
\item{step0}{Line function info at start point.}

\item{step}{Line function info at a point.}

\item{c1}{Sufficient decrease constant.}
}
\value{
\code{TRUE} if \code{step} fulfils the Armijo Rule.
}
\description{
Line Search Criterion.
}

