% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolfe.R
\name{step_curvature_ok}
\alias{step_curvature_ok}
\title{Does Step Fulfil the Standard Curvature Rule}
\usage{
step_curvature_ok(step0, step, c2)
}
\arguments{
\item{step0}{Line function info at start point.}

\item{step}{Line function info at a point along the line searcht.}

\item{c2}{Curvature rule constant.}
}
\value{
\code{TRUE} if \code{step_a} fulfils the curvature rule.
}
\description{
Line Search Criterion.
}

