% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{step_momentum}
\alias{step_momentum}
\title{Step momentum.}
\usage{
step_momentum(init_momentum = 0.5, final_momentum = 0.8,
  switch_iter = 250, verbose = TRUE)
}
\arguments{
\item{init_momentum}{Momentum value for the first \code{switch_iter}
iterations.}

\item{final_momentum}{Momentum value after \code{switch_iter} iterations.}

\item{switch_iter}{Iteration number at which to switch from
\code{init_momentum} to \code{final_momentum}.}

\item{verbose}{if \code{TRUE}, log info about the momentum.}
}
\value{
Step momentum update method, to be used by the optimizer.
}
\description{
Factory function for creating an optimizer update method.
}
\details{
Create a callback for the optimizer to use to update the embedding solution.
Update is in the form of a step momentum function.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# update method:
make_opt(update = step_momentum())
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{\link{optimizer_update}} for details
on the functions and values defined for this method.

Other sneer optimization update methods: \code{\link{constant_momentum}},
  \code{\link{linear_momentum}},
  \code{\link{nesterov_nsc_momentum}},
  \code{\link{no_momentum}}
}

