% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{step_perps}
\alias{step_perps}
\title{Perplexity Values for Step Embedding}
\usage{
step_perps(x, num_perps = 5)
}
\arguments{
\item{x}{Data frame or matrix.}

\item{num_perps}{Number of perplexities.}
}
\value{
Vector containing \code{num_perps} perplexities.
}
\description{
Returns a vector of perplexity values scaled according to the size of the
data frame or matrix suitable for use in a step wise scaling of perplexity.
}
\details{
The perplexities will be returned in descending order, starting at N/2,
where N is the number of observations in the data set, and ending at 32
(or N/4, if there are less than 64 observations in the dataset), with
the perplexities in between equally spaced.
}

