% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolfe.R
\name{step_strong_wolfe_ok}
\alias{step_strong_wolfe_ok}
\title{Does Step Fulfil the Strong Wolfe Condtions}
\usage{
step_strong_wolfe_ok(step0, step, c1, c2)
}
\arguments{
\item{step0}{Line function info at start point.}

\item{step}{Line function info at a point along the line searcht.}

\item{c1}{Sufficient decrease constant.}

\item{c2}{Curvature decrease constant.}
}
\value{
\code{TRUE} if \code{step_a} fulfils the strong Wolfe conditions.
}
\description{
Line Search Criterion.
}

