% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{stiff_to_grads}
\alias{stiff_to_grads}
\title{Convert stiffness matrix to gradient matrix.}
\usage{
stiff_to_grads(ym, km)
}
\arguments{
\item{ym}{Embedded coordinates.}

\item{km}{Stiffness matrix.}
}
\value{
Gradient matrix.
}
\description{
Convert stiffness matrix to gradient matrix.
}

