% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolfe.R
\name{strong_curvature_ok}
\alias{strong_curvature_ok}
\title{Strong Curvature Condition}
\usage{
strong_curvature_ok(d0, da, c2)
}
\arguments{
\item{d0}{Value of gradient of the line at the starting position of the line
search.}

\item{da}{Value of gradient of the line at a distance along the line.}

\item{c2}{Sufficient decrease constant.}
}
\value{
\code{TRUE} if the strong curvature condition is met.
}
\description{
Strong Curvature Condition
}

