% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolfe.R
\name{strong_wolfe_ok}
\alias{strong_wolfe_ok}
\title{Strong Wolfe Condition}
\usage{
strong_wolfe_ok(f0, d0, alpha, fa, da, c1, c2)
}
\arguments{
\item{f0}{Function value at the starting position of the line search.}

\item{d0}{Gradient value at the starting position of the line search.}

\item{alpha}{Step length.}

\item{fa}{Function value at distance \code{alpha} along the line.}

\item{da}{Gradient value at distance \code{alpha} along the line.}

\item{c1}{Sufficient decrease constant.}

\item{c2}{Curvature decrease constant.}
}
\value{
\code{TRUE} if the strong Wolfe conditions are met.
}
\description{
Strong Wolfe Condition
}

