% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summarize}
\alias{summarize}
\title{Summary of distribution of data.}
\usage{
summarize(vals, msg = "")
}
\arguments{
\item{vals}{Array or matrix of data.}

\item{msg}{Label to identify the data summary.}
}
\description{
Summary of distribution of data.
}

