% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{summarize_betas}
\alias{summarize_betas}
\title{Logs parameter search information.}
\usage{
summarize_betas(betas)
}
\arguments{
\item{betas}{Vector of parameters}
}
\description{
Provides summary of the distribution of the beta parameter used to
generate input probabilities in SNE.
}
\details{
Mainly useful for debugging. Also expresses beta as sigma, i.e. the
Gaussian bandwidth \eqn{\frac{1}{\sqrt{2\beta}}}.
}

