% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{swiss_roll}
\alias{swiss_roll}
\title{Swiss Roll Data Set}
\usage{
swiss_roll(n = 1000, nlabels = 6)
}
\arguments{
\item{n}{Number of points to create.}

\item{nlabels}{Number of labels.}
}
\value{
Data frame.
}
\description{
Simulation data.
}
\details{
Creates a series of points randomly sampled from a swiss roll-shaped
manifold: a two-dimensional plane which has been rolled up into a spiral
shape. Or just look at a swiss roll.

The formula for sampling the x, y and z coordinates used in this dataset is
from that given in the Stochastic Proximity Embedding paper by Agrafiotis
and Xu:
\deqn{x = \phi cos\phi, y = \phi sin \phi, z}
{x = phi * cos(phi), y = phi * sin(phi), z}

where \eqn{\phi}{phi} and z are random numbers in the intervals [5, 13]
and [0, 10], respectively.

Labels go from \code{0} to \code{nlabels - 1}. Points are classified into
labels based on their z-coordinate.
}
\references{
Agrafiotis, D. K., & Xu, H. (2002).
A self-organizing principle for learning nonlinear manifolds.
\emph{Proceedings of the National Academy of Sciences}, \emph{99}(25), 15869-15872.
}

