% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{symm_weights_to_pjoint}
\alias{symm_weights_to_pjoint}
\title{Create Joint Probability Matrix from Symmetric Weight Matrix}
\usage{
symm_weights_to_pjoint(wm)
}
\arguments{
\item{wm}{Symmetric matrix of weighted distances.}
}
\value{
List containing:
\item{pm}{Joint probability matrix.}
}
\description{
A weight to probability function.
}
\details{
Creates a joint probability matrix: the grand sum of the elements are
one, and the matrix is symmetric. The weight matrix must be symmetric. Used
in SSNE and t-SNE to generate the output probabilities.
}

