% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{tasne}
\alias{tasne}
\title{t-distributed Asymmetric Stochastic Neighbor Embedding (t-ASNE)}
\usage{
tasne(eps = .Machine$double.eps)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
Creates a list of functions that collectively implement t-ASNE: a method that
I just made up to illustrate how to explore different aspects of embedding
within sneer: this uses the t-distributed distance weighting of t-SNE, but
for probability generation uses the point-wise distribution of ASNE.
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Row probability matrix based on the weight matrix
 \code{wm}.}
 \item{\code{wm}}{Weight matrix generated from the distances between points
 in \code{ym}.}
}
}
\examples{
\dontrun{
embed_prob(method = tasne(), ...)
}
}
\seealso{
t-ASNE uses the \code{\link{exp_weight}} similarity function.
The return value of this function should be used with the
\code{\link{embed_prob}} embedding function.

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{hssne}},
  \code{\link{mmds}}, \code{\link{sammon_map}},
  \code{\link{smmds}}, \code{\link{ssne}},
  \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hssne}},
  \code{\link{probability_embedding_methods}},
  \code{\link{ssne}}, \code{\link{tsne}}
}

