% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{tasne_stiffness}
\alias{tasne_stiffness}
\title{t-ASNE Stiffness Function}
\usage{
tasne_stiffness(pm, qm, wm, beta = 1)
}
\arguments{
\item{pm}{Input probability matrix.}

\item{qm}{Output probabilty matrix.}

\item{wm}{Output weight probability matrix.}

\item{beta}{The precision of the weighting function.}
}
\value{
Stiffness matrix.
}
\description{
t-ASNE Stiffness Function
}

