% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{tdist_gr}
\alias{tdist_gr}
\title{Exponential Gradient}
\usage{
tdist_gr(d2m)
}
\arguments{
\item{d2m}{Matrix of squared distances.}
}
\value{
Matrix containing the gradient (with respect to \code{d2m}).
}
\description{
t-Distributed Kernel Gradient.
}
\details{
Calculates the gradient of the Student-t distribution with one degree of
freedom with respect to d2m, the matrix of squared distances.
}

