% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{tdist_kernel}
\alias{tdist_kernel}
\title{t-Distribution Kernel Factory Function}
\usage{
tdist_kernel()
}
\value{
t-Distributed function and gradient.
}
\description{
Similarity Kernel factory function.
}
\details{
Creates a list implementing the t-distributed kernel function and gradient.
}
\seealso{
Other sneer similiarity kernels: \code{\link{exp_kernel}},
  \code{\link{heavy_tail_kernel}}
}

