% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{tnerv_stiffness}
\alias{tnerv_stiffness}
\title{t-NeRV Stiffness Function}
\usage{
tnerv_stiffness(pm, qm, wm, rev_kl, lambda = 0.5, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}

\item{wm}{Output weight probability matrix.}

\item{rev_kl}{"Reverse" KL divergence between \code{pm} and \code{qm}.}

\item{lambda}{NeRV weighting factor controlling the emphasis placed on
precision versus recall.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix
}
\description{
t-NeRV Stiffness Function
}

