% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricks.R
\name{tricks}
\alias{tricks}
\title{Tricks}
\description{
Miscellaneous data manipulations that may help optimize the embedding.
If you are fancy, you could call them "heuristics", but the t-SNE paper
calls their approach "tricks" and that's less characters to type, so we'll
go with that too.
}
\examples{
# pass one or more tricks to the make_tricks factory function:
# no tricks
make_tricks()

# one trick
make_tricks(early_exaggeration())

# two tricks
make_tricks(early_exaggeration(), late_momentum())

\dontrun{
# in turn, pass the result of make_tricks to an embedding routine
embed_prob(make_tricks(early_exaggeration(), late_momentum()), ...)
}

}
\seealso{
Other sneer tricks: \code{\link{early_exaggeration}},
  \code{\link{late_momentum}}
}
\keyword{internal}

