% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_step.R
\name{tsne_jacobs}
\alias{tsne_jacobs}
\title{Jacobs step size method using parameters from the t-SNE paper.}
\usage{
tsne_jacobs()
}
\value{
A step size method suitable for use with t-SNE.
}
\description{
Factory function for creating an optimizer step size method.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# step size method:
make_opt(step_size = tsne_jacobs())
}
\references{
Laurens van der Maarten, Geoffrey Hinton.
Visualizing Data using t-SNE.
Journal of Machine Learning Research, 2008, 9, 2579-2605.

R. A. Jacobs.
Increased rates of convergence through learning rate adaptation.
Neural Networks, 1:295-307, 1988.
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{\link{optimizer_step_size}} for details
on the functions and values defined for this method.

Other sneer optimization step size methods: \code{\link{bold_driver}},
  \code{\link{jacobs}}
}

