% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{tsne_opt}
\alias{tsne_opt}
\title{t-SNE optimizer.}
\usage{
tsne_opt()
}
\value{
optimizer with parameters from the t-SNE paper.
}
\description{
Convenience factory function which sets the optimizer parameters to that
from the t-SNE paper.
}
\examples{
# Should be passed to the opt argument of an embedding function:
\dontrun{
 embed_sim(opt = tsne_opt(), ...)
}
}
\seealso{
\code{\link{embed_sim}} for how to use this function for configuring
an embedding, and \code{\link{make_opt}} for a more generic function.
}

