% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{tsne_stiffness}
\alias{tsne_stiffness}
\title{t-SNE Stiffness Function}
\usage{
tsne_stiffness(pm, qm, wm, beta = 1)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}

\item{wm}{Output weight probability matrix.}

\item{beta}{The precision of the weighting function.}
}
\value{
Stiffness matrix.
}
\description{
t-SNE Stiffness Function
}

