% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricks.R
\name{tsne_tricks}
\alias{tsne_tricks}
\title{Tricks from t-SNE paper.}
\usage{
tsne_tricks(verbose = TRUE)
}
\arguments{
\item{verbose}{If \code{TRUE} print message about tricks during the
embedding.}
}
\value{
tricks callback parameterized to behave like the t-SNE paper.
}
\description{
Tricks from t-SNE paper.
}
\examples{
# Should be passed to the tricks argument of an embedding function:
\dontrun{
 embed_sim(tricks = tsne_tricks(), ...)
}
}
\seealso{
\code{\link{embed_sim}} for how to use this function for configuring
an embedding.

Other sneer tricks: \code{\link{make_tricks}}
}

