% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{uhsnerv_stiffness}
\alias{uhsnerv_stiffness}
\title{UHSNeRV Stiffness Function}
\usage{
uhsnerv_stiffness(pm, qm, wm, rev_kl, lambda = 0.5, alpha = 1.5e-08,
  beta = 1, eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}

\item{wm}{Output weight probability matrix.}

\item{rev_kl}{"Reverse" KL divergence between \code{pm} and \code{qm}.}

\item{lambda}{NeRV weighting factor controlling the emphasis placed on
precision versus recall.}

\item{alpha}{Tail heaviness of the weighting function.}

\item{beta}{The bandwidth of the weighting function.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix.
}
\description{
If using uniform bandwidths, the stiffness function for UHSNeRV is
simplified compared to the generic non-uniform case for HSNeRV.
}

