% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{undirty}
\alias{undirty}
\title{Mark Output Data as Up to Date}
\usage{
undirty(out)
}
\arguments{
\item{out}{Output data.}
}
\value{
Output data, now marked as being up to date.
}
\description{
This function should be called when the output data has just been updated.
Some parts of the code will avoid unnecessarily repeating the update, but
only if they can detect that there's no change to make.
}

