% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{update_out}
\alias{update_out}
\title{Update Output Data}
\usage{
update_out(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Updated output data.
}
\description{
Updates the output data (matrices dependent on the embedding coordinates,
such as the distance matrix, weight matrix, output probability matrix).
}

