% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{update_out_if_necessary}
\alias{update_out_if_necessary}
\title{Update if Something Has Changed}
\usage{
update_out_if_necessary(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
(Possible) updated output data. Otherwise, \code{out} is returned
unchanged.
}
\description{
This function calls \code{\link{update_out}} only if it detects that there
are changes that needs to be made.
}

