% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{update_probs}
\alias{update_probs}
\title{Update Output Probabilities}
\usage{
update_probs(out, method)
}
\arguments{
\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
List containing:
 \item{\code{d2m}}{Matrix of squared distances.}
 \item{\code{wm}}{Weight matrix.}
 \item{\code{qm}}{Probability Matrix.}
}
\description{
Intermediate data (squared distance matrix and weight matrix) may be useful
in calculating the gradient, so is also returned from this function.
}

