% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{update_solution}
\alias{update_solution}
\title{Output Data Update}
\usage{
update_solution(opt, inp, out, method)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Updated \code{out}.
}
\description{
This function updates the embedded coordinates in the output data, based
on the update information in the Optimizer, as well as updating any
auxiliary output data that is dependent on the coordinates (e.g. distances
and probabilities)
}

