% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{upper_tri}
\alias{upper_tri}
\title{Upper Triangle of a Matrix as a Vector}
\usage{
upper_tri(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
Vector of elements in the upper triangle of the matrix.
}
\description{
Useful if you need to get all of the distances in a distance matrix, without
including self-distances or double counting. However, all index information
is lost.
}

