% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{usnerv}
\alias{usnerv}
\title{Symmetric Neighbor Retrieval Visualizer with uniform bandwidths (USNeRV)}
\usage{
usnerv(lambda = 0.5, beta = 1, eps = .Machine$double.eps,
  verbose = TRUE)
}
\arguments{
\item{lambda}{Weighting factor controlling the emphasis placed on precision
(set \code{lambda} to 0), versus recall (set \code{lambda} to 1). If set to
1, then the method is equivalent to t-SNE. Must be a value between 0 and 1.}

\item{beta}{Bandwidth parameter of the exponential similarity kernel
function.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
USNeRV is a "symmetric" variant of \code{\link{unerv}}. Rather than use the
conditional point-based probabilities of \code{\link{asne}}, it uses the
joint pair-based probabilities of \code{\link{ssne}}. It differs from
\code{\link{snerv}} by only using a uniform bandwidth parameter to generate
the output weight matrix.

When \code{lambda = 1}, this method is equivalent to SSNE.

The probability matrix used in SNeRV:

\itemize{
 \item{represents one probability distribution, i.e. the grand sum of the
 matrix is one.}
 \item{is symmetric, i.e. \code{P[i, j] == P[j, i]} and therefore the
 probabilities are joint probabilities.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix.}
}
}
\examples{
\dontrun{
# default SNeRV settings
embed_prob(method = usnerv(lambda = 0.5), ...)

# equivalent to SSNE
embed_prob(method = usnerv(lambda = 1), ...)

# puts an emphasis on only keeping true neighbors close together
# tends to produce a larger number of small, tight clusters
embed_prob(method = usnerv(lambda = 0), ...)
}
}
\seealso{
\code{\link{snerv}} is a version of this algorithm which relaxes
the requirement for uniform precisions in the output weight kernel (at
the cost of being somewhat less efficient).

The return value of this function should be used with the
  \code{\link{embed_prob}} embedding function.

Other sneer embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{embedding_methods}},
  \code{\link{hsjse_plugin}}, \code{\link{hsjse}},
  \code{\link{hsnerv_plugin}}, \code{\link{hsnerv}},
  \code{\link{hssne_plugin}}, \code{\link{hssne}},
  \code{\link{jse_plugin}}, \code{\link{jse}},
  \code{\link{mmds}}, \code{\link{nerv_plugin}},
  \code{\link{nerv}}, \code{\link{plugin}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sammon_map}}, \code{\link{sjse_plugin}},
  \code{\link{sjse}}, \code{\link{smmds}},
  \code{\link{snerv_plugin}}, \code{\link{snerv}},
  \code{\link{ssne_plugin}}, \code{\link{ssne}},
  \code{\link{tasne_plugin}}, \code{\link{tasne}},
  \code{\link{tnerv_plugin}}, \code{\link{tnerv}},
  \code{\link{tsne_plugin}}, \code{\link{tsne}},
  \code{\link{uhsnerv_plugin}}, \code{\link{uhsnerv}},
  \code{\link{unerv_plugin}}, \code{\link{unerv}},
  \code{\link{usnerv_plugin}}

Other sneer probability embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsjse}}, \code{\link{hsnerv_plugin}},
  \code{\link{hsnerv}}, \code{\link{hssne_plugin}},
  \code{\link{hssne}}, \code{\link{jse_plugin}},
  \code{\link{jse}}, \code{\link{nerv_plugin}},
  \code{\link{nerv}}, \code{\link{plugin}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sjse_plugin}}, \code{\link{sjse}},
  \code{\link{snerv_plugin}}, \code{\link{snerv}},
  \code{\link{ssne_plugin}}, \code{\link{ssne}},
  \code{\link{tasne_plugin}}, \code{\link{tasne}},
  \code{\link{tnerv_plugin}}, \code{\link{tnerv}},
  \code{\link{tsne_plugin}}, \code{\link{tsne}},
  \code{\link{uhsnerv_plugin}}, \code{\link{uhsnerv}},
  \code{\link{unerv_plugin}}, \code{\link{unerv}},
  \code{\link{usnerv_plugin}}
}

