% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerv.R
\name{usnerv_stiffness}
\alias{usnerv_stiffness}
\title{USNeRV Stiffness Function}
\usage{
usnerv_stiffness(pm, qm, rev_kl, beta = 1, lambda = 0.5,
  eps = .Machine$double.eps)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}

\item{rev_kl}{"Reverse" KL divergence between \code{pm} and \code{qm}.}

\item{beta}{Bandwidth of the weighting function.}

\item{lambda}{NeRV weighting factor controlling the emphasis placed on
precision versus recall.}

\item{eps}{Small floating point value used to avoid numerical problems.}
}
\value{
Stiffness matrix
}
\description{
If using uniform bandwidths, the stiffness function for USNeRV is
simplified compared to the generic non-uniform case for SNeRV.
}

