% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{varfilter}
\alias{varfilter}
\title{Filter low variance columns.}
\usage{
varfilter(xm, minvar = 0)
}
\arguments{
\item{xm}{Matrix to filter.}

\item{minvar}{Minimum variance allowed for a column.}
}
\value{
Data with all columns with a variance lower than \code{minvar}
removed.
}
\description{
Removes columns from the data with variance lower than a threshold. Some
techniques can't handle zero variance columns.
}

