% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights}
\alias{weights}
\title{Create Weight Matrix from Output Data}
\usage{
weights(out, method)
}
\arguments{
\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
One or more weight matrix for the embedded coordinates in \code{out}.
}
\description{
Creates one or more matrix of positive weights from the distance matrix of the
embedded coordinates.
}
\details{
The number of matrices returned depends on whether multiscaling has been
chosen. If using multiscaling (see \code{\link{inp_multiscale}}), then a
list of weight matrices will be returned, one for each scale distance.
Otherwise, one matrix is returned.
}

