% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_multi}
\alias{weights_multi}
\title{Create Weight Matrices from Output Data}
\usage{
weights_multi(out, method)
}
\arguments{
\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
One or more weight matrix for the embedded coordinates in \code{out}.
}
\description{
Creates multi-scale weight matrices.
}
\details{
A list of weight matrices will be returned, one for each scale distance.
Otherwise, one matrix is returned.
}

