% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_single}
\alias{weights_single}
\title{Create Weight Matrix from Output Data}
\usage{
weights_single(out, method)
}
\arguments{
\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
Weight matrix for the embedded coordinates in \code{out}.
}
\description{
Creates a matrix of positive weights from the distance matrix of the
embedded coordinates.
}

