% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_to_pcond}
\alias{weights_to_pcond}
\title{Converts a weights matrix to a probability matrix.}
\usage{
weights_to_pcond(wm)
}
\arguments{
\item{wm}{a matrix of weighted distances.}
}
\value{
The probability matrix.
}
\description{
Used in SSNE and TSNE. The probability matrix is such that all elements are
positive and sum to 1.
}

