% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_to_probs_multi}
\alias{weights_to_probs_multi}
\title{Multiscale Weight Matrices to Probability Matrix Conversion}
\usage{
weights_to_probs_multi(wm, method)
}
\arguments{
\item{wm}{List of weight Matrices. Each matrix must have a "type" attribute
with one of the following values:
 \describe{
 \item{"symm"}{A symmetric matrix.}
 \item{"asymm"}{An asymmetric matrix.}
 }}

\item{method}{Embedding method.}
}
\value{
Probability matrix with a type suitable for the embedding
  \code{method}.
}
\description{
Given a list of weight matrices and an embedding method, this function
creates a multiscale probability matrix.
}

