% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_to_prow}
\alias{weights_to_prow}
\title{Converts a weight matrix to a row probability matrix.}
\usage{
weights_to_prow(wm)
}
\arguments{
\item{wm}{a matrix of weighted distances.}
}
\value{
a row probability matrix.
}
\description{
Used in ASNE. The probability matrix is such that all elements are positive
and each row sums to 1.
}

