% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{whiten}
\alias{whiten}
\title{Data whitening.
Whitens the data so that the covariance matrix is the identity matrix
(variances of the data are all one and the covariances are zero).}
\usage{
whiten(xm, scale = FALSE, zca = FALSE, ncomp = min(dim(xm)),
  epsilon = 1e-05, verbose = TRUE)
}
\arguments{
\item{xm}{Matrix to whiten.}

\item{scale}{If \code{TRUE}, then the centered columns of xm are divided by
their standard deviations.}

\item{zca}{If \code{TRUE}, then apply the Zero-Phase Component Analysis
(ZCA) whitening. If ZCA is used then the returned whitened data will have
as many columns as was in the input data, even if a value of \code{ncomp}
was specified.}

\item{ncomp}{Number of components to keep in a reduced dimension
representation of the data.}

\item{epsilon}{Regularization parameter to apply to eigenvalues, to avoid
numerical instabilities with eigenvalues close to zero. Values of 1.e-5 to
0.1 seem common.}

\item{verbose}{If \code{TRUE}, then debug messages will be logged.}
}
\value{
Whitened matrix.
}
\description{
Whitening consists of performing PCA on the centered (and optionally
scaled to unit standard deviation column) input data, optionally removing
eigenvectors corresponding to the smallest eigenvectors, and then rotating
the data to decorrelate it. Each component is additionally scaled by the
square root of the corresponding eigenvalue so that the variances of the
rotated solution are all equal to one.
}
\details{
Because whitening transformations are not unique, a further transformation
may be carried out on the PCA whitened solution to produce the Zero-Phase
Component Analysis (ZCA) whitening, which uses the rotation that best
reproduces the input data in a least-squares sense, while maintaining the
whitened properties. This may be useful for some image datasets where a
whitening which preserves the original data structure as much as possible
may be desirable.

If ZCA is used as the whitening transform then the returned whitened data
will have the same dimensionality as the input data xm, even if a value for
\code{ncomp} was also provided. This is simply due to the form of the
matrix multiplication that produces the ZCA transform. Additionally, ZCA
is most often applied to images where the whitened data is intended to be
visualized in the same way as the input data and so the dimensionality would
need to be identical anyway. Nonetheless, the dimensionality reduction has
still occurred, with the ZCA simply reconstructing the truncated data back
into the original (albeit whitened) space.
}

