% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{asne}
\alias{asne}
\title{Asymmetric Stochastic Neighbor Embedding (ASNE)}
\usage{
asne(beta = 1, eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{beta}{Precision parameter of the exponential similarity kernel
function. This is usually left at its default value of 1.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
The original SNE method, this uses exponential weighting for both the input
and output probabilities. Unlike \code{ssne} and \code{tsne},
the probabilities are defined with respect to points, not pairs: an element
Pij from the NxN probability matrix P should be thought of as a conditional
probabilty, pj|i, the probability that point j would be chosen as a close
neighbor of point i.

The probability matrix used in ASNE:

\itemize{
 \item{represents one N row-wise probability distributions, where N is the
 number of points in the data set, i.e. the row sums of the matrix are all
  one.}
 \item{is asymmetric, i.e. there is no requirement that
 \code{p[i, j] == p[j, i]}.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Row probability matrix based on embedded coordinates.}
}
}
\examples{
\dontrun{
embed_prob(method = asne(), ...)
}
}
\references{
Hinton, G. E., & Roweis, S. T. (2002).
Stochastic neighbor embedding.
In \emph{Advances in neural information processing systems} (pp. 833-840).
}
\seealso{
ASNE uses the \code{kl_cost} cost function and the
  \code{exp_weight} similarity function. The return value of this
  function should be used with the \code{embed_prob} embedding
  function.

Other sneer embedding methods: \code{\link{asne_plugin}},
  \code{\link{hsjse_plugin}}, \code{\link{hsnerv_plugin}},
  \code{\link{hssne_plugin}}, \code{\link{hssne}},
  \code{\link{jse_plugin}}, \code{\link{nerv_plugin}},
  \code{\link{plugin}}, \code{\link{rasne_plugin}},
  \code{\link{rasne}}, \code{\link{rssne_plugin}},
  \code{\link{rssne}}, \code{\link{rtsne_plugin}},
  \code{\link{rtsne}}, \code{\link{sjse_plugin}},
  \code{\link{snerv_plugin}}, \code{\link{ssne_plugin}},
  \code{\link{ssne}}, \code{\link{tasne_plugin}},
  \code{\link{tasne}}, \code{\link{tnerv_plugin}},
  \code{\link{tsne_plugin}}, \code{\link{tsne}},
  \code{\link{uhsnerv_plugin}}, \code{\link{unerv_plugin}},
  \code{\link{usnerv_plugin}}

Other sneer probability embedding methods: \code{\link{asne_plugin}},
  \code{\link{hsjse_plugin}}, \code{\link{hsnerv_plugin}},
  \code{\link{hssne_plugin}}, \code{\link{hssne}},
  \code{\link{jse_plugin}}, \code{\link{nerv_plugin}},
  \code{\link{plugin}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sjse_plugin}}, \code{\link{snerv_plugin}},
  \code{\link{ssne_plugin}}, \code{\link{ssne}},
  \code{\link{tasne_plugin}}, \code{\link{tasne}},
  \code{\link{tnerv_plugin}}, \code{\link{tsne_plugin}},
  \code{\link{tsne}}, \code{\link{uhsnerv_plugin}},
  \code{\link{unerv_plugin}}, \code{\link{usnerv_plugin}}
}

