% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{asne}
\alias{asne}
\title{Asymmetric Stochastic Neighbor Embedding.}
\usage{
asne()
}
\value{
a list with the following members:
\itemize{
 \item \code{cost_fn} Cost function for the embedding: the Kullback-Leibler
 divergence between the input and output probabilities.
 \item \code{weight_fn} Weight function for mapping squared distances in the
 embedded coordinates to weights, which are in turn converted to
 probabilities: ASNE uses the exponential function.
 \item \code{stiffness_fn} Stiffness function.
 \item \code{prob_out_fn} Function to convert weights matrix to a specific
 probability matrix, e.g. row-based or joint.
 \item \code{after_init_fn} Method-specific initialization function to
 invoke after input and output initialization callbacks.
}
}
\description{
Creates a list of functions that collectively implement ASNE.
}
\references{
G.E. Hinton and S.T. Roweis.
Stochastic Neighbor Embedding.
In Advances in Neural Information Processing Systems, volume 15,
pages 833-840, Cambridge, MA, USA, 2002. The MIT Press.
}

