% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{asne}
\alias{asne}
\title{Asymmetric Stochastic Neighbor Embedding}
\usage{
asne(mat_name = "ym", eps = .Machine$double.eps)
}
\arguments{
\item{mat_name}{Name of the matrix in the output data list that will contain
the embedded coordinates.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
a list containing:
 \item{\code{cost_fn}}{Cost function for the embedding: the Kullback-Leibler
 divergence between the input and output probabilities.}
 \item{\code{weight_fn}}{Weight function for mapping squared distances in the
 embedded coordinates to weights, which are in turn converted to
 probabilities: ASNE uses the exponential function.}
 \item{\code{stiffness_fn}}{Stiffness function.}
 \item{\code{update_out_fn}}{Function to calculate and store any needed
 data after a coordinate update.}
 \item{\code{after_init_fn}}{Method-specific initialization function to
 invoke after input and output initialization callbacks.}
 \item{\code{mat_name}}{Name of the matrix in the output data list that will
 contain the embedded coordinates.}
 \item{\code{eps}}{Small floating point value used to prevent numerical
 problems, e.g. in gradients and cost functions.}
}
\description{
Creates a list of functions that collectively implement ASNE.
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates. This name can be changed by
 specifying \code{mat_name}.}
 \item{\code{qm}}{Row probability matrix based on embedded coordinates.}
}
}
\references{
G.E. Hinton and S.T. Roweis.
Stochastic Neighbor Embedding.
In Advances in Neural Information Processing Systems, volume 15,
pages 833-840, Cambridge, MA, USA, 2002. The MIT Press.
}
\seealso{
Other sneer embedding methods: \code{\link{embedding_methods}},
  \code{\link{mmds}}, \code{\link{sammon_map}},
  \code{\link{smmds}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{probability_embedding_methods}},
  \code{\link{ssne}}, \code{\link{tasne}},
  \code{\link{tsne}}
}

