% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{asne}
\alias{asne}
\title{Asymmetric Stochastic Neighbor Embedding (ASNE)}
\usage{
asne(eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
The original SNE method, this uses exponential weighting for both the input
and output probabilities. Unlike \code{\link{ssne}} and \code{\link{tsne}},
the probabilities are defined with respect to points, not pairs: an element
Pij from the NxN probability matrix P should be thought of as a conditional
probabilty, pj|i, the probability that point j would be chosen as a close
neighbor of point i. The probability matrix used in ASNE:

\itemize{
 \item{represents one N row-wise probability distributions, where N is the
 number of points in the data set, i.e. the row sums of the matrix are all
  one.}
 \item{is asymmetric, i.e. there is no requirement that
 \code{p[i, j] == p[j, i]}.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Row probability matrix based on embedded coordinates.}
}
}
\examples{
\dontrun{
embed_prob(method = asne(), ...)
}
}
\references{
Hinton, G. E., & Roweis, S. T. (2002).
Stochastic neighbor embedding.
In \emph{Advances in neural information processing systems} (pp. 833-840).
}
\seealso{
ASNE uses the \code{\link{kl_cost}} cost function and the
  \code{\link{exp_weight}} similarity function. The return value of this
  function should be used with the \code{\link{embed_prob}} embedding
  function.

Other sneer embedding methods: \code{\link{embedding_methods}},
  \code{\link{hsnerv}}, \code{\link{hssne}},
  \code{\link{mmds}}, \code{\link{nerv}},
  \code{\link{sammon_map}}, \code{\link{smmds}},
  \code{\link{snerv}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{hsnerv}},
  \code{\link{hssne}}, \code{\link{nerv}},
  \code{\link{probability_embedding_methods}},
  \code{\link{snerv}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}
}

