% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{asne_stiffness}
\alias{asne_stiffness}
\title{ASNE Stiffness Function}
\usage{
asne_stiffness(pm, qm, beta = 1)
}
\arguments{
\item{pm}{Input probability matrix.}

\item{qm}{Output probabilty matrix.}

\item{beta}{The precision of the weighting function.}
}
\value{
Stiffness matrix.
}
\description{
ASNE Stiffness Function
}

