% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ball}
\alias{ball}
\title{Ball Data Set}
\usage{
ball(n = 1000, rad = 1, nlabels = 5)
}
\arguments{
\item{n}{Number of points to create.}

\item{rad}{Radius of the ball.}

\item{nlabels}{Number of labels.}
}
\value{
Data frame.
}
\description{
Simulation data.
}
\details{
Creates a series of points sampled from a 3D spherical volume.

Labels go from \code{1} to \code{nlabels}. Points are classified into
equally populated labels based on their distance from the origin.
}
\references{
A dataset like this was used in:
Lee, J. A., Peluffo-Ordo'nez, D. H., & Verleysen, M. (2015).
Multi-scale similarities in stochastic neighbour embedding: Reducing
dimensionality while preserving both local and global structure.
\emph{Neurocomputing}, \emph{169}, 246-261.
}

