% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{classical_gradient}
\alias{classical_gradient}
\title{Classical Gradient.}
\usage{
classical_gradient()
}
\value{
Classical gradient calculation method.
}
\description{
Configuration function for optimizer gradient calculation.
}

