% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{degree_centrality}
\alias{degree_centrality}
\title{Degree Centrality}
\usage{
degree_centrality(m)
}
\arguments{
\item{m}{Nearest neighbor graph matrix.}
}
\value{
Degree centrality of \code{m}.
}
\description{
A node centrality measure for graphs.
}
\details{
Sums the (possibly weighted) edges that are incident to each node. For
probability matrices, row and column i is interpreted as containing the
weighted edges that are connected to and from node i.
}

