% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embedding_methods}
\alias{embedding_methods}
\title{Embedding Methods}
\description{
Links to all available embedding methods.
}
\examples{
\dontrun{
embed_prob(method = tsne())
embed_dist(method = mmds())
}
}
\seealso{
Other sneer embedding methods: \code{\link{asne}},
  \code{\link{mmds}}, \code{\link{sammon_map}},
  \code{\link{smmds}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tsne}}
}
\keyword{internal}

