% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{exp_weight}
\alias{exp_weight}
\title{Exponential Weighted Similarity}
\usage{
exp_weight(d2m, beta = 1)
}
\arguments{
\item{d2m}{Matrix of squared distances.}

\item{beta}{exponential parameter.}
}
\value{
Weight matrix.
}
\description{
A similarity function for probability-based embedding.
}
\details{
The weight matrix, \eqn{W} is generated from the squared distances matrix,
\eqn{D^2}{D2} by:

\deqn{W = e^{-\beta D^2}}{exp(-beta * D2)}
}
\seealso{
Other sneer weight functions: \code{\link{heavy_tail_weight}},
  \code{\link{sqrt_exp_weight}}, \code{\link{tdist_weight}}
}

