% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{exp_weight}
\alias{exp_weight}
\title{Exponential weight.}
\usage{
exp_weight(d2m, beta = 1)
}
\arguments{
\item{d2m}{Matrix of squared distances.}

\item{beta}{xponential parameter.}
}
\value{
Weight matrix.
}
\description{
Applies exponential weighting to the squared distances. Can also be thought
of as gaussian weighting of the distances.
}
\details{
\eqn{W = e^{-\beta D^2}}
}

