% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{hssne}
\alias{hssne}
\title{Heavy-Tailed Symmetric Stochastic Neighbor Embedding (HSSNE)}
\usage{
hssne(beta = 1, alpha = 0, eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{beta}{The precision of the kernel similarity function. Becomes
equivalent to the precision in the Gaussian distribution of distances as
\code{alpha} approaches zero. Leave at the default value of 1 if you want
consistency with SSNE and ASNE.}

\item{alpha}{Tail heaviness of the kernel similarity function. Must be
greater than zero. When set to a small value this method is equivalent to
SSNE. When set to one to one, this method behaves like t-SNE.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
HSSNE is a generalization of \code{ssne} and \code{tsne},
which uses the \code{heavy_tail_weight} similarity function to
generate its probabilities.

The heavy tailedness of the weighting function is controlled by the parameter
\eqn{\alpha}{alpha}, As \eqn{\alpha \to 0}{alpha approaches 0}, the weighting
function becomes exponential (like SSNE). At \eqn{\alpha = 1}{alpha = 1},
the weighting function is the t-distribution with one degree of freedom
(t-SNE). Increasing \eqn{\alpha}{alpha} further will take you into the realm
of functions with even heavier tails.

Additionally, HSSNE allows control over \eqn{\beta}{beta}, the degree of
precision (inverse of the spread) of the function. Normally, this set to one
for the output distances in t-SNE and related methods.

The probability matrix used in HSSNE:

\itemize{
 \item{represents one probability distribution, i.e. the grand sum of the
 matrix is one.}
 \item{is symmetric, i.e. \code{P[i, j] == P[j, i]} and therefore the
 probabilities are joint probabilities.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix based on the weight matrix
 \code{wm}.}
 \item{\code{wm}}{Weight matrix generated from the distances between points
 in \code{ym}.}
}
}
\examples{
\dontrun{
# behave like SSNE
embed_prob(method = hssne(alpha = 0), ...)

# behave like t-SNE
embed_prob(method = hssne(alpha = 1), ...)
}
}
\references{
Yang, Z., King, I., Xu, Z., & Oja, E. (2009).
Heavy-tailed symmetric stochastic neighbor embedding.
In \emph{Advances in neural information processing systems} (pp. 2169-2177).
}
\seealso{
HSSNE uses the \code{heavy_tail_weight} similarity function.
The return value of this function should be used with the
\code{embed_prob} embedding function.

Other sneer embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsnerv_plugin}}, \code{\link{hssne_plugin}},
  \code{\link{jse_plugin}}, \code{\link{nerv_plugin}},
  \code{\link{plugin}}, \code{\link{rasne_plugin}},
  \code{\link{rasne}}, \code{\link{rssne_plugin}},
  \code{\link{rssne}}, \code{\link{rtsne_plugin}},
  \code{\link{rtsne}}, \code{\link{sjse_plugin}},
  \code{\link{snerv_plugin}}, \code{\link{ssne_plugin}},
  \code{\link{ssne}}, \code{\link{tasne_plugin}},
  \code{\link{tasne}}, \code{\link{tnerv_plugin}},
  \code{\link{tsne_plugin}}, \code{\link{tsne}},
  \code{\link{uhsnerv_plugin}}, \code{\link{unerv_plugin}},
  \code{\link{usnerv_plugin}}

Other sneer probability embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsnerv_plugin}}, \code{\link{hssne_plugin}},
  \code{\link{jse_plugin}}, \code{\link{nerv_plugin}},
  \code{\link{plugin}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sjse_plugin}}, \code{\link{snerv_plugin}},
  \code{\link{ssne_plugin}}, \code{\link{ssne}},
  \code{\link{tasne_plugin}}, \code{\link{tasne}},
  \code{\link{tnerv_plugin}}, \code{\link{tsne_plugin}},
  \code{\link{tsne}}, \code{\link{uhsnerv_plugin}},
  \code{\link{unerv_plugin}}, \code{\link{usnerv_plugin}}
}

