% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{improve_guess}
\alias{improve_guess}
\title{Narrows bisection bracket range.}
\usage{
improve_guess(bracket, sgn)
}
\arguments{
\item{bracket}{List representing the bounds of the parameter search.
Contains three elements: \code{lower}: lower value, \code{upper}: upper
value, \code{mid}: the midpoint.}

\item{sgn}{Sign of the value of the function evaluted at \code{bracket$mid}.}
}
\value{
Updated \code{bracket} list.
}
\description{
Narrows bisection bracket range.
}

