% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{inp_from_perp}
\alias{inp_from_perp}
\title{Input Probability Initialization By Bisection Search on Perplexity}
\usage{
inp_from_perp(perplexity = 30, input_weight_fn = exp_weight,
  keep_all_results = FALSE, verbose = TRUE)
}
\arguments{
\item{perplexity}{Target perplexity value for the probability distributions.}

\item{input_weight_fn}{Weighting function for distances. It should have the
signature \code{input_weight_fn(d2m, beta)}, where \code{d2m} is a matrix
of squared distances and \code{beta} is a real-valued scalar parameter
which will be varied as part of the search to produce the desired
\code{perplexity}. The function should return a matrix of weights
corresponding to the transformed squared distances passed in as arguments.}

\item{keep_all_results}{If \code{true} then the list returned by the callback
will also contain a vector of \code{beta} parameters that generated the
probability matrix. Otherwise, only the probability matrix is returned.}

\item{verbose}{If \code{TRUE} display messages about progress of
initialization.}
}
\value{
Input initialization method for use by \code{\link{make_init_inp}}.
  Although direct use of this value is intended only by the embedding
  function, the resulting data can be accessed by passing \code{"inp"} to
  the \code{export} list in an embedding function such as
  \code{\link{embed_prob}} or \code{\link{embed_dist}}. If exported, the
  \code{inp} list will contain the follow data created by this method:
  \item{\code{pm}}{Input probabilities.}
  \item{\code{beta}}{Input weighting parameters that produced the
    probabilities. Only provided if \code{keep_all_results} is \code{TRUE}.}
}
\description{
An initialization method for creating input probabilities.
}
\details{
Function to generate a row probability matrix by optimizing a one-parameter
weighting function. This is used to create input probabilities from the
input distances, such that each row of the matrix is a probability
distribution with the specified perplexity.

This is the method described in the original SNE paper, and few methods
deviate very strongly from it, although they may do further processing
on the resulting probability matrix. For example, SSNE and t-SNE convert
this matrix into a single joint probability distribution.
}
\examples{
# Set target perplexity of each probability distribution to 30
inp_from_perp(perplexity = 30)

# Set target perplexity of each probability distribution to 30 but use
# a different weighting function.
inp_from_perp(perplexity = 30, input_weight_fn = sqrt_exp_weight)

# Perplexity of 50, and keep the values of the exponential parameter for
# later processing or reporting.
inp_from_perp(perplexity = 50, keep_all_results = TRUE)

# Should be passed to the init_inp argument of an embedding function:
init_inp = make_init_inp(inp_from_perp(perplexity = 30,
                                         input_weight_fn = exp_weight))
}
\seealso{
\code{\link{make_init_inp}} for how to use this function as part
of an embedding.

Other sneer input initializers: \code{\link{input_initializers}}
}

