% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{inp_prob}
\alias{inp_prob}
\title{Wrap Input Probability Initializer}
\usage{
inp_prob(input_initializer)
}
\arguments{
\item{input_initializer}{Input initializer which creates a probabilty.}
}
\value{
Wrapped initializer with the correct signature for use by an
 embedding function.
}
\description{
This function takes an input initializer which creates a probability, and
wraps it so that the correct handling of the probability matrix can be
carried out when it is invoked.
}
\seealso{
\code{\link{probability_matrices}} describe the type of probability
  matrix used by sneer.
}

