% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{input_initializers}
\alias{input_initializers}
\title{Input Initializers}
\description{
These methods deal with converting the input coordinates or distances
into the structures required by different embedding methods. For instance,
probability-based embedding methods require the construction of a probability
matrix from the input distances.
}
\examples{

\dontrun{
# initializer that uses bisection search to create input probability
# distribution with a perplexity of 50
# pass result to init_inp parameter of an embedding function
embed_prob(method = tsne(), init_inp = inp_from_perp(perplexity = 50), ...)
}
}
\seealso{
Input initializers should be passed to the
\code{init_inp} parameter of embedding functions such as
\code{\link{embed_prob}} or \code{\link{embed_dist}}.

Other sneer input initializers: \code{\link{inp_from_perp}},
  \code{\link{inp_multiscale}}, \code{\link{inp_step_perp}}
}
\keyword{internal}

