% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{kl_cost}
\alias{kl_cost}
\title{Calculate the Kullback Leibler divergence from input and output data.}
\usage{
kl_cost(inp, out)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}
}
\value{
the KL divergence between \code{inp$pm} and \code{out$qm}.
}
\description{
Calculate the Kullback Leibler divergence from input and output data.
}

