% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{kl_cost}
\alias{kl_cost}
\title{Kullback Leibler Divergence Cost Function}
\usage{
kl_cost(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
KL divergence between \code{inp$pm} and \code{out$qm}.
}
\description{
A measure of embedding quality between input and output data.
}
\details{
This cost function requires the following matrices to be defined:
\describe{
 \item{\code{inp$pm}}{Input probabilities.}
 \item{\code{out$qm}}{Output probabilities.}
}
}
\seealso{
Other sneer cost functions: \code{\link{kruskal_stress_cost}},
  \code{\link{mean_relative_error_cost}},
  \code{\link{metric_sstress_cost}},
  \code{\link{metric_stress_cost}},
  \code{\link{normalized_stress_cost}},
  \code{\link{rms_metric_stress_cost}},
  \code{\link{sammon_stress_cost}}
}

