% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricks.R
\name{late_momentum}
\alias{late_momentum}
\title{Early Exaggeration}
\usage{
late_momentum(momentum = 0.9, on_iter = 900, verbose = TRUE)
}
\arguments{
\item{momentum}{Value of the fixed momentum to use.}

\item{on_iter}{Iteration step at which to apply the new momentum value.}

\item{verbose}{If \code{TRUE} report a message when the new momentum is
applied.}
}
\value{
Trick callback. Should be passed to \code{\link{make_tricks}} when
configuring an embedding.
}
\description{
A "trick" for improving the quality of the embedding.
}
\details{
Replaces the momentum scheme of an optimizer with a fixed value. This may
be useful when using an optimizer with a very aggressive momentum schedule,
such as Nesterov Accelerated Gradient method. Reducing the momentum for the
final stage of the embedding may help the optimizer refine a solution.
}
\examples{
\dontrun{
# Apply new momentum after 800 iterations
embed_prob(make_tricks(late_exaggeration(momentum = 0.5, on_iter = 800)),
           ...)
}
}
\references{
Sutskever, I., Martens, J., Dahl, G. and Hinton, G. E.
On the importance of momentum and initialization in deep learning.
30th International Conference on Machine Learning, Atlanta, USA, 2013.
JMLR: W&CP volume 28.
}
\seealso{
Other sneer tricks: \code{\link{early_exaggeration}},
  \code{\link{tricks}}
}

