% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{linear_momentum}
\alias{linear_momentum}
\title{Linear momentum.}
\usage{
linear_momentum(max_iter, init_momentum = 0, final_momentum = 0.9)
}
\arguments{
\item{max_iter}{Number of iterations to scale the momentum over from
\code{init_momentum} to \code{final_momentum}.}

\item{init_momentum}{Momentum value for the first \code{switch_iter}
iterations.}

\item{final_momentum}{Momentum value after \code{switch_iter} iterations.}
}
\value{
A solution update method for use by the optimizer. A list consisting
of:
\itemize{
 \item \code{init_momentum} Initial momentum.
 \item \code{final_momentum} Final momentum.
 \item \code{init} Function to do any needed initialization.
 \item \code{get_update} Function to return the current update, which will
 be added to current solution matrix.
 \item \code{after_step} Function to do any needed updating or internal state
 before the next optimization step.
}
}
\description{
Create an callback for the optimizer to use to update the embedding solution.
Update is in the form of a linear momentum function.
}

