% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_embedding_qplot}
\alias{make_embedding_qplot}
\title{Embedded Coordinates Plot Using \code{ggplot2} Library}
\usage{
make_embedding_qplot(x, attr_name)
}
\arguments{
\item{x}{Data frame containing label information for the embedded data.}

\item{attr_name}{Name of the label column in \code{x}.}
}
\value{
Function with signature \code{plot_fn(ym)} where \code{ym} is a
2D matrix of embedded coordinates of data set \code{x}. On invocation, the
data will be plotted.
}
\description{
Creates a ggplot2 function for embeddings of a given dataset.
}
\note{
Use of this function requires that the \code{ggplot2} package be
installed.
}
\examples{
\dontrun{
# make two different embeddings of the iris dataset
prcomp_iris <- prcomp(iris[, 1:4], center = TRUE, retx = TRUE)
mds_iris <- embed_dist(iris[, 1:4], method = mmds(eps = 1e-4),
                       opt = bold_nag_opt(), init_out = make_init_out(
                         init_config = prcomp_iris$x[, 1:2]),
                       max_iter = 40)
iris_view <- make_embedding_qplot(iris, "Species")
iris_view(prcomp_iris$x)
iris_view(mds_iris$ym)
}
}

