% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{make_init_inp}
\alias{make_init_inp}
\title{Input Initialization}
\usage{
make_init_inp(...)
}
\arguments{
\item{...}{Zero or more \code{\link{input_initializers}}.}
}
\value{
An input initializer to be used by an embedding routine. The input
  data is not of direct interest in most embeddings, but may be useful for
  diagnostic purposes. If exported from an embedding, it will be as a list
  called \code{inp} on the embedding result, and itself contains:
  \item{\code{xm}}{Input coordinates if these were provided.}
  \item{\code{dm}}{Input distances.}
  If other initializers were provided as arguments to this method, they may
  provide extra data - see the documentation of their functions for details.
}
\description{
Factory function for input initialization.
}
\details{
This function is responsible for creating a method to initialize the input
data for embedding. Any
\code{\link{input_initializers}} are passed as parameters to this function
are invoked in order they were passed.
}
\examples{
# Should be passed to the init_inp argument of an embedding function:
\dontrun{
 # distance-based embedding don't need extra input initialization
 embed_dist(init_inp = make_init_inp(), method = mds())

 # probability-based embeddings must also initialize input probabilities
 embed_prob(init_inp = make_init_inp(prob_perp_bisect()), method = tsne())
}
}
\seealso{
\code{\link{embed_dist}} and \code{\link{embed_prob}}
  for how to use this function to configure an embedding and to export input
  data. Also see the documentation for the specific members of the
  \code{\link{input_initializers}} that can be passed to this function.
}

