% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{make_objective_fn}
\alias{make_objective_fn}
\title{Create callback that can be used as an objective function for perplexity
parameter search.}
\usage{
make_objective_fn(d2r, i, weight_fn, perplexity, h_base = exp(1))
}
\arguments{
\item{d2r}{Row of a squared distance matrix.}

\item{i}{Index of the row in the matrix.}

\item{weight_fn}{Function that converts squared distances to weights. Should
have the signature \code{function(D2, beta)}}

\item{perplexity}{the target perplexity the probability generated by the
weight function should produce.}

\item{h_base}{the base of the logarithm to use for Shannon Entropy
calculations.}
}
\value{
Callback function for use in parameter search routine.
}
\description{
The callback has the signature \code{fn(beta)} where \code{beta} is the
parameter being optimized, and returns a list
containing: \itemize{
 \item \code{value} the difference between the Shannon Entropy for the value
  of beta passed as an argument and the target Shannon Entropy
 \item \code{h} Shannon Entropy for the value of beta passed as an
argument.
 \item \code{pm} the probabilities generated by the weighting function.
}
}

