% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{make_preprocess}
\alias{make_preprocess}
\title{Create preprocessing callback}
\usage{
make_preprocess(range_scale_matrix = FALSE, range_scale = FALSE, rmin = 0,
  rmax = 1, auto_scale = FALSE, whiten = FALSE, zwhiten = FALSE,
  whiten_dims = 30, preprocess_fn = NULL, verbose = TRUE)
}
\arguments{
\item{range_scale_matrix}{If \code{TRUE}, input coordinates will be scaled
so that elements are between (\code{rmin}, \code{rmax}).}

\item{range_scale}{If \code{TRUE}, input coordinates will be scaled}

\item{rmin}{Minimum value if using \code{range_scale_matrix} and
\code{range_scale}.}

\item{rmax}{Maximum value if using \code{range_scale_matrix} and
\code{range_scale}.}

\item{auto_scale}{If \code{TRUE}, input coordinates will be centered and
scaled so that each column has a mean of 0 and a standard deviation of 1.}

\item{whiten}{If \code{TRUE}, input coordinates will be whitened.}

\item{zwhiten}{If \code{TRUE}, input coordinates will be whitened using the
ZCA transform.}

\item{whiten_dims}{Number of components to use in the whitening preprocess.
Only use if \code{whiten} or \code{zwhiten} is \code{TRUE}.}

\item{preprocess_fn}{Custom preprocessing function to be applied. Should have
the signature \code{xm} where \code{xm} is the input coordinate matrix, and
returns the processed coordinate matrix.}

\item{verbose}{If \code{TRUE}, log information about preprocessing.}
}
\value{
Function with signature \code{fn(xm)} where \code{xm} is the input
coordinate matrix. Function returns the coordinate matrix after applying
the specified preprocessing.
}
\description{
Creates a callback for use in the embedding routine. Preprocessing will be
applied to the input coordinates (if provided) before embedding-specific
input initialization occurs (e.g. calculation of distances and
probabilities).
}
\details{
Pass the result of this factory function to the preprocess argument of an
embedding function, e.g. \code{embed_sim}. A variety of common preprocessing
options are available. The range scaling and auto scaling options are
mutually exclusive, but can be combined with whitening if required. No
preprocessing will be carried out if the input data is a distance matrix.
}
\examples{
# Scale the input data so the smallest element is 0, and the largest is 1.
make_preprocess(range_scale_matrix = TRUE)

# Scale the input data so the smallest element in each column is -1, and
# the largest is 1.
make_preprocess(range_scale = TRUE, rmin = -1)

# Autoscale each column in the input data, to mean 0 and the sd 1.
make_preprocess(auto_scale = TRUE)

# Whiten the data after range scaling the matrix.
make_preprocess(range_scale_matrix = TRUE, whiten = TRUE)

# Whiten the data using 10 components.
make_preprocess(range_scale_matrix = TRUE, whiten = TRUE, whiten_dims = 10)

# Whiten the data with the ZCA technique, using 10 components.
make_preprocess(range_scale_matrix = TRUE, zwhiten = TRUE, whiten_dims = 10)

# Should be passed to the preprocess argument of an embedding function:
\dontrun{
 embed_sim(preprocess = make_preprocess(range_scale = TRUE, rmin = -1), ...)
}
}
\seealso{
\code{\link{embed_sim}} for how to use this function for configuring
an embedding.
}

