% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricks.R
\name{make_tricks}
\alias{make_tricks}
\title{Tricks}
\usage{
make_tricks(...)
}
\arguments{
\item{...}{Zero or more \code{\link{tricks}}.}
}
\value{
Callback collecting all the supplied tricks, to be invoked by the
embedding routine.
}
\description{
A collection of heuristics to improve embeddings.
}
\details{
Creates a callback which the embedding routine will call before each
optimization step. Represents a collection of miscellaneous tweaks that
don't fit neatly into the rest of the framework, and as such can modify
the state of any other component of the embedding routine: input data,
output data, optimizer or embedding method.

The input to this function can be zero, one or multiple independent tricks,
each of which is created by their own factory function. As a result, the
signature of this function is not informative. For information on the tricks
available, see \code{\link{tricks}}.

Some wrappers around this function which apply sets of tricks from the
literature are available. See the links under the 'See Also' section.
}
\examples{
# Use early exaggeration as described in the t-SNE paper
make_tricks(early_exaggeration(exaggeration = 4, off_iter = 50))

# Should be passed to the tricks argument of an embedding function:
\dontrun{
 embed_prob(tricks = make_tricks(early_exaggeration(
                                 exaggeration = 4, off_iter = 50), ...)
}
}
\seealso{
The result value of this function should be passed to the
\code{\link{tricks}} parameter of embedding routines like
\code{\link{embed_prob}} and \code{\link{embed_dist}}.

Other sneer trick collections: \code{\link{tsne_tricks}}
}

