% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{momentum_scheme}
\alias{momentum_scheme}
\title{Momentum Update Method}
\usage{
momentum_scheme(mu_fn = NULL, calculate = NULL, validate = NULL,
  after_step = NULL, linear_weight = FALSE, init_momentum = 0,
  min_momentum = 0, max_momentum = 1, momentum = list(), msg_fn = NULL,
  verbose = TRUE)
}
\arguments{
\item{mu_fn}{Momentum function with signature \code{mu_fn(iter)} where
\code{iter} is the iteration number and the function returns the momentum
value for that iteration. Ignored if \code{calculate} is specified.}

\item{calculate}{Momentum function with signature
\code{calculate(opt, inp, out, method, iter)} where and the function
returns the momentum value for that iteration.}

\item{validate}{Validation function.}

\item{after_step}{After-step function.}

\item{linear_weight}{If \code{TRUE}, then the gradient descent contribution
to the update will be weighted by \code{1 - momentum}.}

\item{init_momentum}{Initial momentum value.}

\item{min_momentum}{Minimum momentum value.}

\item{max_momentum}{Maximum momentum value.}

\item{momentum}{Momentum method list containing any specific values required
by a particular scheme.}

\item{msg_fn}{Function with signature \code{msg_fn(iter)} where
\code{iter} is the iteration number. Called on each iteration if
\code{verbose} is \code{TRUE}. Can be used to provide information.}

\item{verbose}{If \code{TRUE}, then \code{msg_fn} will be invoked on each
iteration.}
}
\value{
Momentum scheme update method.
}
\description{
Create an update method using a user-defined momentum scheme function.
}

