% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{momentum_update}
\alias{momentum_update}
\title{Update Solution With Momentum}
\usage{
momentum_update(opt, inp, out, method, iter, linear_weight = FALSE)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}

\item{iter}{Iteration number.}

\item{linear_weight}{If \code{TRUE}, then the gradient descent contribution
to the update will be weighted by \code{1 - momentum}.}
}
\value{
Update matrix, consisting of gradient update and momentum term.
}
\description{
Carries out a solution update using a momentum term in addition to the
gradient update.
}

