% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{nesterov_grad_pos}
\alias{nesterov_grad_pos}
\title{Nesterov Accelerated Gradient calculation.}
\usage{
nesterov_grad_pos(opt, inp, out, method)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\value{
List containing:
\itemize{
 \item \code{km} Stiffness matrix.
 \item \code{gm} Gradient matrix.
}
}
\description{
This function calculates the gradient at a position determined by applying
the momentum update to the current solution position.
}
\references{
Sutskever, I., Martens, J., Dahl, G. and Hinton, G. E.
On the importance of momentum and initialization in deep learning.
30th International Conference on Machine Learning, Atlanta, USA, 2013.
JMLR: W&CP volume 28.
}

