% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{nesterov_gradient}
\alias{nesterov_gradient}
\title{Nesterov Accelerated Gradient}
\usage{
nesterov_gradient()
}
\value{
NAG method for gradient calculation.
}
\description{
Factory function for creating an optimizer gradient method.
}
\details{
Calculates the gradient according to the Nesterov Accelerated Gradient
method. Can also provide the position at which the gradient is calculated,
useful for some step length calculation methods.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# gradient method:
make_opt(gradient = nesterov_gradient())
}
\references{
Sutskever, I., Martens, J., Dahl, G., & Hinton, G. (2013).
On the importance of initialization and momentum in deep learning.
In \emph{Proceedings of the 30th international conference on machine learning (ICML-13)}
(pp. 1139-1147).
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{\link{optimization_gradient_interface}}
for details on the functions and values defined for this method.

Other sneer optimization gradient methods: \code{\link{classical_gradient}},
  \code{\link{optimization_gradient}}
}

