% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{nesterov_gradient}
\alias{nesterov_gradient}
\title{Nesterov Accelerated Gradient.}
\usage{
nesterov_gradient()
}
\value{
NAG calculation method.
}
\description{
Configuration function for optimizer gradient calculation.
}
\references{
Sutskever, I., Martens, J., Dahl, G. and Hinton, G. E.
On the importance of momentum and initialization in deep learning.
30th International Conference on Machine Learning, Atlanta, USA, 2013.
JMLR: W&CP volume 28.
}

