% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{nesterov_nsc_momentum}
\alias{nesterov_nsc_momentum}
\title{Momentum schedule for non-strongly convex functions.}
\usage{
nesterov_nsc_momentum(max_momentum = 1)
}
\arguments{
\item{max_momentum}{Maximum value the momentum may take.}
}
\value{
Nesterov momentum update method, to be used by the optimizer.
}
\description{
Factory function for creating an optimizer update method.
}
\details{
Create a callback for the optimizer to use to update the embedding solution,
using momentum scheme suggested by Nesterov.

Update is in the form of a momentum schedule of the form
\deqn{\mu_{t} = 1-\frac{3}{t+5}}{mu_t = 1-[3/(t+5)]}
where \eqn{t} is the iteration number.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# update method:
make_opt(update = nesterov_nsc_momentum())
}
\references{
Sutskever, I., Martens, J., Dahl, G. and Hinton, G. E.
On the importance of momentum and initialization in deep learning.
30th International Conference on Machine Learning, Atlanta, USA, 2013.
JMLR: W&CP volume 28.
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{\link{optimizer_update}} for details
on the functions and values defined for this method.

Other sneer optimization update methods: \code{\link{linear_momentum}},
  \code{\link{no_momentum}}, \code{\link{step_momentum}}
}

