% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{no_momentum}
\alias{no_momentum}
\title{No Momentum Update}
\usage{
no_momentum(...)
}
\arguments{
\item{...}{Base momentum parameters to pass to the
\code{momentum_scheme} factory function.}
}
\value{
Zero-momentum update method, to be used by the optimizer.
}
\description{
Factory function for creating an optimizer update method.
}
\details{
Create an callback for the optimizer to use that uses strict gradient
descent with no momentum term.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# update method:
make_opt(update = no_momentum())
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{optimization_update_interface}
for details on the functions and values defined for this method.

Other sneer optimization update methods: \code{\link{constant_momentum}},
  \code{\link{linear_momentum}},
  \code{\link{nesterov_nsc_momentum}},
  \code{\link{optimization_update}},
  \code{\link{step_momentum}}
}

