% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normalize}
\alias{normalize}
\title{Scale a Vector (or Matrix) to Length 1}
\usage{
normalize(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
\code{x} with elements scaled such that its length equals 1.
}
\description{
Scale a Vector (or Matrix) to Length 1
}

