% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_step.R
\name{optimization_step_size}
\alias{optimization_step_size}
\title{Optimization Step Size Methods}
\description{
The available step size methods that can be used by the optimization routines
in sneer.
}
\examples{
make_opt(step_size = bold_driver())
make_opt(step_size = jacobs())
}
\seealso{
Other sneer optimization step size methods: \code{\link{backtracking}},
  \code{\link{bold_driver}},
  \code{\link{constant_step_size}}, \code{\link{jacobs}},
  \code{\link{tsne_jacobs}}
}
\keyword{internal}

