% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_grad.R
\name{optimizer_gradient}
\alias{optimizer_gradient}
\title{Optimizer gradient methods.}
\description{
Part of the optimizer that calculates the gradient at a position in the
solution space.
}
\section{Interface}{

A gradient method is a function with the signature
\code{grad_fn(opt, inp, out, method)} where the arguments are:
 \describe{
   \item{\code{opt}}{Optimizer.}
   \item{\code{inp}}{Input data.}
   \item{\code{out}}{Output data.}
   \item{\code{method}}{Embedding method.}
 }
The function should return a list containing:
\describe{
 \item{\code{gm}}{Gradient matrix.}
}
}
\keyword{internal}

