% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{pcond_to_pjoint}
\alias{pcond_to_pjoint}
\title{Convert a conditional probability matrix to a joint probability matrix.}
\usage{
pcond_to_pjoint(pm)
}
\arguments{
\item{pm}{a conditional probability matrix or row stochastic matrix.}
}
\value{
a joint probability matrix, such that the elements sum to 1 and
\code{pm[i, j]} = \code{pm[j, i]}.
}
\description{
Used in SSNE and TSNE to convert the input conditional probabilties into a
joint probability matrix.
}

