% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{probability_embedding_methods}
\alias{probability_embedding_methods}
\title{Probability-Based Embeddings}
\description{
The available embedding methods which work by reproducing probabilities
based on the input distances, rather than the input distances directly.
}
\examples{
\dontrun{
embed_prob(method = tsne(), ...)
embed_prob(method = asne(), ...)
}
}
\seealso{
Other sneer probability embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsnerv_plugin}}, \code{\link{hssne_plugin}},
  \code{\link{hssne}}, \code{\link{jse_plugin}},
  \code{\link{nerv_plugin}}, \code{\link{plugin}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sjse_plugin}}, \code{\link{snerv_plugin}},
  \code{\link{ssne_plugin}}, \code{\link{ssne}},
  \code{\link{tasne_plugin}}, \code{\link{tasne}},
  \code{\link{tnerv_plugin}}, \code{\link{tsne_plugin}},
  \code{\link{tsne}}, \code{\link{uhsnerv_plugin}},
  \code{\link{unerv_plugin}}, \code{\link{usnerv_plugin}}
}
\keyword{internal}

