% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{probability_embedding_methods}
\alias{probability_embedding_methods}
\title{Probability-Based Embeddings}
\description{
The available embedding methods which work by reproducing probabilities
based on the input distances, rather than the input distances directly.
}
\examples{
\dontrun{
embed_prob(method = tsne(), ...)
embed_prob(method = asne(), ...)
}
}
\seealso{
Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hsnerv}}, \code{\link{hssne}},
  \code{\link{nerv}}, \code{\link{rasne}},
  \code{\link{rssne}}, \code{\link{rtsne}},
  \code{\link{snerv}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}
}
\keyword{internal}

