% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{prow_to_pjoint}
\alias{prow_to_pjoint}
\title{Convert a row probability matrix to a join probability matrix.}
\usage{
prow_to_pjoint(prow)
}
\arguments{
\item{prow}{Row probability matrix.}
}
\value{
Joint probability matrix.
}
\description{
Given a row probability matrix (elements of each row are non-negative and
sum to one), this function scales each element by such that the elements of
the entire matrix sum to one, and that the matrix is symmetric, i.e.
\code{p[i, j] = p[j, i]}.
}

