% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{rasne}
\alias{rasne}
\title{"Reverse" Asymmetric Stochastic Neighbor Embedding (RASNE)}
\usage{
rasne(eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
Like \code{\link{asne}}, but with the cost function using the "reverse" form
of the Kullback-Leibler divergence, i.e. KL(Q||P).

The probability matrix in RASNE:

\itemize{
 \item{represents one N row-wise probability distributions, where N is the
 number of points in the data set, i.e. the row sums of the matrix are all
  one.}
 \item{is asymmetric, i.e. there is no requirement that
 \code{p[i, j] == p[j, i]}.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Row probability matrix based on embedded coordinates.}
}
}
\examples{
\dontrun{
embed_prob(method = asne(), ...)
}
}
\seealso{
RASNE uses the \code{\link{kl_cost}} cost function and the
  \code{\link{exp_weight}} similarity function. The return value of this
  function should be used with the \code{\link{embed_prob}} embedding
  function.

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{hsjse}},
  \code{\link{hsnerv}}, \code{\link{hssne}},
  \code{\link{jse}}, \code{\link{mmds}},
  \code{\link{nerv}}, \code{\link{rssne}},
  \code{\link{rtsne}}, \code{\link{sammon_map}},
  \code{\link{smmds}}, \code{\link{snerv}},
  \code{\link{ssne}}, \code{\link{tasne}},
  \code{\link{tnerv}}, \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hsjse}}, \code{\link{hsnerv}},
  \code{\link{hssne}}, \code{\link{jse}},
  \code{\link{nerv}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rssne}}, \code{\link{rtsne}},
  \code{\link{snerv}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}
}

