% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{rssne}
\alias{rssne}
\title{"Reverse" Symmetric Stochastic Neighbor Embedding (RSSNE)}
\usage{
rssne(beta = 1, eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{beta}{Precision parameter of the exponential similarity kernel
function.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
Like \code{ssne}, but with the cost function using the "reverse" form
of the Kullback-Leibler divergence, i.e. KL(Q||P).

The probability matrix in RSSNE:

\itemize{
 \item{represents one probability distribution, i.e. the grand sum of the
 matrix is one.}
 \item{is symmetric, i.e. \code{P[i, j] == P[j, i]} and therefore the
 probabilities are joint probabilities.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix based on embedded coordinates.}
}
}
\examples{
\dontrun{
embed_prob(method = rssne(), ...)
}
}
\seealso{
SSNE uses the \code{reverse_kl_cost} cost function and the
  \code{exp_weight} similarity function. The return value of this
  function should be used with the \code{embed_prob} embedding
  function.

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{hssne}}, \code{\link{rasne}},
  \code{\link{rtsne}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hssne}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne}}, \code{\link{rtsne}},
  \code{\link{ssne}}, \code{\link{tasne}},
  \code{\link{tsne}}
}

