% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{scale_prec_to_perp}
\alias{scale_prec_to_perp}
\title{Factory function which creates an output weight function for a given input
perplexity, in the context of multiscale embedding.}
\usage{
scale_prec_to_perp(inp, method, opt, iter, perp, out, verbose = TRUE)
}
\arguments{
\item{inp}{Input data.}

\item{method}{Embedding method.}

\item{opt}{Optimizer.}

\item{iter}{Current iteration.}

\item{perp}{Perplexity of the input probability.}

\item{out}{Output data.}

\item{verbose}{If \code{TRUE} print message about tricks during the
embedding.}
}
\description{
Factory function which creates an output weight function for a given input
perplexity, in the context of multiscale embedding.
}
\seealso{
Other sneer kernel modifiers: \code{\link{transfer_kernel_bandwidths}}
}

