% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{scale_prec_to_perp}
\alias{scale_prec_to_perp}
\title{Factory function which creates an output weight function for a given input
perplexity, in the context of multiscale embedding.}
\usage{
scale_prec_to_perp(inp, out, method)
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}
}
\description{
Factory function which creates an output weight function for a given input
perplexity, in the context of multiscale embedding.
}
\seealso{
Other sneer kernel modifiers: \code{\link{transfer_kernel_bandwidths}}
}

