% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scores_plot}
\alias{scores_plot}
\title{PCA Scores Plot}
\usage{
scores_plot(df, pca_indexes = NULL, label_name = NULL, center = TRUE,
  scale = FALSE, cex = 0.5)
}
\arguments{
\item{df}{Data frame.}

\item{pca_indexes}{Numeric vector containing the column indexes in the
data frame which should be used for the PCA calculation. Default is
\code{NULL}, in which case all numeric columns will be used.}

\item{label_name}{name of the column that contains a value to be used
to color each point. Default is \code{NULL}, in which case the first
encountered factor column will be used.}

\item{center}{If \code{TRUE}, mean-center the columns.}

\item{scale}{If \code{TRUE}, scale the columns to have unit variance.}

\item{cex}{Numeric \strong{c}haracter \strong{ex}pansion factor;
multiplied by \code{\link[graphics]{par}("cex")} yields the final
character size of the labels.}
}
\description{
Carries out PCA on a data frame (using the provided indexes) and then
plots the first two scores, coloring the points with a user-specified
label.
}
\examples{
\dontrun{
 # PCA on the scaled iris dataset, use the "Species" column to display
 scores_plot(iris, 1:4, "Species", scale = TRUE)

 # PCA on the s1k dataset, use all numeric values, and color by the only
 # factor ("Label"). Change text size.
 scores_plot(s1k, cex = 1)
}
}

