% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{set_solution}
\alias{set_solution}
\title{Set Output Data Coordinates}
\usage{
set_solution(opt, inp, coords, method, out = NULL)
}
\arguments{
\item{opt}{Optimizer.}

\item{inp}{Input data.}

\item{coords}{Matrix of coordinates.}

\item{method}{Embedding method.}

\item{out}{Existing output data. Optional.}
}
\value{
\code{out} list with updated with coords and auxiliary data.
}
\description{
This function sets the embedded coordinates in the output data, as well as
recalculating any auxiliary output data that is dependent on the coordinates
(e.g. distances and probabilities)
}

