% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{set_solution}
\alias{set_solution}
\title{Set Output Data Coordinates}
\usage{
set_solution(inp, coords, method, mat_name = "ym", out = NULL)
}
\arguments{
\item{inp}{Input data.}

\item{coords}{Matrix of coordinates.}

\item{method}{Embedding method.}

\item{mat_name}{Name of the \code{out} entry to store \code{coords} in.
Defaults to \code{ym}.}

\item{out}{Existing output data. Optional.}
}
\value{
\code{out} list with updated with coords and auxiliary data.
}
\description{
This function sets the embedded coordinates in the output data, as well as
recalculating any auxiliary output data that is dependent on the coordinates
(e.g. distances and probabilities)
}

