% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{smmds}
\alias{smmds}
\title{Metric Multi-dimensional Scaling (MDS) using SSTRESS Cost Function.}
\usage{
smmds(eps = .Machine$double.eps)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
a list containing:
 \item{\code{cost_fn}}{Cost function for the embedding: for this
 implementation of MDS, the SSTRESS cost function.}
 \item{\code{stiffness_fn}}{Stiffness function for \code{cost_fn}.}
 \item{\code{update_out_fn}}{Function to calculate and store any needed
 data after a coordinate update.}
 \item{\code{eps}}{Small floating point value used to prevent numerical
 problems, e.g. in gradients and cost functions.}
}
\description{
Embedding method.
}
\details{
This function minimizes an embedding using the SSTRESS loss function:

\deqn{SSTRESS = \sum_{i<j} ((r_{ij}^2 - d_{ij})^2)^2}{SSTRESS = sum(rij^2-dij^2)^2}

\eqn{r_{ij}}{rij} is the input distance between point \eqn{i} and point
\eqn{j} and \eqn{d_{ij}}{dij} is the corresponding output distance.
}
\seealso{
Other sneer distance embedding methods: \code{\link{distance_embedding_methods}},
  \code{\link{mmds}}, \code{\link{sammon_map}}

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{mmds}},
  \code{\link{sammon_map}}, \code{\link{ssne}},
  \code{\link{tasne}}, \code{\link{tsne}}
}

