% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{ssne}
\alias{ssne}
\title{Symmetric Stochastic Neighbor Embedding}
\usage{
ssne(eps = .Machine$double.eps)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
a list containing:
 \item{\code{cost_fn}}{Cost function for the embedding: the Kullback-Leibler
 divergence between the input and output probabilities.}
 \item{\code{weight_fn}}{Weight function for mapping squared distances in the
 embedded coordinates to weights, which are in turn converted to
 probabilities: SSNE uses the exponential function.}
 \item{\code{stiffness_fn}}{Stiffness function.}
 \item{\code{update_out_fn}}{Function to calculate and store any needed
 data after a coordinate update.}
 \item{\code{after_init_fn}}{Method-specific initialization function to
 invoke after input and output initialization callbacks. For this method,
 the input probability matrix must be converted from a row probability
 matrix to a joint probability matrix.}
 \item{\code{eps}}{Small floating point value used to prevent numerical
 problems, e.g. in gradients and cost functions.}
}
\description{
Creates a list of functions that collectively implement SSNE.
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix based on embedded coordinates.}
}
}
\references{
J.A. Cook, I. Sutskever, A. Mnih, and G.E. Hinton.
Visualizing similarity data with a mixture of maps.
In Proceedings of the 11th International Conference on Artificial
Intelligence and Statistics, volume 2, pages 67-74, 2007.

Laurens van der Maarten, Geoffrey Hinton.
Visualizing Data using t-SNE.
Journal of Machine Learning Research, 2008, 9, 2579-2605.
}
\seealso{
Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{mmds}},
  \code{\link{sammon_map}}, \code{\link{smmds}},
  \code{\link{tasne}}, \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{probability_embedding_methods}},
  \code{\link{tasne}}, \code{\link{tsne}}
}

