% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{ssne}
\alias{ssne}
\title{Symmetric Stochastic Neighbor Embedding.}
\usage{
ssne()
}
\value{
a list with the following members:
\itemize{
 \item \code{cost_fn} Cost function for the embedding: the Kullback-Leibler
 divergence between the input and output probabilities.
 \item \code{weight_fn} Weight function for mapping squared distances in the
 embedded coordinates to weights, which are in turn converted to
 probabilities: SSNE uses the exponential function.
 \item \code{stiffness_fn} Stiffness function.
 \item \code{prob_out_fn} Function to convert weights matrix to a specific
 probability matrix, e.g. row-based or joint.
 \item \code{after_init_fn} Method-specific initialization function to
 invoke after input and output initialization callbacks. For this method,
 the input probability matrix must be converted from a row probability
 matrix to a joint probability matrix.
}
}
\description{
Creates a list of functions that collectively implement SSNE.
}
\references{
J.A. Cook, I. Sutskever, A. Mnih, and G.E. Hinton.
Visualizing similarity data with a mixture of maps.
In Proceedings of the 11th International Conference on Artificial
Intelligence and Statistics, volume 2, pages 67-74, 2007.

Laurens van der Maarten, Geoffrey Hinton.
Visualizing Data using t-SNE.
Journal of Machine Learning Research, 2008, 9, 2579-2605.
}

