% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{ssne}
\alias{ssne}
\title{Symmetric Stochastic Neighbor Embedding (SSNE)}
\usage{
ssne(eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
SSNE is a variant of \code{\link{asne}} where the probabilities are with
respect to pairs of points, not individual points. The element
\code{P[i, j]} in matrix P should be thought of as the probability of
selecting a pair of points i and j as close neighbours. As a result, unlike
ASNE, the probability matrix in SSNE:

\itemize{
 \item{represents one probability distribution, i.e. the grand sum of the
 matrix is one.}
 \item{is symmetric, i.e. \code{P[i, j] == P[j, i]} and therefore the
 probabilities are joint probabilities.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix based on embedded coordinates.}
}
}
\examples{
\dontrun{
embed_prob(method = ssne(), ...)
}
}
\references{
Cook, J., Sutskever, I., Mnih, A., & Hinton, G. E. (2007).
Visualizing similarity data with a mixture of maps.
In \emph{International Conference on Artificial Intelligence and Statistics} (pp. 67-74).

Van der Maaten, L., & Hinton, G. (2008).
Visualizing data using t-SNE.
\emph{Journal of Machine Learning Research}, \emph{9}(2579-2605).
}
\seealso{
SSNE uses the \code{\link{kl_cost}} cost function and the
  \code{\link{exp_weight}} similarity function. The return value of this
  function should be used with the \code{\link{embed_prob}} embedding
  function.

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{hsnerv}},
  \code{\link{hssne}}, \code{\link{jse}},
  \code{\link{mmds}}, \code{\link{nerv}},
  \code{\link{rasne}}, \code{\link{rssne}},
  \code{\link{rtsne}}, \code{\link{sammon_map}},
  \code{\link{smmds}}, \code{\link{snerv}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hsnerv}}, \code{\link{hssne}},
  \code{\link{jse}}, \code{\link{nerv}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne}}, \code{\link{rssne}},
  \code{\link{rtsne}}, \code{\link{snerv}},
  \code{\link{tasne}}, \code{\link{tnerv}},
  \code{\link{tsne}}
}

