% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{ssne_stiffness}
\alias{ssne_stiffness}
\title{SSNE Stiffness Function}
\usage{
ssne_stiffness(pm, qm, beta = 1)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}

\item{beta}{The precision of the weighting function.}
}
\value{
Stiffness matrix.
}
\description{
The precision parameter \code{beta} is normally left at its default value of
1. Note that unlike the \code{\link{asne_stiffness}} function, a vector of
precisions can not be used as input to \code{beta}: an incorrect gradient
will result.
}

