% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiffness.R
\name{ssne_stiffness}
\alias{ssne_stiffness}
\title{SSNE Stiffness Function}
\usage{
ssne_stiffness(pm, qm)
}
\arguments{
\item{pm}{Input joint probability matrix.}

\item{qm}{Output joint probabilty matrix.}
}
\value{
Stiffness matrix.
}
\description{
SSNE Stiffness Function
}

