% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_update.R
\name{step_momentum}
\alias{step_momentum}
\title{Optimizer update methods.
Step momentum.}
\usage{
step_momentum(init_momentum = 0.5, final_momentum = 0.8,
  switch_iter = 250, verbose = TRUE)
}
\arguments{
\item{init_momentum}{Momentum value for the first \code{switch_iter}
iterations.}

\item{final_momentum}{Momentum value after \code{switch_iter} iterations.}

\item{switch_iter}{Iteration number at which to switch from
\code{init_momentum} to \code{final_momentum}.}

\item{verbose}{if \code{TRUE}, log info about the momentum.}
}
\value{
A solution update method for use by the optimizer. A list consisting
of:
 \item{\code{init_momentum}}{Initial momentum.}
 \item{\code{final_momentum}}{Final momentum.}
 \item{\code{mom_switch_iter}}{Switch iteration.}
 \item{\code{init}}{Function to do any needed initialization.}
 \item{\code{get_update}}{Function to return the current update, which will
 be added to current solution matrix.}
 \item{\code{after_step}}{Function to do any needed updating or internal
 state before the next optimization step.}
}
\description{
Create an callback for the optimizer to use to update the embedding solution.
Update is in the form of a step momentum function.
}
\seealso{
Other sneer optimization update methods: \code{\link{linear_momentum}},
  \code{\link{nesterov_nsc_momentum}},
  \code{\link{no_momentum}}
}

