% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inp.R
\name{symmetrize_matrix}
\alias{symmetrize_matrix}
\title{Create a symmetric matrix from a square matrix.}
\usage{
symmetrize_matrix(pm)
}
\arguments{
\item{pm}{Square matrix to symmetrize.}
}
\value{
Symmetrized matrix such that \code{pm[i, j]} = \code{pm[j, i]}
}
\description{
The matrix is symmetrized by setting \code{pm[i, j]} and \code{pm[j, i]} to
their average, i.e. \code{Pij} = \code{(Pij + Pji)/2} = \code{Pji}.
}
\details{
In SSNE and TSNE, this is used as part of the process of converting the row
stochastic matrix of conditional input probabilities to a joint probability
matrix.
}

