% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{tasne}
\alias{tasne}
\title{t-distributed Asymmetric Stochastic Neighbor Embedding.}
\usage{
tasne()
}
\value{
a list with the following members:
\itemize{
 \item \code{cost_fn} Cost function for the embedding: the Kullback-Leibler
 divergence between the input and output probabilities.
 \item \code{weight_fn} Weight function for mapping squared distances in the
 embedded coordinates to weights, which are in turn converted to
 probabilities: tASNE uses the exponential function.
 \item \code{stiffness_fn} Stiffness function.
 \item \code{prob_out_fn} Function to convert weights matrix to a specific
 probability matrix, e.g. row-based or joint.
 \item \code{after_init_fn} Method-specific initialization function to
 invoke after input and output initialization callbacks.
}
}
\description{
Creates a list of functions that collectively implement t-ASNE: a method that
I just made up to illustrate how to explore different aspects of embedding
within sneer: this uses the t-distributed distance weighting of t-SNE, but
for probability generation uses the point-wise distribution of ASNE.
}

