% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{tasne}
\alias{tasne}
\title{t-distributed Asymmetric Stochastic Neighbor Embedding (t-ASNE)}
\usage{
tasne(eps = .Machine$double.eps)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
a list containing:
 \item{\code{cost_fn}}{Cost function for the embedding: the Kullback-Leibler
 divergence between the input and output probabilities.}
 \item{\code{weight_fn}}{Weight function for mapping squared distances in the
 embedded coordinates to weights, which are in turn converted to
 probabilities: t-ASNE uses the exponential function.}
 \item{\code{stiffness_fn}}{Stiffness function.}
 \item{\code{update_out_fn}}{Function to calculate and store any needed
 data after a coordinate update.}
 \item{\code{after_init_fn}}{Method-specific initialization function to
 invoke after input and output initialization callbacks.}
 \item{\code{eps}}{Small floating point value used to prevent numerical
 problems, e.g. in gradients and cost functions.}
}
\description{
A probability-based embedding method.
}
\details{
Creates a list of functions that collectively implement t-ASNE: a method that
I just made up to illustrate how to explore different aspects of embedding
within sneer: this uses the t-distributed distance weighting of t-SNE, but
for probability generation uses the point-wise distribution of ASNE.
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Row probability matrix based on the weight matrix
 \code{wm}.}
 \item{\code{wm}}{Weight matrix generated from the distances between points
 in \code{ym}.}
}
}
\seealso{
Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{mmds}},
  \code{\link{sammon_map}}, \code{\link{smmds}},
  \code{\link{ssne}}, \code{\link{tsne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{probability_embedding_methods}},
  \code{\link{ssne}}, \code{\link{tsne}}
}

