% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight.R
\name{tdist_weight}
\alias{tdist_weight}
\title{Student-t distribution weighting.}
\usage{
tdist_weight(d2m)
}
\arguments{
\item{d2m}{Matrix of squared distances.}
}
\value{
Weight matrix.
}
\description{
Applies weighting using the Student-t distribution with one degree of
freedom. Compared to the exponential weighting this has a much heavier tail.
}
\details{
\eqn{W = \frac{1}{(1 + D^2)}}
}

