% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{tsne}
\alias{tsne}
\title{t-Distributed Stochastic Neighbor Embedding (t-SNE).}
\usage{
tsne(eps = .Machine$double.eps)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
t-SNE is a variant of \code{\link{ssne}} where the similarity function used
to generate output probabilities is the Student t-Distribution with one
degree of freedom.
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix based on the weight matrix
 \code{wm}.}
 \item{\code{wm}}{Weight matrix generated from the distances between points
 in \code{ym}.}
}
}
\examples{
\dontrun{
embed_prob(method = tsne(), ...)
}
}
\references{
Van der Maaten, L., & Hinton, G. (2008).
Visualizing data using t-SNE.
\emph{Journal of Machine Learning Research}, \emph{9}(2579-2605).
}
\seealso{
t-SNE uses the \code{\link{tdist_weight}} similarity function.
The return value of this function should be used with the
\code{\link{embed_prob}} embedding function.

Other sneer embedding methods: \code{\link{asne}},
  \code{\link{embedding_methods}}, \code{\link{hssne}},
  \code{\link{mmds}}, \code{\link{sammon_map}},
  \code{\link{smmds}}, \code{\link{ssne}},
  \code{\link{tasne}}

Other sneer probability embedding methods: \code{\link{asne}},
  \code{\link{hssne}},
  \code{\link{probability_embedding_methods}},
  \code{\link{ssne}}, \code{\link{tasne}}
}

