% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{tsne}
\alias{tsne}
\title{t-distributed Stochastic Neighbor Embedding.}
\usage{
tsne(mat_name = "ym", eps = .Machine$double.eps)
}
\arguments{
\item{mat_name}{Name of the matrix in the output data list that will contain
the embedded coordinates.}

\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}
}
\value{
a list containing:
 \item{\code{cost_fn}}{Cost function for the embedding: the Kullback-Leibler
 divergence between the input and output probabilities.}
 \item{\code{weight_fn}}{Weight function for mapping squared distances in the
 embedded coordinates to weights, which are in turn converted to
 probabilities: t-SNE uses the Student t-distribution with one degree of
 freedom.}
 \item{\code{stiffness_fn}}{Stiffness function.}
 \item{\code{update_out_fn}}{Function to calculate and store any needed
 data after a coordinate update.}
 \item{\code{after_init_fn}}{Method-specific initialization function to
 invoke after input and output initialization callbacks. For this method,
 the input probability matrix must be converted from a row probability
 matrix to a joint probability matrix.}
 \item{\code{mat_name}}{Name of the matrix in the output data list that will
 contain the embedded coordinates.}
 \item{\code{eps}}{Small floating point value used to prevent numerical
 problems, e.g. in gradients and cost functions.}
}
\description{
Creates a list of functions that collectively implement t-SNE.
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates. This name can be changed by
 specifying \code{mat_name}.}
 \item{\code{qm}}{Joint probability matrix based on the weight matrix
 \code{wm}.}
 \item{\code{wm}}{Weight matrix generated from the distances between points
 in \code{ym}.}
}
}
\references{
Laurens van der Maarten, Geoffrey Hinton.
Visualizing Data using t-SNE.
Journal of Machine Learning Research, 2008, 9, 2579-2605.
}
\seealso{
Other sneer embedding methods: \code{\link{asne}},
  \code{\link{mds}}, \code{\link{sammon}},
  \code{\link{ssne}}, \code{\link{tasne}}
}

