% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sne.R
\name{tsne}
\alias{tsne}
\title{t-Distributed Stochastic Neighbor Embedding (t-SNE).}
\usage{
tsne(eps = .Machine$double.eps, verbose = TRUE)
}
\arguments{
\item{eps}{Small floating point value used to prevent numerical problems,
e.g. in gradients and cost functions.}

\item{verbose}{If \code{TRUE}, log information about the embedding.}
}
\value{
An embedding method for use by an embedding function.
}
\description{
A probability-based embedding method.
}
\details{
t-SNE is a variant of \code{ssne} where the similarity function used
to generate output probabilities is the Student t-Distribution with one
degree of freedom.

The probability matrix in t-SNE:

\itemize{
 \item{represents one probability distribution, i.e. the grand sum of the
 matrix is one.}
 \item{is symmetric, i.e. \code{P[i, j] == P[j, i]} and therefore the
 probabilities are joint probabilities.}
}
}
\section{Output Data}{

If used in an embedding, the output data list will contain:
\describe{
 \item{\code{ym}}{Embedded coordinates.}
 \item{\code{qm}}{Joint probability matrix based on the weight matrix
 \code{wm}.}
 \item{\code{wm}}{Weight matrix generated from the distances between points
 in \code{ym}.}
}
}
\examples{
\dontrun{
embed_prob(method = tsne(), ...)
}
}
\references{
Van der Maaten, L., & Hinton, G. (2008).
Visualizing data using t-SNE.
\emph{Journal of Machine Learning Research}, \emph{9}(2579-2605).
}
\seealso{
t-SNE uses the \code{kl_cost} cost function and the
  \code{tdist_weight} similarity function. The return value of this
  function should be used with the \code{embed_prob} embedding
  function.

Other sneer embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsnerv_plugin}}, \code{\link{hsnerv}},
  \code{\link{hssne_plugin}}, \code{\link{hssne}},
  \code{\link{jse_plugin}}, \code{\link{nerv_plugin}},
  \code{\link{nerv}}, \code{\link{plugin}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sjse_plugin}}, \code{\link{snerv_plugin}},
  \code{\link{snerv}}, \code{\link{ssne_plugin}},
  \code{\link{ssne}}, \code{\link{tasne_plugin}},
  \code{\link{tasne}}, \code{\link{tnerv_plugin}},
  \code{\link{tnerv}}, \code{\link{tsne_plugin}},
  \code{\link{uhsnerv_plugin}}, \code{\link{uhsnerv}},
  \code{\link{unerv_plugin}}, \code{\link{unerv}},
  \code{\link{usnerv_plugin}}, \code{\link{usnerv}}

Other sneer probability embedding methods: \code{\link{asne_plugin}},
  \code{\link{asne}}, \code{\link{hsjse_plugin}},
  \code{\link{hsnerv_plugin}}, \code{\link{hsnerv}},
  \code{\link{hssne_plugin}}, \code{\link{hssne}},
  \code{\link{jse_plugin}}, \code{\link{nerv_plugin}},
  \code{\link{nerv}}, \code{\link{plugin}},
  \code{\link{probability_embedding_methods}},
  \code{\link{rasne_plugin}}, \code{\link{rasne}},
  \code{\link{rssne_plugin}}, \code{\link{rssne}},
  \code{\link{rtsne_plugin}}, \code{\link{rtsne}},
  \code{\link{sjse_plugin}}, \code{\link{snerv_plugin}},
  \code{\link{snerv}}, \code{\link{ssne_plugin}},
  \code{\link{ssne}}, \code{\link{tasne_plugin}},
  \code{\link{tasne}}, \code{\link{tnerv_plugin}},
  \code{\link{tnerv}}, \code{\link{tsne_plugin}},
  \code{\link{uhsnerv_plugin}}, \code{\link{uhsnerv}},
  \code{\link{unerv_plugin}}, \code{\link{unerv}},
  \code{\link{usnerv_plugin}}, \code{\link{usnerv}}
}

