% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_step.R
\name{tsne_jacobs}
\alias{tsne_jacobs}
\title{Jacobs step size method using parameters from the t-SNE paper.}
\usage{
tsne_jacobs()
}
\value{
A step size method suitable for use with t-SNE.
}
\description{
Jacobs step size method using parameters from the t-SNE paper.
}
\references{
Laurens van der Maarten, Geoffrey Hinton.
Visualizing Data using t-SNE.
Journal of Machine Learning Research, 2008, 9, 2579-2605.

R. A. Jacobs.
Increased rates of convergence through learning rate adaptation.
Neural Networks, 1:295-307, 1988.
}

