% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_step.R
\name{tsne_jacobs}
\alias{tsne_jacobs}
\title{Jacobs Step Size Method from t-SNE Paper}
\usage{
tsne_jacobs()
}
\value{
A step size method suitable for use with t-SNE.
}
\description{
Factory function for creating an optimizer step size method.
}
\examples{
# Use as part of the make_opt function for configuring an optimizer's
# step size method:
make_opt(step_size = tsne_jacobs())
}
\references{
Van der Maaten, L., & Hinton, G. (2008).
Visualizing data using t-SNE.
\emph{Journal of Machine Learning Research}, \emph{9}(2579-2605).

Jacobs, R. A. (1988).
Increased rates of convergence through learning rate adaptation.
\emph{Neural networks}, \emph{1}(4), 295-307.
}
\seealso{
The return value of this function is intended for internal use of
the sneer framework only. See \code{optimization_step_size_interface}
for details on the functions and values defined for this method.

Other sneer optimization step size methods: \code{\link{bold_driver}},
  \code{\link{constant_step_size}}, \code{\link{jacobs}},
  \code{\link{optimization_step_size}}
}

