% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{tsne_opt}
\alias{tsne_opt}
\title{t-SNE Optimizer}
\usage{
tsne_opt()
}
\value{
optimizer with parameters from the t-SNE paper.
}
\description{
Optimizer factory function.
}
\details{
Wrapper around \code{make_opt} which sets the optimizer parameters to
that from the t-SNE paper.
}
\examples{
# Should be passed to the opt argument of an embedding function:
\dontrun{
 embed_prob(opt = tsne_opt(), ...)
}
}
\seealso{
\code{embed_prob} and \code{embed_dist} for how to use
 this function for configuring an embedding.

Other sneer optimization methods: \code{\link{back_nag_adapt}},
  \code{\link{back_nag}}, \code{\link{bold_nag_adapt}},
  \code{\link{bold_nag}}, \code{\link{gradient_descent}},
  \code{\link{make_opt}}, \code{\link{nag}},
  \code{\link{optimization_methods}}, \code{\link{ropt}}
}

