% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricks.R
\name{tsne_tricks}
\alias{tsne_tricks}
\title{t-SNE Tricks}
\usage{
tsne_tricks(verbose = TRUE)
}
\arguments{
\item{verbose}{If \code{TRUE} print message about tricks during the
embedding.}
}
\value{
tricks callback parameterized to behave like the t-SNE paper.
}
\description{
Tricks configured according to the details in the t-SNE paper.
}
\examples{
# Should be passed to the tricks argument of an embedding function:
\dontrun{
 embed_prob(tricks = tsne_tricks(), ...)
}
}
\references{
Van der Maaten, L., & Hinton, G. (2008).
Visualizing data using t-SNE.
\emph{Journal of Machine Learning Research}, \emph{9}(2579-2605).
}
\seealso{
\code{\link{embed_prob}} for how to use this function for
configuring an embedding.

Other sneer trick collections: \code{\link{make_tricks}}
}

