% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{update_out}
\alias{update_out}
\title{Update matrices that are dependent on embedding coordinates.}
\usage{
update_out(inp, out, method, mat_name = "ym")
}
\arguments{
\item{inp}{Input data.}

\item{out}{Output data.}

\item{method}{Embedding method.}

\item{mat_name}{Name of the matrix in \code{out} which contains the
coordinates.}
}
\value{
Updated version of \code{out}.
}
\description{
After embedding coordinates are updated, other matrices may need
to be also updated: e.g. distance matrix and for similarity-based
embedding weights and probability matrices.
}

