% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{update_out}
\alias{update_out}
\title{Output Update Factory Function}
\usage{
update_out(keep = c("qm"))
}
\arguments{
\item{keep}{List containing any or all of the following matrix names:
\describe{
 \item{\code{d2m}}{Output squared distances matrix.}
 \item{\code{wm}}{Output weight matrix.}
 \item{\code{qm}}{Output probability matrix.}
}}
}
\value{
The output list, with all the matrices specified by \code{keep}
  added to it.
}
\description{
Embedding methods can specify which of the three matrices created as
part of mapping from embedded coordinates to the output probabilities they
want to keep. The squared distances (\code{d2m}) are not always useful,
except if the plugin gradient is being used to calculate the stiffness
matrix. The weight matrix (\code{wm}) is used by the plugin gradient method
and by some non-plugin method stiffness functions (e.g. \code{\link{tsne}}
or \code{\link{hssne}}). The output probability (\code{qm}) is an integral
part of all cost functions and gradients so should always be retained.
}

