% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_to_pcond}
\alias{weights_to_pcond}
\title{Create Conditional Probability Matrix from Weight Matrix}
\usage{
weights_to_pcond(wm)
}
\arguments{
\item{wm}{Matrix of weighted distances. Asymmetric or symmetric.}
}
\value{
A list containing
 \item{pm}{Conditional probability matrix.}
}
\description{
A weight to probability function.
}
\details{
Creates a conditional probability matrix: the grand sum of the elements are
one, but does not have to be symmetric. The weight matrix can be symmetric
or asymmetric.
}

