% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_to_pcond}
\alias{weights_to_pcond}
\title{Create Conditional Probability Matrix from Weight Matrix}
\usage{
weights_to_pcond(wm)
}
\arguments{
\item{wm}{Matrix of weighted distances.}
}
\value{
Probability matrix.
}
\description{
Used in SSNE and TSNE. The probability matrix is such that all elements are
positive and sum to 1.
}

