% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_to_probs}
\alias{weights_to_probs}
\title{Weight Matrix to Probability Matrix Conversion}
\usage{
weights_to_probs(wm, method)
}
\arguments{
\item{wm}{Probability Matrix. Must have a "type" attribute with one of the
following values:
\describe{
\item{"symm"}{A symmetric matrix.}
\item{"asymm"}{An asymmetric matrix.}
}}

\item{method}{Embedding method.}
}
\value{
Probability matrix with a type suitable for the embedding
  \code{method}.
}
\description{
Given a weight matrix and an embedding method, this function creates a
probability matrix.
}

