% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_sim_stiff.R
\name{weights_to_prow}
\alias{weights_to_prow}
\title{Create Row Probability Matrix from Weight Matrix}
\usage{
weights_to_prow(wm)
}
\arguments{
\item{wm}{Matrix of weighted distances.}
}
\value{
A list containing
 \item{pm}{Row probability matrix.}
}
\description{
Used in ASNE. The probability matrix is such that all elements are positive
and each row sums to 1.
}

